% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_function_imported.R
\name{add_trace}
\alias{add_trace}
\title{I. Internal Functions from Plotly Package
THe functions are from Plotly Package and was called internally only}
\usage{
add_trace(p, ..., data = NULL, inherit = TRUE)
}
\description{
I. Internal Functions from Plotly Package
THe functions are from Plotly Package and was called internally only
}
\references{
https://cran.r-project.org/package=plotly

Sievert C (2020). Interactive Web-Based Data Visualization with R, plotly, and shiny. Chapman and Hall/CRC. ISBN 9781138331457, https://plotly-r.com.
}
