% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_lipomics.R
\name{analyze.lipids}
\alias{analyze.lipids}
\title{Lipid analysis pipeliner}
\usage{
analyze.lipids(inFile, iso = "y")
}
\arguments{
\item{inFile}{Input the file to read in}

\item{iso}{Default is set to "y"}
}
\description{
Lipid analysis pipeliner
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
