% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pca_anova2.R
\name{ANOVA2.Anal}
\alias{ANOVA2.Anal}
\title{Perform Two-way ANOVA}
\usage{
ANOVA2.Anal(mSetObj=NA, thresh=0.05, p.cor="fdr", 
type="time0", aov.type=1, use.interact=1)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{thresh}{Input the p-value threshold}

\item{p.cor}{Select method for p-value correction, bonferroni, holm or fdr}

\item{type}{Select b to perform between-subjects ANOVA, 
and w for within-subjects ANOVA}

\item{aov.type}{Specify 1 for ANOVA type 1, or 3 for ANOVA type 3}

\item{use.interact}{Numeric, whether to consider 
interaction in two-way repeated ANOVA (1) or not (0).}
}
\description{
Perform Two-way ANOVA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
