% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{aof}
\alias{aof}
\title{ANOVA}
\usage{
aof(x, cls)
}
\description{
Perform anova and only return p values and MSres (for Fisher's LSD)
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
