% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_pca_anova2.R
\name{aov.2way}
\alias{aov.2way}
\title{Perform Two-way ANOVA}
\usage{
aov.2way(x)
}
\arguments{
\item{x}{Input data to perform 2-way ANOVA}
}
\description{
Perform Two-way ANOVA 
Perform two-way anova
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
