% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pca_anova2.R
\name{aov.between}
\alias{aov.between}
\title{Perform Two-way ANOVA}
\usage{
aov.between(x)
}
\arguments{
\item{x}{Input data to perform 2-way ANOVA}
}
\description{
Perform Two-way ANOVA 
Perform between-subjects anova
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
