% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_pca_anova2.R
\name{aov.mixed}
\alias{aov.mixed}
\title{Perform Two-way ANOVA}
\usage{
aov.mixed(x)
}
\arguments{
\item{x}{Input the data}
}
\description{
Perform Two-way ANOVA 
Perform within-subjects anova
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
