% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pca_anova2.R
\name{aov.repeated}
\alias{aov.repeated}
\title{Perform Two-way ANOVA}
\usage{
aov.repeated(x, time.fac)
}
\arguments{
\item{x}{Input the data}

\item{time.fac}{Input the time factor}
}
\description{
Perform Two-way ANOVA 
Perform repeated measure one-way anova
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
