% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pca_anova2.R
\name{aov.within}
\alias{aov.within}
\title{Perform Two-way ANOVA}
\usage{
aov.within(x, time.fac)
}
\description{
Perform Two-way ANOVA 
Perform within-subjects anova
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
