% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{ASCAfun.res}
\alias{ASCAfun.res}
\title{Function to perform ASCA}
\usage{
ASCAfun.res(X, Fac)
}
\arguments{
\item{X}{Input list of compounds}

\item{Fac}{Numeric 
McGill University, Canada
License: MIT License}
}
\description{
Perform ASCA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
}
