% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{ASCAfun2}
\alias{ASCAfun2}
\title{Function to perform ASCA}
\usage{
ASCAfun2(X, Desa, Desb, Fac)
}
\arguments{
\item{X}{Numeric, number of compounds}

\item{Desa}{Number of levels in the factor TIME}

\item{Desb}{Number of levels in the other factor}

\item{Fac}{Numeric, the factor}
}
\description{
Perform ASCA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
