% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{BreakLongNames}
\alias{BreakLongNames}
\title{Utilities for creating pathway maps for MetPA}
\usage{
BreakLongNames(long.nm, wrap.len = 10, tol.len = 5)
}
\description{
Perform various utilities for MetPA
break up a long name with brute force
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
