% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_c_utils.R
\name{C_imodwt_r}
\alias{C_imodwt_r}
\title{Internal C fucntion - C_imodwt_r}
\usage{
C_imodwt_r(y, z, N, j, L, ht, gt, XX)
}
\description{
Internal C fucntion - C_imodwt_r
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge University Press.
}
