% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_c_utils.R
\name{C_modwt_r}
\alias{C_modwt_r}
\title{Internal C fucntion - C_modwt_r}
\usage{
C_modwt_r(X, N, j, L, ht, gt, W, V)
}
\description{
Internal C fucntion - C_modwt_r
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge University Press.
}
