% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{calcCV}
\alias{calcCV}
\title{Calculatre CV method}
\usage{
calcCV(xset)
}
\arguments{
\item{xset}{XCMSnExp Object, this object is produced by 'calculateSet_doe' function.}
}
\description{
Calculatre CV method
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
