% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{calcGaussianS}
\alias{calcGaussianS}
\title{Calculatre Gaussian Peak Ratio method}
\usage{
calcGaussianS(mSet, object, useNoise, BPPARAM = bpparam())
}
\arguments{
\item{mSet}{MetaboAnalystR Object, this object is produced by 'calculateSet_doe' function.}

\item{object}{MSnExp object, the trimmed or the original data (Generated by ImportRawMSData function with "inMemory" mode).}

\item{useNoise}{Numeric, the noise level removed to evalute the gaussian peak.}

\item{BPPARAM}{MulticoreParam method, used to set the parallel method. Default is bpparam().}
}
\description{
Calculatre Gaussian Peak Ratio method
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
