% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{calcPPS2}
\alias{calcPPS2}
\title{Calculate PPS method}
\usage{
calcPPS2(xset, isotopeIdentification = c("IPO", "CAMERA"), ...)
}
\arguments{
\item{xset}{xcmsSet Object, this object is produced by 'calculateSet_doe' function, and transformed 
with as(objec,'xcmsSet') function.}

\item{isotopeIdentification}{Character, IsotopeIdentidication method, usually includes 'IPO' and 'CAMERA'.}
}
\description{
Calculate PPS method
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
