% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{calcRCS_GSValues}
\alias{calcRCS_GSValues}
\title{Calculatre RCS and GS method}
\usage{
calcRCS_GSValues(xset)
}
\arguments{
\item{xset}{XCMSnExp Object, this object is produced by 'calculateSet_doe' function.}
}
\description{
Calculatre RCS and GS method
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
