% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_lipomics.R
\name{calculateConcISO}
\alias{calculateConcISO}
\title{Calculate Concentration ISO}
\usage{
calculateConcISO(dat, cls.name, cls.num, min.file, prob.file)
}
\arguments{
\item{dat}{Input the data}

\item{cls.name}{Input the class names}

\item{cls.num}{Input the number of classes}

\item{min.file}{Input the min file}

\item{prob.file}{Input the prob file}
}
\description{
Assuming independent random distribution of FA, the most probable frequency
will be the product of the each component. Note: the data is concentration,
we need to get frequncies - percentage w.r.t the total nmol.
the result is the saved as separate files for each lipid class
data for each FA class, first col is sample name
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
