% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_stats.R
\name{CalculateGlobalTestScore}
\alias{CalculateGlobalTestScore}
\title{Quantitative enrichment analysis with globaltest}
\usage{
CalculateGlobalTestScore(mSetObj = NA, covariates = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{covariates}{Optional covariates to include in the globaltest model. Default is NA (no covariates).}
}
\description{
Various enrichment analysis algorithms
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
