% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{calculateGPRT}
\alias{calculateGPRT}
\title{Alignment Method}
\usage{
calculateGPRT(mSet, param)
}
\arguments{
\item{mSet}{mSet object, the data produced by 'calculatePPKs' function.}

\item{Set_parameters}{Matrix, the parameters combination produced automatically according to}

\item{task}{Numeric, task order for XCMS paramters table to run the peak picking and alignment.}
}
\description{
Alignment Method
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
