% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{CalculateImpVarCutoff}
\alias{CalculateImpVarCutoff}
\title{Calculate the Important Variable Cutoff}
\usage{
CalculateImpVarCutoff(mSetObj, spe.thresh, lev.thresh)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{spe.thresh}{alpha threshold, less is better, default less than 5 percentile based chi-square
note: spe and leverage are vectors, not a single value, but a list to store the result
note: the last model is Model.res, no spe
Calculate leverage cutoff based on permutation
Calculate the reference distribution of leverages
note: leverage.perm is a list with each member in a 3 column matrix}

\item{lev.thresh}{leverage threshold, the higher better, default more than 95 percentile of permuted leverage}
}
\description{
This function calculates the all important features based on 
a specfic cutoff.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
