% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{CalculateOraScore}
\alias{CalculateOraScore}
\title{Over-representation analysis using hypergeometric tests
The probability is calculated from obtaining the equal or higher number
of hits using 1-phyper. Since phyper is a cumulative probability,
to get P(X>=hit.num) => P(X>(hit.num-1))}
\usage{
CalculateOraScore(mSetObj=NA, nodeImp, method)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{nodeImp}{Indicate the pathway topology analysis, "rbc" for relative-betweeness centrality, 
and "dgr" for out-degree centrality.}

\item{method}{is "fisher" or "hyperg"}
}
\description{
Calculate the over representation analysis score
}
\details{
Calculate ORA score
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
