% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{CalculatePairwiseDiff}
\alias{CalculatePairwiseDiff}
\title{Calculate Pairwise Differences}
\usage{
CalculatePairwiseDiff(mat)
}
\description{
mat are log normalized, diff will be ratio
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
