% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{CalculateQeaScore}
\alias{CalculateQeaScore}
\title{Calculate quantitative enrichment score}
\usage{
CalculateQeaScore(mSetObj=NA, nodeImp, method)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{nodeImp}{Indicate the pathway topology analysis, "rbc" for relative-betweeness centrality, 
and "dgr" for out-degree centrality.}

\item{method}{Indicate the pathway enrichment analysis, global test is "gt" and global ancova is "ga".}
}
\description{
Calculate quantitative enrichment score
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
