% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{calculateSet_doe}
\alias{calculateSet_doe}
\title{Cluster of Peak Picking and Alignment}
\usage{
calculateSet_doe(
  object,
  object_mslevel,
  Set_parameters,
  task = 1,
  BPPARAM = bpparam()
)
}
\arguments{
\item{object}{MSnExp object, the trimmed or the original data.}

\item{object_mslevel}{List, the parsed metabolomics scans produced by PeakPicking_prep.}

\item{Set_parameters}{Matrix, the parameters combination produced automatically according to 
the primary parameters input.}

\item{task}{Numeric, task order for XCMS paramters table to run the peak picking and alignment.}

\item{BPPARAM}{MulticoreParam method, used to set the parallel method. Default is bpparam().}
}
\description{
Cluster of Peak Picking and Alignment
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
