% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_stats.R
\name{CalculateSSP}
\alias{CalculateSSP}
\title{Single sample profiling to compare with}
\usage{
CalculateSSP(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
reference concentrations stored in the library
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
