% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{CentroidCheck}
\alias{CentroidCheck}
\title{CentroidCheck}
\usage{
CentroidCheck(filename)
}
\arguments{
\item{filename}{filename to check}
}
\description{
CentroidCheck function used to check centroid or not
}
\author{
Zhiqiang Pang
}
