% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{CheckContainsBlank}
\alias{CheckContainsBlank}
\title{Check for BLANK samples}
\usage{
CheckContainsBlank(mSetObj=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\value{
Returns 1 if "BLANK" is found in class labels; otherwise returns 0.
}
\description{
CheckContainsBlank is used to check whether BLANK samples
are present in the processed class labels.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
