% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{CheckMetaDataConsistency}
\alias{CheckMetaDataConsistency}
\title{Check if data are ready for meta-analysis}
\usage{
CheckMetaDataConsistency(mSetObj = NA, combat = TRUE)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{combat}{Adjust for batch effects, logical variable: TRUE = adjust for batch effects using an empirical Bayes framework (R package sva), FALSE = 
no batch effect adjustment.}
}
\description{
This function determines if all annotated data are ready for meta-analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
