% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{CheckMetaDataIntegrity}
\alias{CheckMetaDataIntegrity}
\title{Check if data are ready for meta-analysis}
\usage{
CheckMetaDataIntegrity()
}
\description{
This function determines if all annotated data are ready for meta-analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
