% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{CleanData}
\alias{CleanData}
\title{Perform data cleaning}
\usage{
CleanData(bdata, removeNA = T, removeNeg = T, removeConst = T)
}
\description{
Cleans data and removes -Inf, Inf, NA, negative and 0
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
