% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{CleanDataMatrix}
\alias{CleanDataMatrix}
\title{Clean the data matrix}
\usage{
CleanDataMatrix(ndata)
}
\arguments{
\item{ndata}{Input the data to be cleaned}
}
\description{
Function used in higher functinos to clean data matrix
}
