% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{ClearStrings}
\alias{ClearStrings}
\title{Remove spaces}
\usage{
ClearStrings(query)
}
\arguments{
\item{query}{Input the query to clear}
}
\description{
Remove from, within, leading and trailing spaces
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
