% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{ClearUserDir}
\alias{ClearUserDir}
\title{Clear folder and memory}
\usage{
ClearUserDir(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Clear the current folder and objects in memory
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
