% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_lipomics.R
\name{computeConc}
\alias{computeConc}
\title{Lipid analysis}
\usage{
computeConc(X, iso = "y")
}
\description{
The upper limit for each combination is considered to be
the minimal of the fatty acid concentration (nmol fatty acid/gram of sample)
X is the lopomics data obtained above
the result is the saved as separate files for each lipid class
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
