% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{ComputeHighLow}
\alias{ComputeHighLow}
\title{Compute the 95% interval for threshold ROC}
\usage{
Utility function, called upon by higher functions
}
\description{
Only for y-axis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
