% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{ComputeMummichogPermPvals}
\alias{ComputeMummichogPermPvals}
\title{Calculate p-values for each Lperm}
\usage{
ComputeMummichogPermPvals(input_cpdlist, total_matched_cpds, pathways,
  matches.res, input_mzlist, cpd2mz_dict)
}
\description{
Calculate p-values for each Lperm
}
