% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{ComputeMummichogSigPvals}
\alias{ComputeMummichogSigPvals}
\title{Calculate p-values for the list of significant compounds}
\usage{
ComputeMummichogSigPvals(mSetObj=NA, enrichOpt, pvalOpt)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{enrichOpt}{Input method for enrichment analysis}

\item{pvalOpt}{Calculate fisher or hypergeometric p-values}
}
\description{
This function performs pathway enrichment by calculate p-values for each pathway in the reference
organism from the list of significant compounds.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
