% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_graphics.R
\name{concplot}
\alias{concplot}
\title{Utility function}
\usage{
concplot(mn, lower, upper, labels = NULL, xlab = "Odds ratio",
  ylab = "Study Reference", xlim = NULL, line.col = "blue",
  text.col = "forestgreen", xaxt = "s", ...)
}
\description{
Utility function
}
