% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{ContainMissing}
\alias{ContainMissing}
\title{Check for missing data}
\usage{
ContainMissing(mSetObj=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
ContainMissing is used to check if any missing data exists in the uploaded file.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
