% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{Convert2AnalObject}
\alias{Convert2AnalObject}
\title{Convert2AnalObject
This function is used to convert mSet object from raw spectra processing for the following analysis.}
\usage{
Convert2AnalObject(mSet, data.type, anal.type, paired = FALSE)
}
\arguments{
\item{mSet}{mSet from raw spectral processing pipeline, OptiLCMS}

\item{data.type}{data type, should be 'spec'}

\item{anal.type}{analysis type, should be 'raw'}

\item{paired}{data is paired or not, use FALSE by default}
}
\description{
Convert2AnalObject
This function is used to convert mSet object from raw spectra processing for the following analysis.
}
