% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{Convert2Mummichog}
\alias{Convert2Mummichog}
\title{Convert mSetObj to proper format for MS Peaks to Pathways module}
\usage{
Convert2Mummichog(mSetObj = NA, rt = FALSE)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}
}
\description{
Following t-test analysis, this functions converts the results from the mSetObj 
to the proper format for mummichog analysis
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
