% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_covariate.R
\name{CovariateScatter.Anal}
\alias{CovariateScatter.Anal}
\title{CovariateScatter.Anal}
\usage{
CovariateScatter.Anal(
  mSetObj,
  imgName = "NA",
  format = "png",
  analysis.var,
  ref = NULL,
  block = "NA",
  thresh = 0.05,
  pval.type = "fdr",
  contrast.cls = "anova"
)
}
\arguments{
\item{mSetObj}{mSetObj object}

\item{imgName}{image name}

\item{format}{image format}

\item{analysis.var}{variable of analysis}

\item{ref}{reference group}

\item{block}{block name}

\item{thresh}{threshold}

\item{pval.type}{pvalue type (raw or fdr)}

\item{contrast.cls}{contrast group}
}
\description{
CovariateScatter.Anal
}
