% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_raw_spectra.R,
%   R/sweave_report_stats.R
\name{CreateAnalNullMsg}
\alias{CreateAnalNullMsg}
\title{Create null message for analysis
Creates a message for the Sweave report}
\usage{
CreateAnalNullMsg()

CreateAnalNullMsg()
}
\description{
Creates a message stating that no analyses were performed on your data.

Creates a message stating that no analyses were performed on your data.
}
