% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateANOVAdoc}
\alias{CreateANOVAdoc}
\title{Create report of analyses}
\usage{
CreateANOVAdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Create ANOVA document
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
