% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateAOV2doc}
\alias{CreateAOV2doc}
\title{Create report of analyses}
\usage{
CreateAOV2doc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
ANOVA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
