% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateASCAdoc}
\alias{CreateASCAdoc}
\title{Create report of analyses}
\usage{
CreateASCAdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Random Forest ASCA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
