% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateBiomarkerIntr}
\alias{CreateBiomarkerIntr}
\title{Create biomarker analysis report: Introduction}
\usage{
CreateBiomarkerIntr()
}
\description{
Report generation using Sweave
Biomarker analysis report introduction
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
