% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateCorAnalysis}
\alias{CreateCorAnalysis}
\title{Create report of analyses}
\usage{
CreateCorAnalysis(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Correlation and Partial Correlation Analysis
}
\author{
Jessica Ewald \email{jessica.ewald@mail.mcgill.ca}
McGill University, Canada
License: MIT License
}
