% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateCorHeatmap}
\alias{CreateCorHeatmap}
\title{Create report of analyses}
\usage{
CreateCorHeatmap(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
For Correlation Heatmap
}
\author{
Jessica Ewald \email{jessica.ewald@mail.mcgill.ca}
McGill University, Canada
License: MIT License
}
