% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateCovAdj}
\alias{CreateCovAdj}
\title{Create report of analyses}
\usage{
CreateCovAdj(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Covariate Adjustment
}
\author{
Jessica Ewald \email{jessica.ewald@mail.mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
