% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateEBAMdoc}
\alias{CreateEBAMdoc}
\title{Create report of analyses}
\usage{
CreateEBAMdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Create EBAM document
Note: the search for delta (SAM) and a0 (EBAM) will not be plotted
it is only exploration, and may cause potential inconsistentcies.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
