% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_enrichment.R
\name{CreateEnrichAnalDoc}
\alias{CreateEnrichAnalDoc}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateEnrichAnalDoc()
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report, analysis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
