% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_enrichment.R
\name{CreateEnrichIntr}
\alias{CreateEnrichIntr}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateEnrichIntr()
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report introduction
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
