% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_enrichment.R
\name{CreateEnrichORAdoc}
\alias{CreateEnrichORAdoc}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateEnrichORAdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report, over
representation analysis (ORA)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
