% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_enrichment.R
\name{CreateEnrichQEAdoc}
\alias{CreateEnrichQEAdoc}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateEnrichQEAdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report
Quantitative enrichment analysis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
