% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_enrichment.R
\name{CreateEnrichSSPdoc}
\alias{CreateEnrichSSPdoc}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateEnrichSSPdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report
Single sampling profiling
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
