% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateFooter}
\alias{CreateFooter}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateFooter()
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report footer
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
