% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{CreateGraph}
\alias{CreateGraph}
\title{Create igraph from the edgelist saved from graph DB and decompose into subnets}
\usage{
CreateGraph(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Function for the network explorer module, prepares user's data for network exploration.
}
