% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateHeader}
\alias{CreateHeader}
\title{Create report of analyses}
\usage{
CreateHeader(usrName)
}
\description{
Report generation using Sweave
Create header
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
