% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateHeatmap2doc}
\alias{CreateHeatmap2doc}
\title{Create report of analyses}
\usage{
CreateHeatmap2doc(mSetObj = NA)
}
\description{
Report generation using Sweave
2-way heatmap
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
