% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{CreateIntegMatchingTable}
\alias{CreateIntegMatchingTable}
\title{CreateIntegMatchingTable}
\usage{
CreateIntegMatchingTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{mSetObj Object}
}
\description{
CreateIntegMatchingTable
}
