% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegPathwayAnalysisRnwReport}
\alias{CreateIntegPathwayAnalysisRnwReport}
\title{Create report of analyses (IntegPathwayAnalysis)}
\usage{
CreateIntegPathwayAnalysisRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Puts together the analysis report
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
