% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegratedPathwayAnalInputDoc}
\alias{CreateIntegratedPathwayAnalInputDoc}
\title{Create integrated pathway  report: Data Input}
\usage{
CreateIntegratedPathwayAnalInputDoc(mSetObj = NA)
}
\description{
Report generation using Sweave
integrated pathway report, data input documentation.
}
\author{
Jasmine Chong
McGill University, viewingCanada
License: GNU GPL (>= 2)
}
