% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegratedPathwayAnalIntr}
\alias{CreateIntegratedPathwayAnalIntr}
\title{Create integrated pathway analysis report: Introduction}
\usage{
CreateIntegratedPathwayAnalIntr()
}
\description{
Report generation using Sweave
Integrated pathwayr analysis report introduction
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
