% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegratedPathwayAnalOverview}
\alias{CreateIntegratedPathwayAnalOverview}
\title{Create integrated pathway  report: Overview}
\usage{
CreateIntegratedPathwayAnalOverview()
}
\description{
Report generation using Sweave
integrated pathway analysis report overview
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
