% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegratedPathwayDoc}
\alias{CreateIntegratedPathwayDoc}
\title{Create integrated pathway analysis report}
\usage{
CreateIntegratedPathwayDoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Biomarker analysis report, ROC Curve Based Model Creation and Evaluation
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
