% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegratedPathwayGeneMapTable}
\alias{CreateIntegratedPathwayGeneMapTable}
\title{Create a x-table for gene name mapping}
\usage{
CreateIntegratedPathwayGeneMapTable(mSetObj = NA)
}
\description{
Report generation using Sweave
Function to create a table for gene name mapping
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
