% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegratedPathwayNameMapTable}
\alias{CreateIntegratedPathwayNameMapTable}
\title{Create a x-table for compound name mapping}
\usage{
CreateIntegratedPathwayNameMapTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Function to create a table for compound name mapping
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
