% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_integmex.R
\name{CreateIntegratedPathwayResultsTable}
\alias{CreateIntegratedPathwayResultsTable}
\title{Create a x-table for pathway results}
\usage{
CreateIntegratedPathwayResultsTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Function to create a table for pathway results
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
