% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{CreateLadder}
\alias{CreateLadder}
\title{R-code for R-SVM}
\usage{
CreateLadder(Ntotal, Nmin = 5)
}
\arguments{
\item{Ntotal}{Total number}

\item{Nmin}{Minimum number, default set to 5}
}
\description{
use leave-one-out / Nfold or bootstrape to permute data for external CV
build SVM model and use mean-balanced weight to sort genes on training set
and recursive elimination of least important genes
}
\author{
Dr. Xin Lu, Research Scientist
Biostatistics Department, Harvard School of Public Health
create a decreasing ladder for recursive feature elimination
}
