% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{CreateMappingResultTable}
\alias{CreateMappingResultTable}
\title{Creates the mapping result table}
\usage{
CreateMappingResultTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Creates the mapping result table
}
