% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateMBdoc}
\alias{CreateMBdoc}
\title{Create report of analyses}
\usage{
CreateMBdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Multivariate Bayes
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
