% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_metamummi.R,
%   R/sweave_report_mummichog.R
\name{CreateMetaAnalTable}
\alias{CreateMetaAnalTable}
\title{Create Mummichog report of analyses}
\usage{
CreateMetaAnalTable(mSetObj = NA)

CreateMetaAnalTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Function to create a summary table of mummichog analysis

Report generation using Sweave
Function to create a summary table of mummichog analysis
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
