% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateMetaAnalysisInputDoc}
\alias{CreateMetaAnalysisInputDoc}
\title{Create MetaAnalysis analysis report: Data Input}
\usage{
CreateMetaAnalysisInputDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Power analysis report, data input documentation.
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
