% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateMetaAnalysisIntr}
\alias{CreateMetaAnalysisIntr}
\title{Create MetaAnalysis analysis report: Introduction}
\usage{
CreateMetaAnalysisIntr()
}
\description{
Report generation using Sweave
MetaAnalysis analysis report introduction
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
