% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateMetaAnalysisNORMdoc}
\alias{CreateMetaAnalysisNORMdoc}
\title{Create MetaAnalysis analysis report: Data Normalization}
\usage{
CreateMetaAnalysisNORMdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Meta-Analysis, data normalization documentation.
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
