% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateMetaAnalysisOutput}
\alias{CreateMetaAnalysisOutput}
\title{Create MetaAnalysis analysis report: Data Normalization}
\usage{
CreateMetaAnalysisOutput(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
MetaAnalysis analysis, data normalization documentation.
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
