% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateMetaAnalysisRnwReport}
\alias{CreateMetaAnalysisRnwReport}
\title{Create report of analyses (Meta-Analysis)}
\usage{
CreateMetaAnalysisRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Puts together the analysis report
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
