% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_metamummi.R
\name{CreateMetaboNetDoc}
\alias{CreateMetaboNetDoc}
\title{Create KEGG Global Metabolic Network section for MetaPath report}
\usage{
CreateMetaboNetDoc(mSetObj)
}
\arguments{
\item{mSetObj}{mSetObj}
}
\description{
Report generation using Sweave
}
