% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_metamummi.R
\name{CreateMetaMummichogInputDoc}
\alias{CreateMetaMummichogInputDoc}
\title{Create analysis report: Functional Meta-Analysis Data Input}
\usage{
CreateMetaMummichogInputDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Mummichog analysis report, data input documentation.
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
