% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_metamummi.R
\name{CreateMetaMummichogIntro}
\alias{CreateMetaMummichogIntro}
\title{Create analysis report: Functional Meta-Analysis Introduction}
\usage{
CreateMetaMummichogIntro()
}
\description{
Report generation using Sweave
Mummichog analysis report introduction
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
