% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_metamummi.R
\name{CreateMetaMummichogResults}
\alias{CreateMetaMummichogResults}
\title{Create analysis report: Functional Meta-Analysis Results}
\usage{
CreateMetaMummichogResults(mSetObj)
}
\arguments{
\item{mSetObj}{mSetObj}
}
\description{
Report generation using Sweave
Mummichog analysis report overview
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
