% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateModelBiomarkersDoc}
\alias{CreateModelBiomarkersDoc}
\title{Create biomarker analysis report: ROC Curve Based Model Creation and Evaluation}
\usage{
CreateModelBiomarkersDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Biomarker analysis report, ROC Curve Based Model Creation and Evaluation
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
