% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateMultiBiomarkersDoc}
\alias{CreateMultiBiomarkersDoc}
\title{Create biomarker analysis report: Multivariate Biomarker Analysis}
\usage{
CreateMultiBiomarkersDoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Biomarker analysis report, Multivariate Biomarker Analysis
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
