% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_mummichog.R
\name{CreateMummichogAnalTable}
\alias{CreateMummichogAnalTable}
\title{Create Mummichog report of analyses}
\usage{
CreateMummichogAnalTable(mSetObj = NA)
}
\description{
Report generation using Sweave
Function to create a summary table of mummichog analysis
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
