% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_mummichog.R
\name{CreateMummichogInputDoc}
\alias{CreateMummichogInputDoc}
\title{Create Mummichog analysis report: Data Input}
\usage{
CreateMummichogInputDoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Mummichog analysis report, data input documentation.
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
