% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_mummichog.R
\name{CreateMummichogIntro}
\alias{CreateMummichogIntro}
\title{Create mummichog analysis report: Introduction}
\usage{
CreateMummichogIntro()
}
\description{
Report generation using Sweave
Mummichog analysis report introduction
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
