% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_metamummi.R
\name{CreateMummichogMetaAnalPathTable}
\alias{CreateMummichogMetaAnalPathTable}
\title{Create analysis report: Functional Meta-Analysis Results Table}
\usage{
CreateMummichogMetaAnalPathTable(mSetObj)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Function to create a summary table of mummichog analysis
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
