% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_mummichog.R
\name{CreateMummichogMetaAnalReport}
\alias{CreateMummichogMetaAnalReport}
\title{Create analysis report: Functional Meta-Analysis}
\usage{
CreateMummichogMetaAnalReport(mSetObj)
}
\description{
Report generation using Sweave
Functional Meta-Analysis Report
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
