% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_mummichog.R
\name{CreateMummichogRnwReport}
\alias{CreateMummichogRnwReport}
\title{Create report of analyses (Biomarker)}
\usage{
CreateMummichogRnwReport(mSetObj, usrName)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{usrName}{Input the name of the user}
}
\description{
Report generation using Sweave
Puts together the analysis report
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
