% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_network.R
\name{CreateNetworkExplorerInputDoc}
\alias{CreateNetworkExplorerInputDoc}
\title{Create network explorer: Data Input}
\usage{
CreateNetworkExplorerInputDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
network explorer report, data input documentation.
}
\author{
Jasmine Chong
McGill University, viewingCanada
License: GNU GPL (>= 2)
}
