% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_network.R
\name{CreateNetworkExplorerIntr}
\alias{CreateNetworkExplorerIntr}
\title{Create integrated pathway analysis report: Introduction}
\usage{
CreateNetworkExplorerIntr()
}
\description{
Report generation using Sweave
Network explorer report introduction
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
