% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_network.R
\name{CreateNetworkGeneMapTable}
\alias{CreateNetworkGeneMapTable}
\title{Create a x-table for gene name mapping}
\usage{
CreateNetworkGeneMapTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Function to create a table for gene name mapping
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
