% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_network.R
\name{CreateNetworkNameMapTable}
\alias{CreateNetworkNameMapTable}
\title{Create a x-table for compound name mapping}
\usage{
CreateNetworkNameMapTable(mSetObj = NA)
}
\description{
Report generation using Sweave
Function to create a table for compound name mapping
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
