% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_pathway.R
\name{CreatePathInputDoc}
\alias{CreatePathInputDoc}
\title{Create report of analyses (Met Pathway)}
\usage{
CreatePathInputDoc()
}
\description{
Report generation using Sweave
Metabolomic pathway analysis
Create data input doc
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
