% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_pathway.R
\name{CreatePathProcessDoc}
\alias{CreatePathProcessDoc}
\title{Create report of analyses (Met Pathway)}
\usage{
CreatePathProcessDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Metabolomic pathway analysis
Create MetPA process
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
