% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreatePathResultDoc}
\alias{CreatePathResultDoc}
\title{Create report of analyses (Met Pathway)}
\usage{
CreatePathResultDoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Metabolomic pathway analysis
Create MetPA results doc
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
