% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreatePathRnwReport}
\alias{CreatePathRnwReport}
\title{Create report of analyses (Met Pathway)}
\usage{
CreatePathRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Metabolomic pathway analysis
write .Rnw file template
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
