% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreatePDFReport}
\alias{CreatePDFReport}
\title{Create report of analyses}
\usage{
CreatePDFReport(mSetObj = NA, usrName)
}
\description{
Report generation using Sweave
Note: most analyses were already performed, only need to embed
the results to the right place without rerunning the whole analysis
through Sweave. Only some auxilliary info (i.e. time, version etc need to
run in R through Sweave
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
