% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_power.R
\name{CreatePowerInputDoc}
\alias{CreatePowerInputDoc}
\title{Create power analysis report: Data Input}
\usage{
CreatePowerInputDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Power analysis report, data input documentation.
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
