% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_power.R
\name{CreatePowerIntr}
\alias{CreatePowerIntr}
\title{Create power analysis report: Introduction}
\usage{
CreatePowerIntr()
}
\description{
Report generation using Sweave
Power analysis report introduction
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
