% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_power.R
\name{CreatePowerOverview}
\alias{CreatePowerOverview}
\title{Create power analysis report: Overview}
\usage{
CreatePowerOverview()
}
\description{
Report generation using Sweave
Power analysis report overview
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
