% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_power.R
\name{CreatePowerParametersDoc}
\alias{CreatePowerParametersDoc}
\title{Create power analysis report: Power Parameter Selection}
\usage{
CreatePowerParametersDoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Power analysis report, parameter selection
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
