% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_power.R
\name{CreatePowerRnwReport}
\alias{CreatePowerRnwReport}
\title{Create report of analyses (Power)}
\usage{
CreatePowerRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Put together the analysis report
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
write .Rnw file template
}
