% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateRandomForest}
\alias{CreateRandomForest}
\title{Create report of analyses}
\usage{
CreateRandomForest(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Random Forest
}
\author{
Jessica Ewald \email{jessica.ewald@mail.mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
