% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{CreateRawRscript}
\alias{CreateRawRscript}
\title{CreateRawRscript}
\usage{
CreateRawRscript(guestName, planString, planString2, rawfilenms.vec)
}
\description{
used to create a running for raw spectra processing
this file will be run by spring daemon by using OptiLCMS rather than relay on the local 
compiling RC files
}
\author{
Guangyan Zhou, Zhiqiang Pang
}
