% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateRHistAppendix}
\alias{CreateRHistAppendix}
\title{Create report of analyses}
\usage{
CreateRHistAppendix()
}
\description{
Report generation using Sweave
Create footer
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
