% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateROCLabelsTable}
\alias{CreateROCLabelsTable}
\title{Create a x-table for newly classified samples}
\usage{
CreateROCLabelsTable(mSetObj = NA)
}
\description{
Report generation using Sweave
Function to create a table for newly classified samples
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
