% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{CreateSemiTransColors}
\alias{CreateSemiTransColors}
\title{Create semitransparant colors}
\usage{
CreateSemiTransColors(cls)

CreateSemiTransColors(cls)
}
\arguments{
\item{cls}{Input class labels}
}
\description{
Create semitransparant colors for a given class label

Create semitransparant colors for a given class label
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
