% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_raw_spectra.R
\name{CreateSpectraIOdoc}
\alias{CreateSpectraIOdoc}
\title{Create report of analyses}
\usage{
CreateSpectraIOdoc(mSetObj = NULL)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Read and process raw data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
