% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateStatFooter}
\alias{CreateStatFooter}
\title{Create report of analyses}
\usage{
CreateStatFooter()
}
\description{
Report generation using Sweave
Create footer
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
