% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_stats.R
\name{CreateStatIntr}
\alias{CreateStatIntr}
\title{Create report of analyses}
\usage{
CreateStatIntr()
}
\description{
Report generation using Sweave
Create header
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
