% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateStatRnwReport}
\alias{CreateStatRnwReport}
\title{Create report of analyses}
\usage{
CreateStatRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Write .Rnw file template
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
