% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateTimeSeriesAnalNullMsg}
\alias{CreateTimeSeriesAnalNullMsg}
\title{Create null analysis message for time-series sweave report}
\usage{
CreateTimeSeriesAnalNullMsg()
}
\description{
Creates empty time-series analysis message
}
