% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateTimeSeriesIOdoc}
\alias{CreateTimeSeriesIOdoc}
\title{Create report of analyses (Met Pathway)}
\usage{
CreateTimeSeriesIOdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Metabolomic pathway analysis, time-series
Read and process the raw data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
