% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateTimeSeriesRnwReport}
\alias{CreateTimeSeriesRnwReport}
\title{Create report of analyses (Met Pathway)}
\usage{
CreateTimeSeriesRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Metabolomic pathway analysis
Create timeseries .Rnw file template
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
