% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateUnivarBiomarkersDoc}
\alias{CreateUnivarBiomarkersDoc}
\title{Create power analysis report: Biomarker Univariate Analysis}
\usage{
CreateUnivarBiomarkersDoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Biomarker analysis report, Univariate Analysis
}
\author{
Jasmine Chong 
McGill University, Canada
License: GNU GPL (>= 2)
}
