% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateVennMetaTable}
\alias{CreateVennMetaTable}
\title{Create MetaAnalysis table of results for Venn Diagram}
\usage{
CreateVennMetaTable(mSetObj = NA)
}
\description{
Report generation using Sweave
Function to create a table containing meta-analysis results.
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
