% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{creatPeakTable}
\alias{creatPeakTable}
\title{creatPeakTable}
\usage{
creatPeakTable(xset)
}
\description{
creatPeakTable
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
