% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{CrossReferencing}
\alias{CrossReferencing}
\title{Various functions for mapping b/w names & database identifiers
Given a list of compound names or ids, find matched name or ids from selected databases}
\usage{
CrossReferencing(mSetObj = NA, q.type, hmdb = T, pubchem = T, chebi = F,
  kegg = T, metlin = F)
}
\description{
Given a list of compound names or ids
find matched name or ids from selected databases
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
