% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\docType{data}
\name{currency}
\alias{currency}
\title{Global variables define currency compounds}
\format{An object of class \code{character} of length 26.}
\usage{
currency
}
\description{
Global variables define currency compounds
}
\keyword{datasets}
