% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{CVTest.LRmodel}
\alias{CVTest.LRmodel}
\title{Calculate ROC performance with CV}
\usage{
CVTest.LRmodel(data.in, fmla.in, kfold = 10, run.stepwise = FALSE)
}
\arguments{
\item{data.in}{Input matrix of data}

\item{fmla.in}{Input for generalized linear model}

\item{kfold}{Numeric}

\item{run.stepwise}{Logical}
}
\description{
Calculate ROC performance with CV
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
