% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{Densitygrouping_slave}
\alias{Densitygrouping_slave}
\title{Densitygrouping_slave}
\usage{
Densitygrouping_slave(
  x,
  bw,
  densFrom,
  densTo,
  densN,
  sampleGroups,
  sampleGroupTable,
  minFraction,
  minSamples,
  maxFeatures
)
}
\description{
Densitygrouping_slave
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
