% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_peaks.R
\name{descendMin}
\alias{descendMin}
\title{Perform utilities for peak grouping}
\usage{
descendMin(y, istart = which.max(y))
}
\arguments{
\item{y}{Input peaks}

\item{istart}{Performs which.max on y}
}
\description{
Perform various utilities for peak grouping
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
