% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{doCompoundMapping}
\alias{doCompoundMapping}
\title{Perform compound mapping}
\usage{
doCompoundMapping(cmpd.vec, q.type)
}
\arguments{
\item{cmpd.vec}{Input compound vector}

\item{q.type}{Query type}
}
\description{
Perform compound mapping
}
