% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{doEmblProtein2EntrezMapping}
\alias{doEmblProtein2EntrezMapping}
\title{Utility function for PerformNetEnrichment}
\usage{
doEmblProtein2EntrezMapping(emblprotein.vec)
}
\description{
Utility function for PerformNetEnrichment
}
