% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{doGeneIDMapping}
\alias{doGeneIDMapping}
\title{Perform various annotation}
\usage{
doGeneIDMapping(q.vec, org, type)
}
\description{
Gene ID mapping, gene annotation, compound
mapping, KEGG mapping
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
