% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R
\name{doKEGG2NameMapping}
\alias{doKEGG2NameMapping}
\title{Perform KEGG to compound name mapping}
\usage{
doKEGG2NameMapping(kegg.vec)
}
\arguments{
\item{kegg.vec}{Input vector of KEGG compounds}
}
\description{
Perform KEGG to compound name mapping
}
