% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R, R/networks_data.R
\name{doKOFiltering}
\alias{doKOFiltering}
\title{Utility function}
\usage{
doKOFiltering(ko.vec, type)

doKOFiltering(ko.vec, type)
}
\arguments{
\item{ko.vec}{Input the vector containing KOs}

\item{type}{Input the type}
}
\description{
Returns matched KO in the same order (NA if no match)

Returns matched KO in the same order (NA if no match)
}
