% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{doLogisticRegMdl}
\alias{doLogisticRegMdl}
\title{Develop a Logistic Regression Model with all of the combined k-fold CV subsets}
\usage{
doLogisticRegMdl(mSetObj, x.train, y.train, x.test, y.test)
}
\arguments{
\item{x.train}{Input the X training set}

\item{y.train}{Input the Y training set}

\item{x.test}{Input the X test set}

\item{y.test}{Input the Y test set}
}
\description{
Develop a Logistic Regression Model with all of the combined k-fold CV subsets
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
