% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{.compute.mummichog.RT.fgsea}
\alias{.compute.mummichog.RT.fgsea}
\title{Internal function for calculating GSEA, with RT}
\usage{
.compute.mummichog.RT.fgsea(mSetObj, permNum)

.compute.mummichog.RT.fgsea(mSetObj, permNum)
}
\description{
Internal function for calculating GSEA, with RT

Internal function for calculating GSEA, with RT
}
