% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming.R
\name{.emptyscan.remove}
\alias{.emptyscan.remove}
\title{Function for 'Empty scan' removal}
\usage{
.emptyscan.remove(raw_data, ms_list)
}
\description{
Function for 'Empty scan' removal (internal use only)
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}
}
