% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_lib_utils.R
\name{.get.my.lib}
\alias{.get.my.lib}
\title{Read RDS files from the internet}
\usage{
.get.my.lib(filenm, sub.dir = NULL)
}
\arguments{
\item{filenm}{Input the name of the file to download}
}
\description{
Function downloads the required file and reads it only if not already in working directory.
Need to specify the file URL and the destfile.
}
