% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{.init.Permutations}
\alias{.init.Permutations}
\title{Internal function to perform PSEA, no retention time}
\usage{
.init.Permutations(mSetObj, permNum)

.init.Permutations(mSetObj, permNum)
}
\description{
Internal function to perform PSEA, no retention time

Internal function to perform PSEA, no retention time
}
