% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{.init.RT.Permutations}
\alias{.init.RT.Permutations}
\title{Internal function to perform PSEA, with RT}
\usage{
.init.RT.Permutations(mSetObj, permNum)

.init.RT.Permutations(mSetObj, permNum)
}
\description{
Internal function to perform PSEA, with RT

Internal function to perform PSEA, with RT
}
