% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{.readDataTable}
\alias{.readDataTable}
\title{Read data table}
\usage{
.readDataTable(fileName)
}
\arguments{
\item{fileName}{Input filename}
}
\description{
Read in data table
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
