% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{EBAM.A0.Init}
\alias{EBAM.A0.Init}
\title{For EBAM analysis}
\usage{
EBAM.A0.Init(mSetObj = NA, isPaired, isVarEq)
}
\description{
deteriming a0, only applicable for z.ebam (default)
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
