% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{EBAM.Cmpd.Init}
\alias{EBAM.Cmpd.Init}
\title{For EBAM analysis}
\usage{
EBAM.Cmpd.Init(mSetObj = NA, method = "z.ebam", A0 = 0,
  isPaired = FALSE, isVarEq = TRUE)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
note: if method is wilcoxon, the A0 and var equal will be ignored
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
