% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{EBAM.Init}
\alias{EBAM.Init}
\title{For EBAM analysis}
\usage{
EBAM.Init(
  mSetObj = NA,
  isPaired,
  isVarEq,
  nonPar,
  A0 = -99,
  delta,
  imgA0,
  imgSig
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{isPaired}{Logical}

\item{isVarEq}{Logical}
}
\description{
deteriming a0, only applicable for z.ebam (default)
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
