% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{ExportMissingHeatmapJSON}
\alias{ExportMissingHeatmapJSON}
\title{Export Missing Value Heatmap to JSON for Plotly.js (v3)}
\usage{
ExportMissingHeatmapJSON(mSetObj = NA,
                           fileName = "missing_heatmap.json")
}
\arguments{
\item{mSetObj}{MetaboAnalyst object.}

\item{fileName}{Output JSON file name.}
}
\description{
Prepares a JSON file with the missingness matrix (0 = present, 1 = missing) formatted for Plotly.js v3 heatmap.
}
