% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{ExtractMS2data}
\alias{ExtractMS2data}
\title{Extract MS2 Data}
\usage{
ExtractMS2data(filename, peakParams, mzmin, mzmax)
}
\arguments{
\item{filename}{Name of the file (e.g. mzML, mzXML)}

\item{peakParams}{Object containing parameters for peak picking.}

\item{mzmin}{Minimum m/z when selecting a precursor from peak list}

\item{mzmax}{Maximum m/z when selecting a precursor from peak list}
}
\description{
This function returns a list of spectra that matches with 
a user specified precursor m/z.
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
