% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{FastRunningShow_auto}
\alias{FastRunningShow_auto}
\title{FastRunningShow_auto}
\usage{
FastRunningShow_auto(fullUserPath)
}
\description{
FastRunningShow_auto is used to showcase the AUTO running pipeline
}
\author{
Zhiqiang Pang
}
