% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{FastRunningShow_customized}
\alias{FastRunningShow_customized}
\title{FastRunningShow_customized}
\usage{
FastRunningShow_customized(fullUserPath)
}
\description{
FastRunningShow_customized is used to showcase the customized running pipeline
}
\author{
Zhiqiang Pang
}
