% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{FC.Anal.paired}
\alias{FC.Anal.paired}
\title{Fold change analysis, paired}
\usage{
FC.Anal.paired(mSetObj = NA, fc.thresh = 2, percent.thresh = 0.75,
  cmp.type = 0)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{fc.thresh}{Fold-change threshold, numeric input}

\item{percent.thresh}{Numeric input, from 0 to 1 to indicate the significant count threshold}

\item{cmp.type}{Comparison type, 0 for group 1 minus group 2, and 1 for group}
}
\description{
Perform paired fold change analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
