% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{FeatureCorrelation}
\alias{FeatureCorrelation}
\title{Pattern hunter}
\usage{
FeatureCorrelation(mSetObj = NA, dist.name, varName)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{dist.name}{Input the name of the distance measure}

\item{varName}{Input the variable name}
}
\description{
Calculate correlation of all other feature to a given feature name
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
