% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{fgsea2}
\alias{fgsea2}
\title{Pre-ranked gsea adapted for untargeted metabolomics}
\usage{
fgsea2(
  mSetObj,
  pathways,
  stats,
  ranks,
  nperm,
  minSize = 1,
  maxSize = Inf,
  nproc = 0,
  gseaParam = 1,
  BPPARAM = NULL
)
}
\description{
Pre-ranked gsea adapted for untargeted metabolomics
}
