% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{FilterVariable}
\alias{FilterVariable}
\title{Methods for non-specific filtering of variables}
\usage{
FilterVariable(mSetObj=NA, filter, qcFilter, rsd)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{qc.filter}{Filter the variables based on QC samples - True (T), or use non-QC based filtering - False (F).}

\item{rsd}{Define the relative standard deviation cut-off. Variables with a RSD greater than this number
will be removed from the dataset. It is only necessary to specify this argument if qc.filter is True (T). 
Otherwise, it will not be used in the function.}

\item{var.filter}{Select the filter option, "rsd" which is the relative standard deviation, "nrsd" which
is the non-parametric relative standard deviation, "mean" which is the mean, "sd" which is the standard
deviation, "mad" which is the median absolute deviation, or "iqr" which is the interquantile range.}

\item{var.cutoff}{var.cutoff value}

\item{int.filter}{int.filter value}

\item{int.cutoff}{int.cutoff value, numeric}

\item{blank.subtraction}{blank.subtraction boolean value}

\item{blank.threshold}{blank.threshold value}

\item{filter.cutoff}{percent to be filtered, for example, 5 (5\%)}
}
\description{
This is a function that filters the dataset, dependent on the user-specified method
for filtering. The function applies a filtering method, ranks the variables within the dataset,
and removes variables based on its rank. The final dataset should contain no more than
than 5000 variables for effective computing.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
