% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{findEqualGreaterM}
\alias{findEqualGreaterM}
\title{Perform utilities for peak grouping}
\usage{
findEqualGreaterM(x, values)
}
\description{
Perform various utilities for peak grouping
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
