% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{FormatPeakList}
\alias{FormatPeakList}
\title{Format Peak List}
\usage{
FormatPeakList(annotPeaks, annParams, filtIso = TRUE,
  filtAdducts = FALSE, missPercent = 0.5)
}
\arguments{
\item{annotPeaks}{The object created using the PerformPeakAnnotation.}

\item{annParams}{The object created using the SetAnnotationParam function, 
containing user's specified or default parameters for downstream 
raw MS data pre-processing.}

\item{filtIso}{Logical, filter out all isotopes except for [M]+ for 
positive ion mode and [M]- for negative ion mode. By default it is 
set to true.}

\item{filtAdducts}{Logical, filter out all adducts except [M+H]+ for
positive ion more and [M-H]- for negative ion mode. By default it is set to false.}

\item{missPercent}{Numeric, specify the threshold to remove features
missing in X\% of samples. For instance, 0.5 specifies to remove features
that are missing from 50\% of all samples per group. Method is only valid
when there are two groups.}
}
\description{
This function formats the CAMERA output to a usable format for MetaboAanlyst.
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
