% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{GeneratePeakList}
\alias{GeneratePeakList}
\title{GeneratePeakList}
\usage{
GeneratePeakList(userPath)
}
\description{
GeneratePeakList is used to generate the peak summary list for result page
}
\author{
Zhiqiang Pang
}
