% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{genLogisticRegMdl}
\alias{genLogisticRegMdl}
\title{Develop a Logistic Regression Model with all of the combined k-fold CV subsets}
\usage{
genLogisticRegMdl(x.train, y.train, x.test, y.test)
}
\description{
Develop a Logistic Regression Model with all of the combined k-fold CV subsets
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
