% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{Get.asca.tss}
\alias{Get.asca.tss}
\title{Function for ASCA permutation}
\usage{
Get.asca.tss(dummy, perm = T)
}
\description{
Dummy is used only for the purpose to maintain lapply API
this is used for permutation on ANOVA paritions,
not on the SCA/PCA part, so the number of selected components is
not applicable in this step
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
