% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{Get.ConcRef}
\alias{Get.ConcRef}
\title{Get the concentration reference}
\usage{
Get.ConcRef(mSetObj = NA, cmpd.nm)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{cmpd.nm}{Input the compound name}
}
\description{
Get the concentration reference
}
