% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{Get.Leverage}
\alias{Get.Leverage}
\title{Fast leverage calculation for permutation purpose}
\usage{
Get.Leverage(XKw, Fac)
}
\description{
note, the leverage combines all components
the importance feature is for the factor not per components
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
