% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Get.rpart.summary}
\alias{Get.rpart.summary}
\title{Get the text description of a recursive partitioning (rpart) result}
\usage{
Get.rpart.summary(x)
}
\arguments{
\item{x}{An Rpart object}
}
\description{
x must be an rpart object
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
