% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{Get.tss}
\alias{Get.tss}
\title{Compute sum of squares (SSQ) for each row of a matrix which may have NA}
\usage{
Get.tss(x, cl)
}
\description{
Columns have labels cl=consecutive integers
note: this is desgined for ASCA parition data
in which Within group (WSS) is zero, so, we only need TSS
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
