% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Get.VIP}
\alias{Get.VIP}
\title{Calculate variable importance of projection (VIP) score for PLS object}
\usage{
Get.VIP(pls.obj, comp = 2)
}
\description{
Users give a pls object ('oscorespls'=T), function calculates VIP score
usually one VIP for each component, return is the average of all VIP
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
