% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetAbundanceLabel}
\alias{GetAbundanceLabel}
\title{Determine value label for plotting}
\usage{
GetAbundanceLabel(data.type)
}
\arguments{
\item{data.type}{Input concentration or intensity data}
}
\description{
Concentration or intensity data type
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
