% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetAccuracyInfo}
\alias{GetAccuracyInfo}
\title{Export biomarker accuracy information}
\usage{
GetAccuracyInfo(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Export biomarker accuracy information
}
