% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{GetAllSOMClusterMembers}
\alias{GetAllSOMClusterMembers}
\title{SOM analysis}
\usage{
GetAllSOMClusterMembers(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Get members for given cluster index, return a character string
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
