% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{GetCircleInfo}
\alias{GetCircleInfo}
\title{Export information about selected circle}
\usage{
GetCircleInfo(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Export information about selected circle
}
