% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetCMD}
\alias{GetCMD}
\title{Retrieve last command from the Rhistory.R file}
\usage{
GetCMD(regexp)
}
\arguments{
\item{regexp}{Retrieve last command from Rhistory file}
}
\description{
Fetches the last command from the Rhistory.R file
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
