% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{getCmpdID}
\alias{getCmpdID}
\title{Perform utilities for MetPa}
\usage{
getCmpdID(dirName)
}
\description{
Get all the KEGG compounds from the pathway databases
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
