% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{GetCompoundDetails}
\alias{GetCompoundDetails}
\title{Function to get adduct details from a specified compound}
\usage{
GetCompoundDetails(mSetObj = NA, cmpd.id)

GetCompoundDetails(mSetObj = NA, cmpd.id)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object.}

\item{cmpd.id}{Input the name of the selected compound.}
}
\description{
Function to get adduct details from a specified compound.
The results will be both printed in the console as well as saved
as a csv file. Note that performing this function multiple times will
overwrite previous queries.

Function to get adduct details from a specified compound.
The results will be both printed in the console as well as saved
as a csv file. Note that performing this function multiple times will
overwrite previous queries.
}
