% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetConvertFullPath}
\alias{GetConvertFullPath}
\title{Perform utilities for cropping images}
\usage{
GetConvertFullPath()
}
\description{
Obtain the full path to convert (from imagemagik)
for cropping images
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
