% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{getDataFromTextArea}
\alias{getDataFromTextArea}
\title{Transform two column text to data matrix}
\usage{
getDataFromTextArea(txtInput, sep.type = "space")
}
\arguments{
\item{txtInput}{Input text}

\item{sep.type}{Indicate the seperator type for input text. Default set to "space"}
}
\description{
Transform two column input text to data matrix (single column data frame)
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
