% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{getDataFromTextInput}
\alias{getDataFromTextInput}
\title{Transform 2 column text to data matrix}
\usage{
getDataFromTextInput(mSetObj = NA, txtInput, sep.type = "space")
}
\description{
Transform two column input text to data matrix (single column data frame)
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
