% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{GetEllipseRadius}
\alias{GetEllipseRadius}
\title{Compute 2D ellipse radius (t-value) for a group
chi-squared: sqrt(qchisq(level, 2))
F-distribution: sqrt(2 * qf(level, 2, n-1))}
\usage{
GetEllipseRadius(level, method, n)
}
\arguments{
\item{level}{confidence level}

\item{method}{"chisq" or "f"}

\item{n}{group sample size}
}
\description{
Compute 2D ellipse radius (t-value) for a group
chi-squared: sqrt(qchisq(level, 2))
F-distribution: sqrt(2 * qf(level, 2, n-1))
}
