% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{GetFinalNameMap}
\alias{GetFinalNameMap}
\title{Return the final (after user selection) map as dataframe}
\usage{
GetFinalNameMap(mSetObj = NA, lipid = FALSE)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{lipid}{Logical}
}
\description{
Returns three columns: original name, HMDB name and KEGG ID,
for enrichment and pathway analysis, respectively
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
