% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetFisherPvalue}
\alias{GetFisherPvalue}
\title{Get fisher p-values}
\usage{
GetFisherPvalue(numSigMembers, numSigAll, numMembers, numAllMembers)
}
\arguments{
\item{numSigMembers}{Number of significant members}

\item{numSigAll}{Number of all significant features}

\item{numMembers}{Number of members}

\item{numAllMembers}{Number of all members}
}
\description{
Get fisher p-values
}
