% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{GetGroupNames}
\alias{GetGroupNames}
\title{Get all group names}
\usage{
GetGroupNames(mSetObj = NA, exp.fac = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Get all group names
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
