% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetImpFeatureMat}
\alias{GetImpFeatureMat}
\title{Get important feature matrix}
\usage{
GetImpFeatureMat(mSetObj = NA, feat.outp, bestFeatNum)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{feat.outp}{Input the list that contains the ranked features in each 
cross validation (CV) and returns a two column matrix, col 1 = median ranking 
and col 2 =  mean importance measure}

\item{bestFeatNum}{Numeric}
}
\description{
feat.outp is a list that contains the ranked features in each 
cross validation (CV) and returns a two column matrix, col 1 = median ranking 
and col 2 =  mean importance measure
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
