% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetLassoFreqs}
\alias{GetLassoFreqs}
\title{Compute lasso frequency}
\usage{
GetLassoFreqs(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Not part of default, need to perform function 
to compute lasso frequency
msg: There are more than 500 variables and n<m
You may wish to restart and set use.Gram=FALSE
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
