% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{GetLimmaResTable}
\alias{GetLimmaResTable}
\title{Get result table from eBayes fit object}
\usage{
GetLimmaResTable(fit.obj)
}
\arguments{
\item{fit.obj}{eBayes fit object to parse to a table}
}
\description{
Get result table from eBayes fit object
}
