% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{GetMapTable}
\alias{GetMapTable}
\title{Get mapping table}
\usage{
GetMapTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Return results from compound name mapping in a table
}
