% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{GetMatchedPathCompoundIDs}
\alias{GetMatchedPathCompoundIDs}
\title{Return matched compound/metabolite IDs for a pathway}
\usage{
GetMatchedPathCompoundIDs(
  mSetObj = NA,
  msetNm,
  id.type = "native",
  uniqueOnly = TRUE
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{msetNm}{Input pathway name or pathway ID}

\item{id.type}{Output ID type: "native", "kegg", or "hmdb"}

\item{uniqueOnly}{Return unique IDs only (default TRUE)}
}
\description{
Return matched IDs (hits) for a selected pathway from pathway analysis results.
}
