% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetMeanROC}
\alias{GetMeanROC}
\title{Compute data points on the ROC curve}
\usage{
GetMeanROC(perf)
}
\arguments{
\item{perf}{Performance object from ROCR}
}
\description{
perf is the performance object from ROCR
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
