% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{GetMetaResultMatrix}
\alias{GetMetaResultMatrix}
\title{Single.type return logFC or p value for individual data analysis}
\usage{
GetMetaResultMatrix(mSetObj = NA, single.type = "fc")
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{single.type}{Default is "fc"}
}
\description{
Single.type return logFC or p value for individual data analysis
}
