% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{GetMetaSigHitsTable}
\alias{GetMetaSigHitsTable}
\title{Export the significant hits from meta-analysis}
\usage{
GetMetaSigHitsTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Export the significant hits from meta-analysis
}
