% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{GetMsetLibCheckMsg}
\alias{GetMsetLibCheckMsg}
\title{Get the library check messages}
\usage{
GetMsetLibCheckMsg(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Get the library check messages
}
