% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{GetMsetLibSearchResult}
\alias{GetMsetLibSearchResult}
\title{Return metset search results}
\usage{
GetMsetLibSearchResult(mSetObj = NA)
}
\description{
since String[][] is not supported, have to return as 1D vector, 
matrix can be directly convert to vector, note default will be column first
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
