% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{GetMsetNames}
\alias{GetMsetNames}
\title{Return the selected metset library to java for display}
\usage{
GetMsetNames(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Return the selected metset library to java for display
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
