% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_covariate.R
\name{GetMultiRFSigMat}
\alias{GetMultiRFSigMat}
\title{Random Forest Significance matrix}
\usage{
GetMultiRFSigMat(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Significance measure, double brackets
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
