% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{GetMummichogMZHits}
\alias{GetMummichogMZHits}
\title{Function to return the unique m/zs from the selected pathways 
based on the compounds}
\usage{
GetMummichogMZHits(mSetObj = NA, msetNm)
}
\description{
Function to return the unique m/zs from the selected pathways 
based on the compounds
}
