% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{GetMummichogPathSetDetails}
\alias{GetMummichogPathSetDetails}
\title{Function to get compound details from a specified pathway}
\usage{
GetMummichogPathSetDetails(mSetObj = NA, msetNm)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object.}

\item{msetNm}{Input the name of the pathway}
}
\description{
Function to get compound details from a specified pathway.
The results will be both printed in the console as well as saved
as a csv file. Note that performing this function multiple times will
overwrite previous queries.
}
