% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{GetMumStatRowNames}
\alias{GetMumStatRowNames}
\title{Plot volcano for mummichog table input}
\usage{
GetMumStatRowNames()
}
\arguments{
\item{mSetObj}{mSet Object}

\item{imgName}{Image file name prefix}

\item{dpi}{Resolution
Get mummichog stat result table data for details view}
}
\description{
Plots logFC vs -log10(p) from the statistical analysis results
}
