% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{getndp}
\alias{getndp}
\title{Perform utilities for MetPa}
\usage{
getndp(x, tol = 2 * .Machine$double.eps)
}
\description{
Count the number of digits in the values
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
