% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{GetNMDRStudy}
\alias{GetNMDRStudy}
\title{Function to retrieve dataset from the Metabolomics Workbench.}
\usage{
GetNMDRStudy(mSetObj=NA, StudyID)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects).}

\item{StudyID}{Input the StudyID of the study from the Metabolomics Workbench. 
Use the ListNMDRStudies function to obtain a list of all available studies
from the Metabolomics Workbench.}
}
\description{
This function uses the httr R package to make an API
call to the Metabolomics Workbench to download and save a dataset
based on the Study ID into the current working directory.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}, Jasmine Chong
McGill University, Canada
License: MIT License
}
