% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{GetORA.pathNames}
\alias{GetORA.pathNames}
\title{Export pathway names from ORA analysis}
\usage{
GetORA.pathNames(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Export pathway names from ORA analysis
}
