% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_stats.R
\name{GetORATable}
\alias{GetORATable}
\title{Get ORA table}
\usage{
GetORATable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Get ORA table
}
