% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{GetPathwayCompoundColors}
\alias{GetPathwayCompoundColors}
\title{Get compound colors for KEGG pathway network viewer}
\usage{
GetPathwayCompoundColors(mSetObj = NA, pathName)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{pathName}{Input the name of the selected pathway}
}
\value{
Vector of strings in format "compoundId|color|pvalue"
}
\description{
Returns matched compounds with their colors based on significance
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
