% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{GetPathwayLibraryCompounds}
\alias{GetPathwayLibraryCompounds}
\title{Get pathway library compounds for KEGG pathway network viewer}
\usage{
GetPathwayLibraryCompounds(mSetObj = NA, pathName)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{pathName}{Input the name or id of the selected pathway}
}
\value{
Character vector of KEGG compound IDs (e.g. C00022)
}
\description{
Returns all compound KEGG IDs in current organism pathway library for the selected pathway.
}
