% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{GetPvalCutoffForSigFeatures}
\alias{GetPvalCutoffForSigFeatures}
\title{Compute p-value cutoff to exactly capture significant features from PerformMumTableStat}
\usage{
GetPvalCutoffForSigFeatures(mSetObj = NA)
}
\arguments{
\item{mSetObj}{mSet Object}
}
\description{
After running PerformMumTableStat, find the p-value threshold that
  captures all features where inx.imp == TRUE, keeping background peaks for permutation analysis
}
