% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetRandomSubsetIndex}
\alias{GetRandomSubsetIndex}
\title{Random subset of numbers}
\usage{
GetRandomSubsetIndex(total, sub.num = 50)
}
\description{
Obtain a random subset of numbers from a total number
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
