% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{GetRCommandHistory}
\alias{GetRCommandHistory}
\title{Export R Command History}
\usage{
GetRCommandHistory(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Export R Command History
}
