% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetRFConf.Table}
\alias{GetRFConf.Table}
\title{Classification performance table for random forest analysis}
\usage{
GetRFConf.Table(mSetObj = NA)
}
\description{
Classification performance table for random forest analysis
}
