% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetRFConfMat}
\alias{GetRFConfMat}
\title{Random Forest Confusion Matrix}
\usage{
GetRFConfMat(mSetObj = NA)
}
\description{
Return double[][] confusion matrix
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
