% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetRFOOB}
\alias{GetRFOOB}
\title{Random Forest OOB}
\usage{
GetRFOOB(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Get the OOB error for the last signif
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
