% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetRFSigMat}
\alias{GetRFSigMat}
\title{Random Forest Significance matrix}
\usage{
GetRFSigMat(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Significance measure, double brackets
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
