% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetROC.coords}
\alias{GetROC.coords}
\title{Return ROC corodinates with confidence intervals}
\usage{
GetROC.coords(mSetObj = NA, fld.nm, val, plot = TRUE, imgNm)
}
\description{
Return ROC corodinates with confidence intervals
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
