% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetROCLassoFreq}
\alias{GetROCLassoFreq}
\title{Get p-values from lasso}
\usage{
GetROCLassoFreq(data, cls)
}
\arguments{
\item{data}{Input data}

\item{cls}{Input class labels}
}
\description{
Get p-values from lasso
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
