% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetROCTtestP}
\alias{GetROCTtestP}
\title{Get p-values for ROC}
\usage{
GetROCTtestP(data, cls)
}
\arguments{
\item{data}{Input data}

\item{cls}{Input class labels}
}
\description{
ROC p-vaues, used in higher function

Fits a linear model per feature and returns logFC (model coefficient)
  and moderated p-values via empirical Bayes. Works on normalized/transformed data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
Compute logFC and moderated p-values using limma linear model
}
