% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{GetSampleSizeLadder}
\alias{GetSampleSizeLadder}
\title{Retrieve sample size ladder}
\usage{
GetSampleSizeLadder(maxNum)
}
\arguments{
\item{maxNum}{Numeric}
}
\description{
Return sample size ladder, used in higher functions
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
