% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{GetSelectedDataNames}
\alias{GetSelectedDataNames}
\title{Retrieve data names}
\usage{
GetSelectedDataNames(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Retrieve data names
}
