% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{GetSelectedDataNumber}
\alias{GetSelectedDataNumber}
\title{Retrieve selected data numbers}
\usage{
GetSelectedDataNumber(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Retrieve selected data numbers
}
