% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetShortNames}
\alias{GetShortNames}
\title{Perform utilities for MetPa}
\usage{
GetShortNames(nm.vec, max.len = 45)
}
\description{
Return shorter names
break long names at space, append "..." to indicate the abbrev
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
