% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_pca_anova2.R
\name{GetSigTable.Aov2}
\alias{GetSigTable.Aov2}
\title{Sig table for AOV2}
\usage{
GetSigTable.Aov2(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig table for AOV2
}
