% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{GetSigTable.ASCA}
\alias{GetSigTable.ASCA}
\title{Table of features well modelled by ASCA}
\usage{
GetSigTable.ASCA(mSetObj = NA, nm)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{nm}{Input the name of the well modelled features}
}
\description{
Table of features well modelled by ASCA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
