% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{GetSigTable.Corr}
\alias{GetSigTable.Corr}
\title{Sig table for Correlation Analysis}
\usage{
GetSigTable.Corr(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig table for Correlation Analysis
}
