% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{GetSigTable.EBAM}
\alias{GetSigTable.EBAM}
\title{Sig table for EBAM}
\usage{
GetSigTable.EBAM(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig table for EBAM
}
