% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetSigTable.FC}
\alias{GetSigTable.FC}
\title{Sig Table for Fold-Change Analysis}
\usage{
GetSigTable.FC(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig Table for Fold-Change Analysis
}
