% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_mb.R
\name{GetSigTable.MB}
\alias{GetSigTable.MB}
\title{Sig table for MB analysis}
\usage{
GetSigTable.MB(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig table for MB analysis
}
