% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetSigTable}
\alias{GetSigTable}
\title{Create Latex table}
\usage{
GetSigTable(mat, method, data.type)
}
\arguments{
\item{mat}{Input matrix}

\item{method}{Input method to create table}

\item{data.type}{Input the data type}
}
\description{
generate Latex table
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
