% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetSigTable.RF}
\alias{GetSigTable.RF}
\title{Sig table for random forest analysis}
\usage{
GetSigTable.RF(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig table for random forest analysis
}
