% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetSigTable.SVM}
\alias{GetSigTable.SVM}
\title{Sig table for SVM}
\usage{
GetSigTable.SVM(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig table for SVM
}
