% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetSigTable.Volcano}
\alias{GetSigTable.Volcano}
\title{Sig table for Volcano Analysis}
\usage{
GetSigTable.Volcano(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Sig table for Volcano Analysis
}
