% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{GetSOMClusterMembers}
\alias{GetSOMClusterMembers}
\title{SOM analysis}
\usage{
GetSOMClusterMembers(mSetObj = NA, i, j)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{i}{Index of X}

\item{j}{Index of Y}
}
\description{
Get members for given cluster index, return a character string
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
