% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{getSpectraInclusion}
\alias{getSpectraInclusion}
\title{getSpectraInclusion}
\usage{
getSpectraInclusion()
}
\description{
read the spectra file info for project loading
}
\author{
Zhiqiang Pang
}
