% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_stats.R
\name{GetSSPTable}
\alias{GetSSPTable}
\title{Replace the last column of the ssp.mat with the final selection from users}
\usage{
GetSSPTable(mSetObj = NA)
}
\description{
Replace the last column of the ssp.mat with the final selection from users
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
