% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{GetSuggestedSAMDelta}
\alias{GetSuggestedSAMDelta}
\title{For SAM analysis}
\usage{
GetSuggestedSAMDelta(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
obtain a default delta with reasonable number
of sig features and decent FDR
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
