% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetSVMSigMat}
\alias{GetSVMSigMat}
\title{Recursive Support Vector Machine (R-SVM) Significance Measure}
\usage{
GetSVMSigMat(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Return significance measure, double[][]
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
