% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetTopInx}
\alias{GetTopInx}
\title{Volcano indices}
\usage{
GetTopInx(vec, n, dec = T)
}
\arguments{
\item{vec}{Vector containing volcano indices}

\item{n}{Numeric}

\item{dec}{Logical, default set to TRUE}
}
\description{
Get indices of top n largest/smallest number
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
