% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetTtestRes}
\alias{GetTtestRes}
\title{Retrieve T-test p-values}
\usage{
GetTtestRes(mSetObj = NA, paired = FALSE, equal.var = TRUE, nonpar = F)
}
\description{
Utility method to get p values
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
