% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetTTSigMat}
\alias{GetTTSigMat}
\title{T-test matrix}
\usage{
GetTTSigMat(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Return a double matrix with 2 columns - p values and lod
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
