% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetTTSubsetIndex}
\alias{GetTTSubsetIndex}
\title{Subset data by p value (t-test)}
\usage{
GetTTSubsetIndex(mSetObj = NA, sub.num = 50)
}
\description{
Get a subsets of data ranked by their p values from t tests
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
