% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetUnivReport}
\alias{GetUnivReport}
\title{Utility method to perform the univariate analysis automatically}
\usage{
GetUnivReport(mSetObj = NA)
}
\description{
The approach is computationally expensive,and fails more often 
get around: make it lazy unless users request, otherwise the default t-test will also be affected
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
