% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetVariableLabel}
\alias{GetVariableLabel}
\title{Determine variable label for plotting}
\usage{
GetVariableLabel(data.type)
}
\description{
Determine data type, binned spectra, nmr peak, or ms peak
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
