% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{GetVennGeneNames}
\alias{GetVennGeneNames}
\title{Get Venn names}
\usage{
GetVennGeneNames(mSetObj = NA, areas)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{areas}{Input areas to retrieve names}
}
\description{
Get Venn names
}
