% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_peaks.R, R/util_peaks.R
\name{GroupPeakList}
\alias{GroupPeakList}
\title{Group peak list}
\usage{
GroupPeakList(mSetObj=NA, mzwid, bw, minfrac, minsamp, max)

GroupPeakList(mSetObj=NA, mzwid, bw, minfrac, minsamp, max)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{mzwid, }{define the width of overlapping m/z slices to use for creating peak density chromatograms
and grouping peaks across samples}

\item{bw, }{define the bandwidth (standard deviation or half width at half maximum) of gaussian smoothing
kernel to apply to the peak density chromatogram}

\item{minfrac, }{define the minimum fraction of samples necessary in at least one of the sample groups for
it to be a valid group}

\item{minsamp, }{define the minimum number of samples necessary in at least one of the sample groups for 
it to be a valid group}

\item{max, }{define the maximum number of groups to identify in a single m/z slice}
}
\description{
Group peaks from the peak list based on position
using the XCMS grouping algorithm (align peaks wrt, rt, and mz).
For NMR peaks, need to change ppm -> mz and add dummy rt.
If the data is 2-column MS, first need to add dummy rt.
If the data is 3-column MS, the data can be used directly.
The default mzwid for MS is 0.25 m/z, and for NMR is 0.03 ppm.
The default bw is 30 for LCMS, and 5 for GCMS.

Group peaks from the peak list based on position
using the XCMS grouping algorithm (align peaks wrt, rt, and mz).
For NMR peaks, need to change ppm -> mz and add dummy rt.
If the data is 2-column MS, first need to add dummy rt.
If the data is 3-column MS, the data can be used directly.
The default mzwid for MS is 0.25 m/z, and for NMR is 0.03 ppm.
The default bw is 30 for LCMS, and 5 for GCMS.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
