% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{heckbert}
\alias{heckbert}
\title{Perform utilities for MetPa}
\usage{
heckbert(dmin, dmax, m)
}
\description{
function to calculate tick mark based on Heckbert algorithm
available in the "labeling" package implemented by Justin Talbot
adapted from the imagemap package
Heckbert's labeling algorithm
Heckbert, P. S. (1990) Nice numbers for graph labels, Graphics Gems I, Academic Press Professional, Inc.
}
\author{
Justin Talbot \email{jtalbot@stanford.edu}
}
