% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{HMDBID2KEGGID}
\alias{HMDBID2KEGGID}
\title{Given a vector of HMDBIDs, return a vector of KEGG IDs}
\usage{
HMDBID2KEGGID(ids)
}
\arguments{
\item{ids}{Input the vector of HMDB Ids}
}
\description{
This function, when given a vector of HMDBIDs, returns a vector of KEGG ID. HMDB standing
for the Human Metabolome Database.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
