% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_graphics.R
\name{image.plot}
\alias{image.plot}
\title{Plot a strip of color key beside a figure}
\usage{
image.plot(..., add = FALSE, nlevel = 64, horizontal = FALSE,
  legend.shrink = 0.9, legend.width = 1.2, legend.mar = ifelse(horizontal,
  3.1, 5.1), legend.lab = NULL, graphics.reset = FALSE, bigplot = NULL,
  smallplot = NULL, legend.only = FALSE, col = tim.colors(nlevel),
  lab.breaks = NULL, axis.args = NULL, legend.args = NULL,
  midpoint = FALSE)
}
\description{
Adapted from the image.plot in fields package to correct label
so that the small p value is bigger, located on top of the color key
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
