% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{ImportRawMSData}
\alias{ImportRawMSData}
\title{Import raw MS data}
\usage{
ImportRawMSData(grpA = "X", numA = 1, grpB = "Y", numB = 1,
  format = "png", dpi = 72, width = 9)
}
\arguments{
\item{grpA}{Character, input the first group label.}

\item{numA}{Numeric, input the number of samples in group A.}

\item{grpB}{Character, input the second group label.}

\item{numB}{Numeric, input the number of samples in group B.}

\item{format}{Character, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}
}
\description{
This function handles the reading in of 
raw MS data (.mzML, .CDF and .mzXML). Users must set 
their working directory to the file containing their raw 
data and specify the group labels. The function will output
two chromatograms into the user's working directory, a 
base peak intensity chromatogram (BPIC) and a total ion 
chromatogram (TIC).
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai88@gmail.com}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
