% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{ImputeMissingVar}
\alias{ImputeMissingVar}
\title{Data processing: Replace missing variables}
\usage{
ImputeMissingVar(mSetObj, method)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{method}{Select the option to replace missing variables, either 
replacement based on the minimum ("min), the mean ("mean"), or the median ("median") value of each feature columns,
or several options to impute the missing values, using k-nearest neighbour ("KNN"), probabilistic PCA ("PPCA"), 
Bayesian PCA ("BPCA") method, or Singular Value Decomposition ("svdImpute")}
}
\description{
Replace missing variables by min/mean/median/KNN/BPCA/PPCA/svdImpute.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
