% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{InitializaPlan}
\alias{InitializaPlan}
\title{InitializePlan}
\usage{
InitializaPlan()
}
\description{
this function is used to initialize a plan before submit job to spring daemon
}
\author{
Zhiqiang Pang
}
