% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{InitMSObjects}
\alias{InitMSObjects}
\title{InitMSObjects}
\usage{
InitMSObjects(data.type = NULL, anal.type = NULL, paired = FALSE)
}
\arguments{
\item{data.type}{should be "raw"}

\item{anal.type}{should be "spec"}

\item{paired}{should be "FALSE"}
}
\description{
InitMSObjects
}
