% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{InitPowerAnal}
\alias{InitPowerAnal}
\title{Function for power analysis}
\usage{
InitPowerAnal(mSetObj, clsOpts)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{clsOpts}{For data with >2 groups, specify the two classes on which to perform power analysis, 
otherwise for data with 2 groups, "NA" will automatically select the 2 groups.}
}
\description{
Perform power analysis, requires the SSPA R package.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
