% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{InitStatAnalMode}
\alias{InitStatAnalMode}
\title{Introduction for statistical analysis module report
Initialize Statistical Analysis Report}
\usage{
InitStatAnalMode()
}
\description{
Introduction for statistical analysis module report
Initialize Statistical Analysis Report
}
