% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pca_anova2.R
\name{iPCA.Anal}
\alias{iPCA.Anal}
\title{Perform PCA analysis, prepare file for interactive liveGraphics3D}
\usage{
iPCA.Anal(mSetObj, fileNm)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{fileNm}{select a file name}
}
\description{
Perform PCA analysis, prepare file for interactive liveGraphics3D
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
