% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{IsSpectraProcessingOK}
\alias{IsSpectraProcessingOK}
\title{Check if the spectra processing is ok}
\usage{
IsSpectraProcessingOK(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects).}
}
\description{
Check if the spectra processing is ok
}
