% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{KEGGID2HMDBID}
\alias{KEGGID2HMDBID}
\title{Given a vector of KEGGIDs, return a vector of HMDB ID}
\usage{
KEGGID2HMDBID(ids)
}
\description{
This functionn, when given a vector of KEGGIDs, returns a vector of HMDB IDs. HMDB standing
for the Human Metabolome Database.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
