% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{KEGGID2Name}
\alias{KEGGID2Name}
\title{Given a vector containing KEGGIDs, returns a vector of KEGG compound names}
\usage{
KEGGID2Name(ids)
}
\arguments{
\item{ids}{Vector of KEGG ids}
}
\description{
This function, given a vector containing KEGGIDs, returns a vector of KEGG compound names.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
