% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{KEGGPATHID2SMPDBIDs}
\alias{KEGGPATHID2SMPDBIDs}
\title{Given a vector containing KEGG pathway IDs, return a vector containing SMPDB IDs (only for hsa)}
\usage{
KEGGPATHID2SMPDBIDs(ids)
}
\description{
This function, when given a vector of KEGG pathway IDs, return a vector of SMPDB IDs (only for hsa).
SMPDB standing for the Small Molecule Pathway Database, and hsa standing for human serum albumin.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
