% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{Kmeans.Anal}
\alias{Kmeans.Anal}
\title{K-means analysis}
\usage{
Kmeans.Anal(mSetObj = NA, clust.num)
}
\description{
Perform K-means analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
