% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{LoadKEGGLib}
\alias{LoadKEGGLib}
\title{Load KEGG library}
\usage{
LoadKEGGLib(libType, libNm)
}
\description{
Load different libraries
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
