% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{LoadSmpLib}
\alias{LoadSmpLib}
\title{Load pathway library}
\usage{
LoadSmpLib(mSetObj = NA)
}
\description{
Load pathway library
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
