% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{LSD.test}
\alias{LSD.test}
\title{Calculate Fisher's Least Significant Difference (LSD)}
\usage{
LSD.test(y, trt, alpha = 0.05)
}
\description{
Adapted from the 'agricolae' package
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
