% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{make_cpdlist}
\alias{make_cpdlist}
\title{Utility function to create compound lists for permutation analysis}
\usage{
make_cpdlist_function(mSetObj=NA, match_dict, refmzlist, input_mzlist)
}
\arguments{
\item{mummichog}{Input the name of the created mummichog object (see mummichog_init).}

\item{input_mzlist}{The vector of randomly drawn m/z features.}
}
\description{
From a vector of m/z features, this function outputs a vector of compounds.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
