% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{make_ecpdlist}
\alias{make_ecpdlist}
\title{Utility function to create compound lists for permutation analysis}
\usage{
make_ecpdlist(mSetObj=NA, input_mzs)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj}

\item{input_mzs}{The vector of randomly drawn m/z features.}
}
\description{
From a vector of m/z features, this function outputs a vector of compounds.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
