% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{map}
\alias{map}
\title{sPLS-DA Map}
\usage{
map(Y)
}
\arguments{
\item{Y}{Input data}
}
\description{
map variable for (s)plsda
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
