% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maplist_utils.R
\name{MapListIds}
\alias{MapListIds}
\title{Mapping user supplied IDs}
\usage{
MapListIds(listNm, geneIDs, org, idType)
}
\arguments{
\item{listNm}{File name of the list}

\item{geneIDs}{Raw string of the gene list}

\item{org}{Three letters annotation of organism (i.e hsa, mmu)}

\item{idType}{The orginal ID type}

\item{dataSetObj}{Input the name of the created datasetObj (see Init.Data).}
}
\description{
Mapping user supplied IDs to internal IDs (in most cases Entrez)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
