% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{Match.Pattern}
\alias{Match.Pattern}
\title{Match pattern for correlation analysis}
\usage{
Match.Pattern(mSetObj = NA, dist.name = "pearson", pattern = NULL)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj}

\item{dist.name}{Input the distance method, default is set to pearson}

\item{pattern}{Set the pattern, default is set to NULL}
}
\description{
Match pattern for correlation analysis
}
