% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depends_refs.R
\name{melt}
\alias{melt}
\title{melt
Convert an object into a molten data frame. This function is from reshape2 package.}
\usage{
melt(data, ..., na.rm = FALSE, value.name = "value")
}
\arguments{
\item{data}{Data set to melt}

\item{...}{further arguments passed to or from other methods.}

\item{na.rm}{Should NA values be removed from the data set? This will
convert explicit missings to implicit missings.}

\item{value.name}{name of variable used to store values}
}
\description{
This the generic melt function. See the following functions
for the details about different data structures:
}
