% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_data.R
\name{MergeDatasets}
\alias{MergeDatasets}
\title{Utility function for PrepareQueryJson}
\usage{
MergeDatasets(dataSet1, dataSet2)
}
\arguments{
\item{dataSet1}{Input the first dataset}

\item{dataSet2}{Input the second dataset}
}
\description{
Utility function for PrepareQueryJson
}
