% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{MergeDuplicates}
\alias{MergeDuplicates}
\title{Perform miscellaneous tasks}
\usage{
MergeDuplicates(data, dim = 2)
}
\description{
Perform misc tasks

dim 1 => row,  dim 2 => column
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

merge duplicated columns or rows by their mean
}
