% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{MetaboAnalystMatrix}
\alias{MetaboAnalystMatrix}
\title{Matrix creation for MetaboAnalyst}
\usage{
MetaboAnalystMatrix(xset)
}
\arguments{
\item{xset}{The xcmsSet object created}
}
\description{
Converts processed raw lc/ms data from XCMS 
to a suitable matrix for use in MetaboAnalyst
}
