% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{MetaboliteMappingExact}
\alias{MetaboliteMappingExact}
\title{Mapping from different metabolite IDs}
\usage{
MetaboliteMappingExact(mSetObj = NA, q.type, lipid = F)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}

\item{q.type}{Inpute the query-type, "name" for compound names, "hmdb" for HMDB IDs, "kegg" for KEGG IDs, "pubchem"
for PubChem CIDs, "chebi" for ChEBI IDs, "metlin" for METLIN IDs, and "hmdb_kegg" for a both KEGG and HMDB IDs.}

\item{lipid}{Boolean, if features are lipids, a different database will be used for
compound matching.}
}
\description{
For compound names to other ids, can do exact or approximate matches
For other IDs, except HMDB ID, all others may return multiple/non-unique hits
Multiple hits or non-unique hits will allow users to manually select
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
