% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{MetaPathNormalization}
\alias{MetaPathNormalization}
\title{MetaPathNormalization}
\usage{
MetaPathNormalization(
  mSetObj = NA,
  sampleNor,
  tranform,
  scale = "NULL",
  name,
  name2
)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{sampleNor}{sample Normalization option}

\item{tranform}{sample transformation option}

\item{scale}{sample scale option}

\item{name}{file name 1 with absolute path}

\item{name2}{file name 2 with absolute path or "null"}
}
\description{
MetaPathNormalization
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca} 
Zhiqiang Pang\email{zhiqiang.pang@mail.mcgill.ca}
McGill University, Canada
License: MIT License
}
