% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{mSet2xcmsSet}
\alias{mSet2xcmsSet}
\title{mSet2xcmsSet}
\usage{
mSet2xcmsSet(mSet)
}
\description{
mSet2xcmsSet
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
