% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{MSspec.fillPeaks}
\alias{MSspec.fillPeaks}
\title{Fill in missing peaks}
\usage{
MetaboAnalystR:::MSspec.fillPeaks(mSetObj)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
For each sample in the processed MS spectra data, this function will fill in missing peaks 
using the fillPeaks function from the XCMS package. First, the function will identify any peak groups that are missing 
any peaks from the samples and will then fill in those peaks by rereading the raw data and 
integrating signals at those regions to create a new peak.
}
