% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{MSspec.rtCorrection}
\alias{MSspec.rtCorrection}
\title{Retention time correction for LC/GC-MS spectra}
\usage{
MetaboAnalystR:::MSspec.rtCorrection(mSetObj, bw)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{bw:}{define the bandwidth (standard deviation or half width at half maximum) of gaussian smoothing
kernel to apply to the peak density chromatogram}
}
\description{
Performs retention time correction for LC/GC-MS spectra using the XCMS package. Following
retention time correction, the object dataSet will be regrouped.
}
