% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{multi.stat}
\alias{multi.stat}
\title{Get multiple category statistics}
\usage{
multi.stat(pred, resp)
}
\description{
Get multiple category statistics
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
