% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming.R
\name{mz.trim_random}
\alias{mz.trim_random}
\title{Data trimming Method Based on Random MS}
\usage{
mz.trim_random(raw_data, ms_list)
}
\arguments{
\item{raw_data}{MSnExp object, the raw data that has been read in memory.}

\item{ms_list}{List, the names list of all scans.}
}
\description{
Trim raw data scan signal randomly in the mz dimension.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
