% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming.R
\name{mz.trim_specific}
\alias{mz.trim_specific}
\title{Data trimming Method Based on Specific MS}
\usage{
mz.trim_specific(raw_data, ms_list, mz, mzdiff = 100)
}
\arguments{
\item{raw_data}{MSnExp object, the raw data that has been read in memory.}

\item{ms_list}{List, the names list of all scans.}

\item{mz}{Numeric, the specifric mz value that will be kept or removed.}

\item{mzdiff}{Numeric, the deviation (ppm) for the 'mz' values. Default is 100.}
}
\description{
Trim data based on specific mz values. Positive values will be specially retained, 
while the negative values will be removed.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
