% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{Noise_evaluate}
\alias{Noise_evaluate}
\title{Noise_evaluation based on Kernal density model}
\usage{
Noise_evaluate(raw_data)
}
\arguments{
\item{raw_data}{MSnExp object, the (trimmed) data in memory produced by 'PerformDataTrimming'.}
}
\description{
This functions handles the evaluation on the data noise (noise and prefilter parameters) 
and the identification on the molecule weights deviation evaluation.
}
\references{
McLean C (2020). Autotuner: Automated parameter selection for untargeted metabolomics data processing
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
