% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{Normalization}
\alias{Normalization}
\title{Normalization}
\usage{
Normalization(mSetObj, rowNorm, transNorm, scaleNorm, ref=NULL, ratio=FALSE, ratioNum=20)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{rowNorm}{Select the option for row-wise normalization, "QuantileNorm" for Quantile Normalization, 
"ProbNormT" for Probabilistic Quotient Normalization without using a reference sample,
"ProbNormF" for Probabilistic Quotient Normalization based on a reference sample, 
"CompNorm" for Normalization by a reference feature,
"SumNorm" for Normalization to constant sum, 
"MedianNorm" for Normalization to sample median, and 
"SpecNorm" for Normalization by a sample-specific factor.}

\item{transNorm}{Select option to transform the data, "LogNorm" for Log Normalization,
and "CrNorm" for Cubic Root Transformation.}

\item{scaleNorm}{Select option for scaling the data, "MeanCenter" for Mean Centering,
"AutoNorm" for Autoscaling, "ParetoNorm" for Pareto Scaling, amd "RangeNorm" for Range Scaling.}

\item{ref}{Input the name of the reference sample or the reference feature, use " " around the name.}

\item{ratio}{This option is only for biomarker analysis.}

\item{ratioNum}{Relevant only for biomarker analysis.}
}
\description{
This function performs row-wise normalization, transformation, and 
scaling of your metabolomic data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}, Jasmine Chong
McGill University, Canada
}
