% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{OPLSDA.Permut}
\alias{OPLSDA.Permut}
\title{Perform OPLS-DA permutation}
\usage{
OPLSDA.Permut(mSetObj = NA, num = 100)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{num}{Input the number of permutations, default is set to 100.}
}
\description{
Orthogonal PLS-DA (from ropls) 
perform permutation, using training classification accuracy as
indicator, for two or multi-groups
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
