% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{optimizxcms.doe.peakpicking}
\alias{optimizxcms.doe.peakpicking}
\title{Core Optimization Function of DoE}
\usage{
optimizxcms.doe.peakpicking(
  object = NULL,
  params = params,
  BPPARAM = bpparam(),
  nSlaves = 4,
  plot = F,
  ...
)
}
\arguments{
\item{BPPARAM}{MulticoreParam method, used to set the parallel method. Default is bpparam().}

\item{plot}{Logical, weather to plot the Contours plots of the DoE results.}

\item{objet}{MSnExp object, the trimmed or the original data.}

\item{param}{List, the parameters lists set by 'SetPeakParam' function. The noise, prefilter and ppm values should 
be defined by AutoTuner in the previous steps.}

\item{nSlave}{Numeric, core number used to perform the parallel based optimization.}
}
\description{
This function is the core for parameters optimization with Design of Experiment (DoE)
method.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
