% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{parseFisher}
\alias{parseFisher}
\title{Return only the signicant comparison names}
\usage{
parseFisher(fisher, cut.off)
}
\description{
Return only the signicant comparison names, used in higher function
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
