% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{parseTukey}
\alias{parseTukey}
\title{Return only the signicant comparison names}
\usage{
parseTukey(tukey, cut.off)
}
\arguments{
\item{tukey}{Input tukey output}

\item{cut.off}{Input numeric cut-off}
}
\description{
Return only the signicant comparison names, used in higher function
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
