% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{PathMapping}
\alias{PathMapping}
\title{Perform pathway mapping}
\usage{
PathMapping(mSetObj = NA, qvec)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{qvec}{Input the query vector}
}
\description{
Perform pathway mapping of compounds
}
