% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PCA.Anal}
\alias{PCA.Anal}
\title{Perform PCA analysis}
\usage{
PCA.Anal(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object
McGill University, Canada
License: MIT License}
}
\description{
Perform PCA analysis, obtain variance explained, store item to PCA object
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
}
