% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{PCA.GENES}
\alias{PCA.GENES}
\title{Obtain principal components into a matrix that has more variables than individuals}
\usage{
PCA.GENES(X)
}
\arguments{
\item{X}{Input matrix that has as columns the compounds that were considered as variables in the PCA analysis}
}
\description{
X is a matrix that has as columns the compounds that were considered as variables in the PCA analysis.
First we center the matrix by columns (Xoff) and then we obtain the eigenvalues and the 
eigenvectors of the matrix Xoff%*%t(Xoff) and we
use the equivalences between the loadings and scores to obtain the solution
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
