% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{PeakPicking_centWave_slave}
\alias{PeakPicking_centWave_slave}
\title{PeakPicking_centWave_slave}
\usage{
PeakPicking_centWave_slave(x, param)
}
\description{
PeakPicking_centWave_slave
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
