% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{PeakPicking_core}
\alias{PeakPicking_core}
\title{Calculate PPS method}
\usage{
PeakPicking_core(
  object,
  object_mslevel,
  param,
  BPPARAM = bpparam(),
  msLevel = 1L
)
}
\arguments{
\item{object_mslevel}{List, prepared by findChromPeaks_prep function.}

\item{param}{Parameters list.}

\item{BPPARAM}{Parallel Method.}

\item{msLevel}{msLevel. Only 1 is supported currently.}

\item{xset}{MSnExp object.}
}
\description{
Peak picking method. Specifically used for parameters optimization
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
