% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{PeakPicking_prep}
\alias{PeakPicking_prep}
\title{Data Preparation for ChromPeaking Finding}
\usage{
PeakPicking_prep(object)
}
\arguments{
\item{object}{MSnExp object.}
}
\description{
Data Preparation for ChromPeaking Finding
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
