% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{perf.splsda}
\alias{perf.splsda}
\title{Performance for PLS-DA and sPLS-DA}
\usage{
perf.splsda(object, dist = c("all", "max.dist", "centroids.dist",
  "mahalanobis.dist"), constraint = FALSE, validation = c("Mfold", "loo"),
  folds = 5, nrepeat = 1, auc = FALSE, progressBar = FALSE, cpus, ...)
}
\description{
Performance prediction for PLS-DA and sPLS-DA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
