% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{Perform.ASCA}
\alias{Perform.ASCA}
\title{Perform ASCA}
\usage{
Perform.ASCA(mSetObj = NA, a = 1, b = 2, x = 2, res = 2)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{a}{specify the number of components for facA}

\item{b}{specify the number of components for facB}

\item{x}{specify the number of components for interaction AB}

\item{res}{specify the number of model residuals
type is string, indicating the type of analysis
"abc" separately
"aab" facA joins with AB
"bab" facB joins with AB}
}
\description{
The ASCA algorithm was adapted from the ASCA-genes method
(analysis of variance (ANOVA) simultaneous component analysis)
by Maria Jose Nueda (mj.nueda@ua.es) and Ana Conesa (aconesa@ivia.es)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
