% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{Perform.permutation}
\alias{Perform.permutation}
\title{Permutation}
\usage{
Perform.permutation(perm.num, fun)
}
\arguments{
\item{perm.num}{Numeric, input the number of permutations to perform}

\item{fun}{Dummy function}
}
\description{
Perform permutation, options to change number of cores used
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
