% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Perform.UnivROC}
\alias{Perform.UnivROC}
\title{Perform Classical Univariate ROC}
\usage{
Perform.UnivROC(mSetObj=NA, feat.nm, imgName, format="png", dpi=72, isAUC, isOpt, optMethod, isPartial, measure, cutoff)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{feat.nm}{Input the name of the feature to perform univariate ROC analysis}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, png, of pdf.}

\item{dpi}{Input the dpi. If the image format is pdf, users need not define the dpi. For png images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{isAUC}{Logical, select T to compute the 95 percent confidence interval band and "F" to not}

\item{isOpt}{Logical, show the optimal cutoff, T to show it and F to not}

\item{optMethod}{Select the optimal cutoff by using either closest.topleft for closest to top-left corner or 
youden for farthest to the diagonal line (Youden)}

\item{isPartial}{Logical, input T to calculate a partial ROC curve, and F to not}

\item{measure}{Select the parameter to limit the calculation of the partial ROC curve, se for the X-axis (maximum false-positive rate)
and sp for the Y-axis, representing the minimum true positive-rate}

\item{cutoff}{Input the threshold to limit the calculation of the partial ROC curve, the number must be between 0 and 1.}
}
\description{
Perform Classical Univariate ROC
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
