% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{perform_opls}
\alias{perform_opls}
\title{Perform OPLS}
\usage{
perform_opls(x, y = NULL, predI = NA, orthoI = 0, crossvalI = 7,
  log10L = FALSE, permI = 20, scaleC = c("none", "center", "pareto",
  "standard")[4], ...)
}
\description{
Orthogonal PLS functions (adapted from ropls package for web-based usage)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
