% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PerformAdductMapping}
\alias{PerformAdductMapping}
\title{PerformAdductMapping}
\usage{
PerformAdductMapping(mSetObj = NA, add.mode)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object}

\item{add.mode}{Adduct mode, positive or negative}
}
\description{
This function reads in the user's adduct list and 
saves it as a matrix.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
