% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{PerformApproxMatch}
\alias{PerformApproxMatch}
\title{Perform approximate compound matches}
\usage{
PerformApproxMatch(mSetObj = NA, q)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj}

\item{q}{Input the q vector}
}
\description{
Given a query, perform approximate compound matching
}
