% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{PerformCmpdMapping}
\alias{PerformCmpdMapping}
\title{Perform compound mapping for integrative analysis methods}
\usage{
PerformCmpdMapping(mSetObj = NA, cmpdIDs, org, idType)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{cmpdIDs}{Input the list of compound IDs}

\item{org}{Input the organism code}

\item{idType}{Input the ID type}
}
\description{
Perform compound mapping
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
