% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_covariate.R
\name{PerformCovariateAdjustment}
\alias{PerformCovariateAdjustment}
\title{Perform Covariate Adjustment and Update Data}
\usage{
PerformCovariateAdjustment(
  mSetObj = NA,
  primary.var,
  blocking.factor = "NA",
  use.combat = FALSE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{primary.var}{Primary metadata variable (disease/condition)}

\item{blocking.factor}{Optional blocking factor for random effects (default "NA")}

\item{use.combat}{Logical, use ComBat instead of limma removeBatchEffect (default FALSE)}
}
\value{
1 on success, 0 on failure
}
\description{
Adjusts data for confounding covariates and replaces normalized data with adjusted data
}
