% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PerformCV.explore}
\alias{PerformCV.explore}
\title{Perform Monte-Carlo Cross Validation (MCCV)}
\usage{
PerformCV.explore(mSetObj, cls.method, rank.method="auroc", lvNum=2, propTraining=2/3)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{cls.method}{Select the classification method, "rf" for random forest classification, "pls" for PLS-DA,
and "svm" for support vector machine}

\item{rank.method}{Select the ranking method, "rf" for random forest mean decrease accuracy, 
"fisher" for Fisher's univariate ranking based on area under the curve
"auroc" for univariate ranking based on area under the curve, "tt" for T-test univariate ranking based on area under the curve,
"pls" for partial least squares, and "svm" for support vector machine}

\item{lvNum}{Input the number of latent variables to include in the analyis, only for PLS-DA classification}

\item{propTraining}{Input the proportion of samples to use for training}
}
\description{
Classification MCCV, aims to find the best feature 
subsets using default model parameters
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
