% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PerformCV.test}
\alias{PerformCV.test}
\title{Perform MCCV for manually selected features}
\usage{
PerformCV.test(mSetObj, method, lvNum, propTraining=2/3, nRuns=100)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{method}{Select the classification method, "rf" for random forest classification, "pls" for PLS-DA,
and "svm" for support vector machine}

\item{lvNum}{Input the number of latent variables to include in the analyis, only for PLS-DA classification}

\item{propTraining}{Input the proportion of samples to use for training, by default it is 2/3}

\item{nRuns}{Input the number of MCCV runs, by default it is 100}
}
\description{
MCCV for manually selected features (no additional feature selection)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
