% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_idanot_utils.R
\name{PerformDataAnnot}
\alias{PerformDataAnnot}
\title{Perform data annotation}
\usage{
PerformDataAnnot(
  dataName = "",
  org = "hsa",
  dataType = "array",
  idType = "entrez",
  lvlOpt = "mean"
)
}
\arguments{
\item{org}{three letters annotation of organism (i.e hsa, mmu)}

\item{dataType}{Either "array" (microarray) or "count" (rna-seq)}

\item{idType}{original id type of genes}

\item{lvlOpt}{merging original ID to entrez gene IDs. "NA" to keep original IDs without merging}

\item{dataSetObj}{Input the name of the created datasetObj (see Init.Data).}
}
\description{
Read data and perform gene ID mapping using built in databases
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
