% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{PerformDataTrimming}
\alias{PerformDataTrimming}
\title{Perform ROI Extraction from raw MS data (PerformDataTrimming)}
\usage{
PerformDataTrimming(
  datapath,
  mode = "ssm",
  write = FALSE,
  mz,
  mzdiff,
  rt,
  rtdiff,
  rt.idx = 1/15,
  rmConts = TRUE,
  plot = TRUE,
  running.controller = NULL
)
}
\arguments{
\item{datapath}{Character, the path of the raw MS data files' or folder's path (.mzXML, .CDF and .mzML) 
for parameters training.}

\item{mode}{Character, mode for data trimming to select the chraracteristic peaks. 
Default is 'ssm'. Users could select random trimed according to mz value (mz_random) or 
RT value (rt_random). Besides, specific peaks at certain mz (mz_specific) or 
RT (rt_specific) could also be extracted. 'none' will not trim the data.}

\item{write}{Logical, if true, will write the trimmed data to the directory 'trimmed' folder 
in the datapath. The data in memory will be kept.}

\item{mz}{Numeric, mz value(s) for specific selection. Positive values means including (the values 
indicted) and negative value means excluding/removing.}

\item{mzdiff}{Numeric, the deviation (ppm) of mz value(s).}

\item{rt}{Numeric, rt value for specific selection. Positive values means including 
and negative value means excluding.}

\item{rtdiff}{Numeric, the deviation (seconds) of rt value(s).}

\item{rt.idx}{Numeric, the relative rt (retention time) range, from 0 to 1. 1 means all retention time
will be retained, while 0 means none. Default is 1/15. If default rt.idx produce too few peaks, 
please consider increasing this value.}

\item{rmConts}{LOgical, whether to exclude/remove the potential contamination for parameters optimization. Default is TRUE.}

\item{plot}{Logical, if TRUE, will plot the chromatogram of the trimmed data.}

\item{running.controller}{The resuming pipeline running controller. Optional. Don't need to define by hand.}
}
\value{
will return an mSet objects with extracted ROI
}
\description{
This function performs the raw data trimming. This function will output 
an trimmed MSnExp file to memory or hardisk according to the choice of users must 
provide the data path for 'datapath', and optionally provide other corresponding parameters.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
