% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{PerformDetailMatch}
\alias{PerformDetailMatch}
\title{Perform detailed name match}
\usage{
PerformDetailMatch(mSetObj = NA, q)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{q}{Input the query.}
}
\description{
Given a query, perform compound matching.
}
