% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformEachDEAnal}
\alias{PerformEachDEAnal}
\title{Performs differential expression analysis on individual data}
\usage{
PerformEachDEAnal(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
This function performs DE analysis on individual data using the common matrix, which
will be used/compared in later steps of the analysis (according to the p-value). The DE for each feature
may be adjusted using the p-value.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
