% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{PerformExpressNormalization}
\alias{PerformExpressNormalization}
\title{Perform data normalization}
\usage{
PerformExpressNormalization(
  dataName,
  norm.opt,
  var.thresh,
  abundance,
  count.thresh,
  filterUnmapped
)
}
\arguments{
\item{norm.opt}{Normalization method to be used}

\item{var.thresh}{Relative abundance threshold}

\item{count.thresh}{Count threshold}

\item{filterUnmapped, }{boolean, whether to filter unmapped genes or not}
}
\description{
Normalize gene expression data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
