% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{PerformGeneMapping}
\alias{PerformGeneMapping}
\title{Perform integrated gene mapping}
\usage{
PerformGeneMapping(mSetObj = NA, geneIDs, org, idType)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{geneIDs}{Input the list of gene IDs}

\item{org}{Input the organism code}

\item{idType}{Input the ID type}
}
\description{
Used for the pathinteg module
}
