% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{PerformIndNormalization}
\alias{PerformIndNormalization}
\title{Perform normalization for individually-uploaded datasets for meta-analysis}
\usage{
PerformIndNormalization(mSetObj = NA, dataName, norm.opt, auto.opt)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{dataName}{Input the name of the individual dataset for normalization.}

\item{norm.opt}{Performs log2 normalization "log", or no normalization "none".}

\item{auto.opt}{Performs auto-scaling of data (1), or no (0).}
}
\description{
This function performs normalization of individuall-uploaded datasets prior to meta-analysis.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
