% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{PerformIntegGeneMapping}
\alias{PerformIntegGeneMapping}
\title{Perform integrated gene mapping}
\usage{
PerformIntegGeneMapping(mSetObj = NA, geneIDs, org, idType)
}
\description{
Used for the pathinteg module
}
