% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_enrich.R
\name{PerformKOEnrichAnalysis_KO01100}
\alias{PerformKOEnrichAnalysis_KO01100}
\title{Performs KO enrichment analysis based on the KO01100 map}
\usage{
PerformKOEnrichAnalysis_KO01100(mSetObj = NA, category, file.nm)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{category}{Input the option to perform enrichment analysis, "pathway"}

\item{file.nm}{Input name of file to save}
}
\description{
This function performs KO enrichment analysis based on the KO01100 map
and saves the .JSON file
}
\author{
Othman Soufan, Jeff Xia \email{jeff.xia@mcgill.ca}, {othman.soufan@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
