% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R
\name{PerformKOEnrichAnalysis_List}
\alias{PerformKOEnrichAnalysis_List}
\title{Utility function for PerformKOEnrichAnalysis_KO01100}
\usage{
PerformKOEnrichAnalysis_List(file.nm)
}
\arguments{
\item{file.nm}{Input the file name}
}
\description{
Please note: only return hits in map KO01100
}
