% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_data.R
\name{PerformMapping}
\alias{PerformMapping}
\title{Utility function for PrepareQueryJson
geneIDs is text one string, need to make to vector}
\usage{
PerformMapping(inputIDs, type)
}
\arguments{
\item{inputIDs}{Input list of IDs}

\item{type}{Input the type of IDs}
}
\description{
Utility function for PrepareQueryJson
geneIDs is text one string, need to make to vector
}
