% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R
\name{PerformMapping_ko01100}
\alias{PerformMapping_ko01100}
\title{Utility function for PrepareKeggQueryJson
geneIDs is text one string, need to make to vector}
\usage{
PerformMapping_ko01100(inputIDs, type)
}
\arguments{
\item{inputIDs}{Input list of IDs}

\item{type}{Input the type of IDs}
}
\description{
Utility function for PrepareKeggQueryJson
geneIDs is text one string, need to make to vector
}
