% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_mb.R
\name{performMB}
\alias{performMB}
\title{Timecourse analysis}
\usage{
performMB(mSetObj, topPerc)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{topPerc}{select the cut-off, default is 10}
}
\description{
Adapted from the timecourse package by Yu Chuan Tai
This method is only applicable for time-series, not for general case
two/multiple factor analysis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
