% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformMetaMerge}
\alias{PerformMetaMerge}
\title{Meta-Analysis Method: Direct merging of datasets}
\usage{
PerformMetaMerge(mSetObj = NA, BHth = 0.05)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{BHth}{Numeric input to set the significance level. By default it is 0.05.}
}
\description{
This function is one of three methods to perform meta-analysis. Direct merging of individual data into 
a mega-dataset results in an analysis of that mega-dataset as if the individual data were derived from the same experiment. This 
method thereby ignores any inherent bias and heterogeneity between the different data. Because of this, there exists several confounders 
such as different experimental protocols, technical platforms, and raw data processing procedures that can mask true underlying differences. 
It is therefore highly suggested that this approach be used only when individual data are very similar (i.e. from the same lab, same platform, 
without batch effects)."
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
