% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{PerformMirrorPlotting}
\alias{PerformMirrorPlotting}
\title{PerformMirrorPlotting}
\usage{
PerformMirrorPlotting(
  mSetObj = NA,
  fragDB_path = NA,
  peak_idx,
  sub_idx,
  interactive = T,
  ppm,
  dpi,
  format,
  width,
  height
)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{fragDB_path}{Fragmentation database path}

\item{peak_idx}{peak_idx}

\item{sub_idx}{sub_idx}

\item{interactive}{interactive or not}

\item{ppm}{ppm}

\item{dpi}{dpi}

\item{format}{format}

\item{width}{width}

\item{height}{height}

\item{featurelabel}{featurelabel}

\item{imageNM}{imageNM}
}
\description{
PerformMirrorPlotting
}
