% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_compatibility.R
\name{PerformMS1ResultsFormatting}
\alias{PerformMS1ResultsFormatting}
\title{PerformMS1ResultsFormatting
This function is used to format the results from other tools into the generic format of MetaboAnalystR
Currently,we are supporting the compatibility for four commonly used open-source tools: MS-DIAL, MZmine, Asari and XCMS online
The first parameter file_path should be a valid file of the result.
User need to specify the type in the 2nd argument, type. This argument can be msdial, mzmine, asari and xcms
Please note, if your original data does not contain meta information, you need to manually add them in the generated "metaboanalyst_input.csv" file
The formatted file is 'sample in columns'.}
\usage{
PerformMS1ResultsFormatting(file_path, type, meta_data = NA)
}
\arguments{
\item{file_path}{}

\item{type}{}

\item{meta_data}{this is path to a table containing two columns, the first column is the sample names and second column is the group information}
}
\description{
PerformMS1ResultsFormatting
This function is used to format the results from other tools into the generic format of MetaboAnalystR
Currently,we are supporting the compatibility for four commonly used open-source tools: MS-DIAL, MZmine, Asari and XCMS online
The first parameter file_path should be a valid file of the result.
User need to specify the type in the 2nd argument, type. This argument can be msdial, mzmine, asari and xcms
Please note, if your original data does not contain meta information, you need to manually add them in the generated "metaboanalyst_input.csv" file
The formatted file is 'sample in columns'.
}
