% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem_ms.R
\name{performMS2searchBatch}
\alias{performMS2searchBatch}
\title{performMS2searchBatch}
\usage{
performMS2searchBatch(
  mSetObj = NA,
  ppm1 = 10,
  ppm2 = 25,
  dbpath = "",
  frgdbpath = "",
  database = "all",
  similarity_meth = 0,
  precMZ = NA,
  sim_cutoff = 30,
  ionMode = "positive",
  unit1 = "ppm",
  unit2 = "ppm",
  ncores = 1
)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{ppm1}{ppm value for ms1}

\item{ppm2}{ppm value for ms2}

\item{dbpath}{database path}

\item{frgdbpath}{fragmentation database path}

\item{database}{database option}

\item{similarity_meth}{similarity computing method}

\item{precMZ}{mz of precursor}

\item{sim_cutoff}{filtration cutoff of similarity score. will be enabled soon.}

\item{ionMode}{ion mode, for ESI+, is should be 1. for ESI-, it should be 0}

\item{unit1}{ppm or da for ms1 matching}

\item{unit2}{ppm or da for ms2}

\item{ncores}{number of cpu cores used to search}
}
\description{
performMS2searchBatch
}
\author{
Zhiqiang Pang
}
