% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming.R
\name{PerformMSDataOutput}
\alias{PerformMSDataOutput}
\title{Function MS Generation}
\usage{
PerformMSDataOutput(raw_data)
}
\arguments{
\item{raw_data}{MS data in R environment with "MSnExp" class.}
}
\description{
Output the MS data. This function will generate .mzML MS data in the working dirctory.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
