% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{PerformMultiMatch}
\alias{PerformMultiMatch}
\title{Perform multiple name matches}
\usage{
PerformMultiMatch(mSetObj = NA, q, lipid)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{q}{Input the query.}
}
\description{
Given a query, performs compound name matching.
}
