% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{PerformMummichog}
\alias{PerformMummichog}
\title{Main function to perform mummichog}
\usage{
mSet <- PerformMummichog(mummichog = NA, enrichOpt, pvalOpt, permNum = 100)
}
\arguments{
\item{mummichog}{Input the name of the created mummichog object (see mummichog_init).}
}
\description{
This is the main function that performs the mummichog analysis.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
