% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{PerformMummichogPermutations}
\alias{PerformMummichogPermutations}
\title{Perform permutations and calculates p-values}
\usage{
PerformMummichogPermutations(mSetObj=NA, permNum)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}

\item{permNum}{Numeric, the number of permutations to perform}
}
\description{
This function performs permutations to create a background distribution of p-values for pathway
enrichment analysis.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
