% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PerformMumTableStat}
\alias{PerformMumTableStat}
\title{Perform statistical analysis for table input (ranking only, no conversion)}
\usage{
PerformMumTableStat(
  mSetObj = NA,
  ranking.method = "classical",
  pval.cutoff = 0.05,
  fc.cutoff = 0,
  ref.group = NA,
  contrast.group = NA,
  covariates = NA
)
}
\arguments{
\item{mSetObj}{mSet Object}

\item{ranking.method}{"classical" or "limma"}
}
\value{
Number of significant features (FDR < 0.05)
}
\description{
Computes t-test/ANOVA or limma statistics for feature ranking.
  Results are stored in mSetObj$analSet$tt for later use by Convert2Mummichog.
}
