% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{PerformParamsOptimization}
\alias{PerformParamsOptimization}
\title{Perform Parameters Optimization}
\usage{
PerformParamsOptimization(
  mSet,
  param = NULL,
  method = "DoE",
  ncore = 4,
  running.controller = NULL
)
}
\arguments{
\item{mSet}{mSet object, usually generated by 'PerformROIExtraction' 
or 'PerformDataTrimming' here.}

\item{param}{List, Parameters defined by 'SetPeakParam' function.}

\item{method}{Character, method of parameters optimization, including 
"DoE' only. Default is "DoE". Other method 
is under development.}

\item{ncore}{Numeric, CPU threads number used to perform the parallel 
based optimization. If thers is memory issue,
please reduce the 'ncore' used here. For default, 2/3 CPU threads of 
total will be used.}

\item{running.controller}{The resuming pipeline running controller. Optional. 
Don't need to define by hand.}
}
\value{
will a parameter object can be used for following processing
}
\description{
This function is used to optimize the critical 
parameters of peak picking and alignment for 
the following data processing. It utilizes the trimed data and 
the internal instrument-specific parameters.
Parallel computing will be performed. The number of cores user 
want to use could be specified.
}
\details{
PerformParamsOptimization
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} 
Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
