% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{PerformPeakAlignment}
\alias{PerformPeakAlignment}
\title{PerformPeakAlignment}
\usage{
PerformPeakAlignment(mSet, param)
}
\arguments{
\item{mSet}{the mSet object generated by PerformPeakPicking function.}

\item{param}{param list generated by updateRawSpectraParam function.}
}
\description{
PerformPeakAlignment
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
