% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{PerformPeakFiling}
\alias{PerformPeakFiling}
\title{PerformPeakFiling}
\usage{
PerformPeakFiling(mSet, param, BPPARAM = bpparam())
}
\arguments{
\item{mSet}{the mSet object generated by PerformPeakPicking function.}

\item{param}{param list generated by updateRawSpectraParam function.}

\item{BPPARAM}{parallel method used for data processing. Default is bpparam().}
}
\description{
PerformPeakFiling
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
