% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{PerformPeakPicking}
\alias{PerformPeakPicking}
\title{PerformPeakPicking}
\usage{
PerformPeakPicking(object, param, BPPARAM = bpparam())
}
\arguments{
\item{object}{the raw data object read by ImportRawMSData function.}

\item{param}{param list generated by updateRawSpectraParam function.}

\item{BPPARAM}{parallel method used for data processing. Default is bpparam().}
}
\description{
PerformPeakPicking
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
