% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{PerformPeakProfiling}
\alias{PerformPeakProfiling}
\title{PerformPeakProfiling}
\usage{
PerformPeakProfiling(mSet, Params, plotSettings, ncore)
}
\arguments{
\item{mSet}{mSet}

\item{Params}{Params}

\item{plotSettings}{plotSettings}

\item{ncore}{number of cores}
}
\description{
PerformPeakProfiling
}
