% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{PerformPowerProfiling}
\alias{PerformPowerProfiling}
\title{Perform power profiling}
\usage{
PerformPowerProfiling(mSetObj, fdr.lvl, smplSize)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{fdr.lvl}{Specify the false-discovery rate level.}

\item{smpSize}{Specify the maximum sample size, the number must be between 60-1000.}
}
\description{
Perform power profiling of data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
