% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformPvalCombination}
\alias{PerformPvalCombination}
\title{Meta-Analysis Method: Combining p-values}
\usage{
PerformPvalCombination(mSetObj = NA, method = "stouffer", BHth = 0.05)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{method}{Method of p-value combination. By default it is "stouffer", else it is "fisher".}

\item{BHth}{Numeric input to set the significance level. By default it is 0.05.}
}
\description{
This function is one of three methods to perform meta-analysis. Here, p-values are combined using either
the Fisher's method or the Stouffer's method.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
