% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformVoteCounting}
\alias{PerformVoteCounting}
\title{Meta-Analysis Method: Vote Counting}
\usage{
PerformVoteCounting(mSetObj = NA, BHth = 0.05, minVote)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{BHth}{Numeric input to set the significance level. By default it is 0.05.}

\item{minVote}{Numeric input to set the minimum vote-count.}
}
\description{
This function is one of three methods to perform meta-analysis. Here, significant features are selected based on a selected criteria (i.e. an adjusted p-value
<0.05 and the same direction of FC) for each dataset. The votes are then calculated for each feature by counting the total of number of times
a feature is significant across all included datasets. However, this method is statistically inefficient and should be considered the
last resort in situations where other methods to perform meta-analysis cannot be applied.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
