% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{plot.dist}
\alias{plot.dist}
\title{Batch Distance Plotting}
\usage{
plot.dist(mSetObj = NA, imgName = "dist", format = "png", width = NA, dpi = 72)
}
\description{
Scatter sample trend comparison between all sample of different batches
}
