% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{plot.sample.dist}
\alias{plot.sample.dist}
\title{Batch Distance Plotting}
\usage{
plot.sample.dist(
  mSetObj = NA,
  imgName,
  format = "png",
  width = NA,
  dpi = default.dpi
)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{imgName}{imgName}

\item{format}{format}

\item{width}{width}

\item{dpi}{dpi}
}
\description{
Scatter sample trend comparison between all sample of different batches
}
