% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{plot_dist}
\alias{plot_dist}
\title{Batch Distance Plotting}
\usage{
plot_dist(mSetObj = NA, imgName = "dist", format = "png", width = NA, dpi = 72)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{imgName}{imgName}

\item{format}{format}

\item{width}{width}

\item{dpi}{dpi}
}
\description{
Scatter sample trend comparison between all sample of different batches
}
