% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{PlotASCA.Permutation}
\alias{PlotASCA.Permutation}
\title{Plot ASCA permutation}
\usage{
PlotASCA.Permutation(imgName, format="png", dpi=72, width=NA)
}
\description{
Plot plsda classification performance using different components
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
