% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{PlotAscaImpVar}
\alias{PlotAscaImpVar}
\title{Plot the important variables for each factor}
\usage{
PlotASCAImpVar(imgName, format="png", dpi=72, width=NA, type)
}
\arguments{
\item{type}{select model a, b, or ab}
}
\description{
Plot the important variables for each factor
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
