% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{PlotCmpdSummary}
\alias{PlotCmpdSummary}
\title{Plot compound summary
change to use dataSet$proc instead of dataSet$orig in
case of too many NAs}
\usage{
PlotCmpdSummary(
  mSetObj = NA,
  cmpdNm,
  meta = "NA",
  meta2 = "NA",
  count = 0,
  format = "png",
  dpi = default.dpi,
  width = NA
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{cmpdNm}{Input the name of the compound to plot}

\item{meta}{meta is "NA"}

\item{meta2}{only applicable for multifac module, secondary factor}

\item{count}{img count number}

\item{format}{Input the format of the image to create}

\item{dpi}{Input the dpi of the image to create}

\item{width}{Input the width of the image to create}
}
\description{
Plot compound summary
change to use dataSet$proc instead of dataSet$orig in
case of too many NAs
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
