% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_meta.R
\name{PlotCochranQ}
\alias{PlotCochranQ}
\title{Compute Cochran's Q test}
\usage{
PlotCochranQ(imgNm)
}
\arguments{
\item{imgNm}{Image name to output}
}
\description{
Plot Q-Q plot for estimating statistical heterogeneity. Cochran’s Q test is calculated as the weighted sum of squared differences between individual study effects and the pooled effect across studies. When the estimated Q values have approximately a chi-squared distribution, it suggests FEM assumption is appropriate. If it deviates significantly from a chi-squared distribution, REM should usually be used.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
