% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{PlotCorr}
\alias{PlotCorr}
\title{Pattern hunter, correlation plot}
\usage{
PlotCorr(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA)
}
\description{
Plot correlation
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
