% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{PlotCorrHeatMap}
\alias{PlotCorrHeatMap}
\title{Pattern hunter, corr heatmap}
\usage{
PlotCorrHeatMap(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, target, cor.method, colors, viewOpt, fix.col, no.clst, top,
  topNum)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{target}{Input "row" to select features, or "col" to select samples.}

\item{cor.method}{Indicate the correlation method, 'pearson', 'spearman', or 'kendall'.}

\item{colors}{Indicate the colors for the heatmap, "bwm" for default, "gbr" for red/green, "heat" for heat colors,
"topo" for topo colors, and "gray" for gray scale.}

\item{viewOpt}{Indicate "overview" to get an overview of the heatmap, and "detail" to get a detailed view of the heatmap.}

\item{fix.col}{Logical, fix colors (TRUE) or not (FALSE).}

\item{no.clst}{Logical, indicate if the correlations should be clustered (TRUE) or not (FALSE).
McGill University, Canada
License: GNU GPL (>= 2)}
}
\description{
Plot correlation heatmap
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
}
