% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_covariate.R
\name{PlotCovariateAdjustmentPCA}
\alias{PlotCovariateAdjustmentPCA}
\title{PlotCovariateAdjustmentPCA}
\usage{
PlotCovariateAdjustmentPCA(
  mSetObj = NA,
  covariate,
  imgName = "covariate_pca",
  format = "png",
  dpi = 96
)
}
\arguments{
\item{mSetObj}{Input the name of the bound mSet Object}

\item{covariate}{Name of the primary covariate to show in the plot}

\item{imgName}{Image file name}

\item{format}{Image format ("png" or "pdf")}

\item{dpi}{Resolution for PNG}
}
\value{
1 on success, 0 on error
}
\description{
Generate side-by-side PCA plots showing data before and after covariate adjustment
}
