% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotDetailROC}
\alias{PlotDetailROC}
\title{Plot detailed ROC}
\usage{
PlotDetailROC(mSetObj = NA, imgName, thresh, sp, se, dpi = 72, format = "png")
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Input a name for the plot}

\item{thresh}{Input the threshold}

\item{sp}{Specificity}

\item{se}{Sensitivity}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{format}{Select the image format, "png", or "pdf".}
}
\description{
Plot detailed ROC
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
