% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{PlotEBAM.A0}
\alias{PlotEBAM.A0}
\title{For EBAM analysis}
\usage{
PlotEBAM.A0(imgName, format, dpi, width)
}
\description{
plot ebam a0 plot also return the analSet$ebam.a0 object 
so that the suggested a0 can be obtained
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
