% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{PlotEBAM.Cmpd}
\alias{PlotEBAM.Cmpd}
\title{Plot EBAM}
\usage{
PlotEBAM.Cmpd(mSetObj=NA, imgName, format, dpi, width)
}
\description{
Plot EBAM
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
