% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{PlotEIC}
\alias{PlotEIC}
\title{Plot EIC}
\usage{
PlotEIC(
  raw_data,
  rt_mn,
  rt_mx,
  mz_mn,
  mz_mx,
  aggreg = "sum",
  format = "png",
  dpi = 72,
  width = 9
)
}
\arguments{
\item{raw_data}{The object created using the InspectRawMSData function,
containing the raw MS data.}

\item{rt_mn}{Numeric, specify the minimum bound of the retention time range.}

\item{rt_mx}{Numeric, specify the maximum bound of the retention time range.}

\item{mz_mn}{Numeric, specify the minimum bound of the m/z range.}

\item{mz_mx}{Numeric, specify the maximum bound of the m/z range.}

\item{aggreg}{Character, if "sum", creates a total ion chromatogram. 
If "max", creates a base peak chromatogram. By default it is set 
to "sum".}

\item{format}{Character, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}
}
\description{
This functionn creates an extracted ion chromatogram (EIC) for a specific
m/z and retention time. This is used for quality-control of raw m/s data.
}
