% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_graphics.R
\name{PlotEnrichNet.Overview}
\alias{PlotEnrichNet.Overview}
\title{Barplot height is enrichment fold change}
\usage{
PlotEnrichNet.Overview(folds, pvals, layoutOpt = layout.fruchterman.reingold)
}
\description{
Used in higher functions, the color is based on p values
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
