% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_stats.R
\name{PlotEnrichPieChart}
\alias{PlotEnrichPieChart}
\title{Create a pie chart for compound classes}
\usage{
PlotEnrichPieChart(
  mSetObj = NA,
  enrichType,
  imgName,
  format = "png",
  dpi = 72,
  width = 8,
  maxClass = 15,
  colPal = "Set1"
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Numeric, input the width, the default is 8.}

\item{maxClass}{Numeric, input the maximum number of lipid classes to
include in the pie-chart. By default this is set to 15.}

\item{colPal}{Character, input the preferred R Color Brewer palette to be
used for the pie chart. By default this is set to "Set1".}
}
\description{
This function creates a pie-chart for compound classes
in the enrichment analysis if the library is based on chemical ontologies.
}
