% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{PlotHCTree}
\alias{PlotHCTree}
\title{Plot Dendrogram}
\usage{
PlotHCTree(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA,
  smplDist, clstDist)
}
\description{
Dendogram
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
