% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{PlotHeatMap}
\alias{PlotHeatMap}
\title{Create Heat Map Plot}
\usage{
PlotHeatMap(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA,
  dataOpt, scaleOpt, smplDist, clstDist, palette, viewOpt = "detail",
  rowV = T, colV = T, var.inx = NA, border = T, grp.ave = F)
}
\description{
Plot a heatmap based on results from t-tests/ANOVA, VIP or randomforest
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
