% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotImpBiomarkers}
\alias{PlotImpBiomarkers}
\title{Plot selected compounds by their percentage frequency}
\usage{
PlotImpBiomarkers(mSetObj=NA, imgName, format="png", dpi=default.dpi, 
mdl.inx, measure = "freq", feat.num = 15)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Input a name for the plot}

\item{format}{elect the image format, "png", of "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{mdl.inx}{Model index, -1 selects the model with the best AUC, input 1-6 to 
view the important features of one of the top six models}

\item{measure}{Choose to rank features by the frequency of being selected "freq", or the 
mean importance measure "mean"}

\item{feat.num}{Input the number of features to include in the plot, by default it is 15.}
}
\description{
Plot the important variables of single biomarker model ranked by order of importance
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
