% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotImpVar}
\alias{PlotImpVar}
\title{Plot PLS important variables,}
\usage{
PlotImpVar(mSetObj = NA, imp.vec, xlbl, feat.num = 15, color.BW = FALSE)
}
\description{
Plot PLS important variables, BHan: added bgcolor parameter for B/W color
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
