% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_covariate.R
\name{PlotImpVarMeta}
\alias{PlotImpVarMeta}
\title{Plot PLS important variables,}
\usage{
PlotImpVarMeta(
  mSetObj = NA,
  imp.vec,
  xlbl,
  feat.num = 15,
  color.BW = FALSE,
  type = "meta"
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imp.vec}{Input the vector of important variables}

\item{xlbl}{Input the x-label}

\item{feat.num}{Numeric, set the feature numbers, default is set to 15}

\item{color.BW}{Use black-white for plot (T) or colors (F)}

\item{type}{type, default is "type"}
}
\description{
Plot PLS important variables, BHan: added bgcolor parameter for B/W color
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
