% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_kegg.R
\name{PlotInmexPath}
\alias{PlotInmexPath}
\title{Plot integrated methods pathway analysis}
\usage{
PlotInmexPath(mSetObj=NA, path.id, width, height)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{path.id}{Input the ID of the pathway to plot.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{height}{Input the height of the image to create.}
}
\description{
Only update the background info for matched node
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
