% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{PlotIntegPaths}
\alias{PlotIntegPaths}
\title{PlotIntegPaths}
\usage{
PlotIntegPaths(mSetObj = NA, format = "png", dpi = 72, width = 9,
  labels = FALSE)
}
\arguments{
\item{format}{Character, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}

\item{labels}{Logical, default is FALSE. If set to true, it will
label the all of the identified pathways.}
}
\description{
Plot both the original mummichog and the GSEA results
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
