% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{PlotInteraction}
\alias{PlotInteraction}
\title{Plot ASCA interaction plots}
\usage{
PlotInteraction(imgName, format="png", dpi=72, colowBW, width=NA)
}
\description{
Plot ASCA interaction plots
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
