% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotKEGGPath}
\alias{PlotKEGGPath}
\title{Plot KEGG pathway}
\usage{
PlotKEGGPath(mSetObj = NA, pathName, format = "png", width = NA,
  dpi = 72)
}
\description{
Orthogonal PLS-DA (from ropls)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
