% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_mb.R
\name{PlotMBTimeProfile}
\alias{PlotMBTimeProfile}
\title{Plot MB Time Profile}
\usage{
PlotMBTimeProfile(mSetObj = NA, cmpdNm, format = "png", dpi = 72,
  width = NA)
}
\description{
Plot MB Time Profile
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
