% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{PlotMetaCorrHeatmap}
\alias{PlotMetaCorrHeatmap}
\title{Generate correlation heatmap for metadata}
\usage{
PlotMetaCorrHeatmap(
  mSetObj = NA,
  cor.opt = "pearson",
  imgName,
  format = "png",
  dpi = 96,
  width = NA,
  cor.method = "univariate",
  colorGradient = "default",
  interactive = FALSE
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{cor.opt}{Method for computing correlation coefficient}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{cor.method}{Character, method to compute correlation for mixed data types: "univariate" (default) or similar.}

\item{colorGradient}{Character, color scheme for the heatmap. Default is "default".}

\item{interactive}{Logical, whether to return an interactive plotly figure. Default is FALSE.}
}
\description{
Plot correlation coefficients between metadata
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
