% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_meta.R
\name{PlotMetaPCA}
\alias{PlotMetaPCA}
\title{Plot PCA plot for meta-analysis samples}
\usage{
PlotMetaPCA(imgNm, dpi, format, factor)
}
\arguments{
\item{imgNm}{name of the image to output}
}
\description{
Plot PCA plot for meta-analysis samples
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
