% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PlotMetaVenn}
\alias{PlotMetaVenn}
\title{Meta-Analysis: Plot Venn Diagram}
\usage{
PlotMetaVenn(mSetObj = NA, imgNM = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{imgNM}{Input the name of the created Venn Diagram}
}
\description{
This function plots a venn diagram of the individual studies.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
