% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotMetpaPath}
\alias{PlotMetpaPath}
\title{Plot KEGG pathway}
\usage{
PlotMetpaPath(
  mSetObj = NA,
  pathName,
  width = NA,
  height = NA,
  format = "png",
  dpi = default.dpi
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{pathName}{Input the name of the selected KEGG pathway}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{height}{height value for the image.}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}
}
\description{
Plot KEGG pathway
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
