% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{PlotMS2Spectra}
\alias{PlotMS2Spectra}
\title{Plot selected M2 spectra for an m/z feature}
\usage{
PlotMS2Spectra(ms2, spectra = 1)
}
\arguments{
\item{ms2}{Spectrum2 class object containing all of the spectra for the selected
m/z feature.}
}
\description{
This function creates a plot of the user selected precursor m/z.
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
