% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_graphics.R
\name{PlotMSEA.Overview}
\alias{PlotMSEA.Overview}
\title{Plot MSEA overview}
\usage{
PlotMSEA.Overview(folds, pvals)
}
\arguments{
\item{folds}{Input the fold-change values}

\item{pvals}{Input the p-values}
}
\description{
Barplot height is enrichment fold change
color is based on p values, used in higher functions
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
