% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{plotMSfeature}
\alias{plotMSfeature}
\title{plotMSfeature}
\usage{
plotMSfeature(FeatureNM, format = "png", dpi = 72, width = NA)
}
\description{
plotMSfeature is used to plot MS feature stats for different groups
}
\author{
Zhiqiang Pang
}
