% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PlotMumVolcano}
\alias{PlotMumVolcano}
\title{PlotMumVolcano}
\usage{
PlotMumVolcano(
  mSetObj = NA,
  imgName,
  dpi = 150,
  format = "png",
  pval.cutoff = 0.05,
  fc.cutoff = 0
)
}
\arguments{
\item{format}{Image format}
}
\description{
PlotMumVolcano
}
