% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Utility function to plot network for analysis report (CreateGraph)}
\usage{
PlotNetwork(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA)
}
\description{
Utility function to plot network for analysis report (CreateGraph)
}
