% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotOPLS.Imp}
\alias{PlotOPLS.Imp}
\title{PlotOPLS.Imp
OPLS VIP plot}
\usage{
PlotOPLS.Imp(
  mSetObj = NA,
  imgName,
  format = "png",
  dpi = 72,
  width = NA,
  type = "vip",
  feat.nm = "tscore",
  feat.num = 15,
  color.BW = FALSE
)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{imgName}{image name}

\item{format}{format}

\item{dpi}{dpi}

\item{width}{width}

\item{type}{type}

\item{feat.nm}{feat.nm}

\item{feat.num}{feat.num}

\item{color.BW}{color of background}
}
\description{
PlotOPLS.Imp
OPLS VIP plot
}
