% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotOPLS.Permutation}
\alias{PlotOPLS.Permutation}
\title{Perform OPLS-DA permutation}
\usage{
PlotOPLS.Permutation(mSetObj = NA, imgName, format = "png", dpi = 72,
  num = 100, width = NA)
}
\description{
Orthogonal PLS-DA (from ropls) 
perform permutation, using training classification accuracy as
indicator, for two or multi-groups
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
