% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotOPLS.Splot}
\alias{PlotOPLS.Splot}
\title{S-plot for OPLS-DA}
\usage{
PlotOPLS.Splot(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, plotType)
}
\description{
Orthogonal PLS-DA (from ropls) 
S-plot for important features from OPLS-DA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
