% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotOPLS2DScore}
\alias{PlotOPLS2DScore}
\title{Create OPLS-DA score plot}
\usage{
PlotOPLS2DScore(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, inx1, inx2, reg = 0.95, show = 1, grey.scale = 0)
}
\description{
Orthogonal PLS-DA (from ropls) score plot
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
