% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_graphics.R
\name{PlotORAMembership}
\alias{PlotORAMembership}
\title{Plot the compound x metabolite-set membership heatmap.}
\usage{
PlotORAMembership(
  mSetObj = NA,
  imgName,
  format = "png",
  dpi = default.dpi,
  width = NA,
  topN = 100
)
}
\arguments{
\item{mSetObj}{Input mSetObj}

\item{imgName}{Filename root (gets dpi + format suffix appended)}

\item{format}{png / pdf}

\item{dpi}{default.dpi}

\item{width}{optional inches}

\item{topN}{Cap on the number of sets shown; default 100 matches the viewer}
}
\description{
Static PNG companion to the Plotly viewer on OraView. Rows are
the input compounds that hit at least one of the top-N enriched sets;
columns are the sets ranked by p-value. Cells are binary (compound in set
or not). Saved as Cairo PNG so the Dashboard gallery + Sweave report + slide
builder pick it up alongside the existing ora bar chart / ora_dot.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
