% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotPathSummary}
\alias{PlotPathSummary}
\title{Plot a scatterplot (circle) overview of the matched pathways}
\usage{
PlotPathSummary(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, x, y)
}
\description{
x axis is the pathway impact factor
y axis is the p value (from ORA or globaltest) 
return the circle information
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
