% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotPathSummarySlide}
\alias{PlotPathSummarySlide}
\title{Plot pathway summary for slide reports using ggplot2 style}
\usage{
PlotPathSummarySlide(
  mSetObj = NA,
  imgName,
  format = "png",
  dpi = 96,
  width = NA,
  num.labels = 5
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi.}

\item{width}{Input the width.}

\item{num.labels}{Number of top pathways to label (default 5)}
}
\description{
Creates a modern ggplot2-based pathway overview plot for slide reports
Same style as PlotPeaks2Paths (Mummichog) with yellow-to-red gradient and labeled top pathways
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
