% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{PlotPathwayMetaAnalysis}
\alias{PlotPathwayMetaAnalysis}
\title{PlotPathwayMetaAnalysis}
\usage{
PlotPathwayMetaAnalysis(
  mSetObj = NA,
  imgName,
  plotType = "heatmap",
  heatmap_colorType = "brewer",
  heatmap_palette = "RdYlBu",
  heatmap_interactive = FALSE,
  heatmap_square = TRUE,
  heatmap_allPaths = TRUE,
  heatmap_npaths = 25,
  heatmap_vertical = TRUE,
  heatmap_fontSize = 9,
  pvalCutoff = 0.05,
  overlap = 0.25,
  networkType = "static",
  layout = "kk",
  net_palette = "YlOrRd",
  netTextSize = 2.5,
  netPlotSize = 7.5,
  bubble_interactive = FALSE,
  bubbleMaxPaths = 15,
  bubble_colorType = "brewer",
  bubble_palette = "RdBu",
  bubbleFontSize = 9,
  bubblePlotSize = 7,
  format = "png",
  width = 7,
  height = 5,
  dpi = 300
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object.}

\item{imgName}{name of image.}

\item{plotType}{Use "heatmap" to create a heatmap summary, "network" to create 
a network summary, or "bubble" to create a bubble plot summary of the meta-analysis
results.}

\item{heatmap_colorType}{Character, "brewer" for R Color Brewer scales or
"viridis" for viridis color scales. Used for creating the heatmap
color scheme.}

\item{heatmap_palette}{Character, input the preferred color palette according
to R Color Brewer or viridis (e.g. "RdBu").}

\item{heatmap_interactive}{Boolean. FALSE to create a non-interactive plot
and TRUE for plotly generated interactive plot.}

\item{heatmap_square}{Boolean. TRUE for the heatmap to be squares versus
rectangles (FALSE).}

\item{heatmap_allPaths}{Boolean. TRUE to use all paths when plotting the heatmap.
FALSE to use a subset of paths, number defined in npaths.}

\item{heatmap_npaths}{Numeric. The number of pathways to subset the pathway
results.}

\item{heatmap_vertical}{Boolean. TRUE, heatmap plot will be vertical. FALSE, heatmap plot
will be horizontal.}

\item{heatmap_fontSize}{Numeric, input the preferred font size to be used in the heatmap
plot.}

\item{pvalCutoff}{The size of the nodes in the network correspond to the number of
studies in which that pathway was significant. This pvalCutoff (Numeric) is thus used
to determine whether or not a pathway was found to be significant in each
individual study.}

\item{overlap}{Numeric, this number is used to create edges between the nodes.
By default it is set to 0.25, meaning that if 2 pathways (nodes) share 25% of
the same compounds/empirical compounds, they will be connected by a node.}

\item{networkType}{Character, "static" to create a static image or
"interactive" to create an interactive network saved as an html 
in your working directory.}

\item{layout}{Character, layout from ggraph. "kk" for the spring-based algorithm by Kamada and Kawai
as default. "drl" for force directed algorithm from the DrL toolbox. "lgl" for Large Graph Layout. "fr" for
force-directed of Fruchterman and Reingold.}

\item{net_palette}{Character, input the color code for the nodes in the network. Default is
"YlOrRd". Uses the hcl palettes from the grDevices. Use hcl.pals()
to view the name of all available palettes.}

\item{netTextSize}{Numeric, input the preferred font size to be used in the network
plot.}

\item{netPlotSize}{Numeric, input the preferred dimensions (in inches) of the network
to be saved.}

\item{bubble_interactive}{logical}

\item{bubbleMaxPaths}{maximum number of pathways}

\item{bubble_colorType}{Character, "brewer" for R Color Brewer scales or 
"viridis" for viridis color scales. Used for creating the bubble plot
color scheme.}

\item{bubble_palette}{Character, use two/three colors max if using R ColorBrewer palettes
for pleasing looking plots.}

\item{bubbleFontSize}{font size of the bubble plot}

\item{bubblePlotSize}{plot size of the bubble plot}

\item{format}{format of the image}

\item{width}{width of the image}

\item{height}{height of the image}

\item{dpi}{dpi of the image}
}
\description{
Function to create summary plot of MS Peaks to Paths
meta-analysis at the pathway level. This function creates a summary plot of the
MS Peaks to Paths meta-analysis at the pathway level. The plot
can either be a heatmap or a network, both of which can
be made interactive. 
NETWORK: The size of the nodes in the network correspond to the number of
studies in which that pathway was significant. The color of the nodes correspond
to the meta-p-value for each pathway, with (default coloring) red being the most 
significant and yellow the least.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
