% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_batch_check.R
\name{PlotPCA.overview}
\alias{PlotPCA.overview}
\title{Scatter plot colored by different batches}
\usage{
PlotPCA.overview(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA)
}
\description{
Scatter plot colored by different batches
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
