% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPCA2DScore}
\alias{PlotPCA2DScore}
\title{Create 2D PCA score plot}
\usage{
PlotPCA2DScore(mSetObj=NA, imgName, format="png", dpi=72, width=NA, pcx, pcy, reg = 0.95, show=1, grey.scale = 0)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{pcx}{Specify the principal component on the x-axis}

\item{pcy}{Specify the principal component on the y-axis}

\item{reg}{Numeric, input a number between 0 and 1, 0.95 will display the 95 percent confidence regions, and 0 will not.}

\item{show}{Display sample names, 1 = show names, 0 = do not show names.}

\item{grey.scale}{Use grey-scale colors, 1 = grey-scale, 0 = not grey-scale.}
}
\description{
Rotate PCA analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
