% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPCA3DLoading}
\alias{PlotPCA3DLoading}
\title{PlotPCA3DLoading}
\usage{
PlotPCA3DLoading(mSetObj = NA, imgName, format = "json", inx1, inx2, inx3)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{imgName}{imgName}

\item{format}{format}

\item{inx1}{inx1}

\item{inx2}{inx2}

\item{inx3}{inx3}
}
\description{
PlotPCA3DLoading
}
