% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPCA3DScore}
\alias{PlotPCA3DScore}
\title{Create 3D PCA score plot}
\usage{
PlotPCA3DScore<- function(mSetObj=NA, imgName, format="json", inx1, inx2, inx3)
}
\description{
Rotate PCA analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
