% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{PlotPCAPairSummaryMeta}
\alias{PlotPCAPairSummaryMeta}
\title{Plot PCA Pair Summary with Metadata}
\usage{
PlotPCAPairSummaryMeta(mSetObj=NA, imgName, format="png", dpi=default.dpi, width=NA, pc.num, meta, metaShape=NULL)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Base name of the output image file}

\item{format}{Output image format, default is "png"}

\item{dpi}{Image resolution in dots per inch, default is default.dpi}

\item{width}{Image width in inches; use NA for automatic sizing}

\item{pc.num}{Number of principal components to include}

\item{meta}{Metadata column name used for coloring}

\item{metaShape}{Optional metadata column name used for point shape}
}
\description{
Generate PCA pairwise summary plots with metadata overlays.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
