% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PlotPeaks2Paths}
\alias{PlotPeaks2Paths}
\title{PlotPeaks2Paths}
\usage{
PlotPeaks2Paths(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = 9, labels = "default", num_annot = 5)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object}

\item{imgName}{Input a name for the plot}

\item{format}{Character, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}

\item{Labels}{Character, indicate if the plot should be labeled. By default
it is set to "default", and the 5 top-ranked pathways per each algorithm will be plotted.
Users can adjust the number of pathways to be annotated per pathway using the "num_annot" 
parameter.}
}
\description{
Plots either the original mummichog or GSEA results.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
