% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLS.Classification}
\alias{PlotPLS.Classification}
\title{Plot PLS-DA classification performance using different components}
\usage{
PlotPLS.Classification(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA)
}
\description{
Plot plsda classification performance using different components
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
