% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLS.Imp}
\alias{PlotPLS.Imp}
\title{Plot PLS important features}
\usage{
PlotPLS.Imp(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA,
  type, feat.nm, feat.num, color.BW = FALSE)
}
\description{
Plot PLS important features, BHan: added bgcolor parameter for B/W color
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
