% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLS2DScore}
\alias{PlotPLS2DScore}
\title{Plot PLS score plot}
\usage{
PlotPLS2DScore(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, inx1, inx2, reg = 0.95, show = 1, grey.scale = 0,
  use.sparse = FALSE)
}
\description{
Plot PLS score plot
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
