% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLS3DScore}
\alias{PlotPLS3DScore}
\title{Plot PLS 3D score plot}
\usage{
PlotPLS3DScore(mSetObj = NA, imgName, format = "json", inx1, inx2, inx3)
}
\description{
Plot PLS 3D score plot
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
