% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLSBiplot}
\alias{PlotPLSBiplot}
\title{Plot PLS-DA biplot}
\usage{
PlotPLSBiplot(mSetObj=NA, imgName, format="png", dpi=default.dpi, width=NA, inx1, inx2, topnum=10)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Base name of the output image file}

\item{format}{Output image format, default is "png"}

\item{dpi}{Image resolution in dots per inch, default is default.dpi}

\item{width}{Image width in inches; use NA for automatic sizing}

\item{inx1}{Index of the first component for x-axis}

\item{inx2}{Index of the second component for y-axis}

\item{topnum}{Number of top variables to label in the biplot}
}
\description{
Generate a PLS-DA biplot that overlays sample scores and top contributing
variables based on VIP values for selected components.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
