% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLSLoading}
\alias{PlotPLSLoading}
\title{Plot PLS loading plot, also set the loading matrix for display}
\usage{
PlotPLSLoading(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, inx1, inx2, plotType, lbl.feat = 1)
}
\description{
Plot PLS loading plot, also set the loading matrix for display
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
