% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLSPairSummary}
\alias{PlotPLSPairSummary}
\title{Plot PLS pairwise summary}
\usage{
PlotPLSPairSummary(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, pc.num)
}
\description{
Plot PLS pairwise summary
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
