% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{PlotPowerEffectSize}
\alias{PlotPowerEffectSize}
\title{Plot power effect size}
\usage{
PlotPowerEffectSize(mSetObj, imgName, format, dpi, width)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Specify the name to save the image as.}

\item{format}{Specify the format of the image to save it as, either "png" or "pdf"}

\item{dpi}{Specify the dots-per-inch (dpi). By default it is 72, for publications
the recommended dpi is 300.}

\item{width}{Specify the width of the image. NA specifies a width of 9, 0 specifies a width
of 8, otherwise input a chosen width.}
}
\description{
Create power effect size plot
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
