% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{PlotPowerProfile}
\alias{PlotPowerProfile}
\title{Plot power profile}
\usage{
PlotPowerProfile(mSetObj=NA, fdr.lvl, smplSize, imgName, format, dpi, width)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{fdr.lvl}{Specify the false-discovery rate level.}

\item{smplSize}{Specify the maximum sample size, the number must be between 60-1000.}

\item{imgName}{Specify the name to save the image as.}

\item{format}{Specify the format of the image to save it as, either "png" or "pdf".}

\item{dpi}{Specify the dots-per-inch (dpi). By default it is 72, for publications
the recommended dpi is 300.}

\item{width}{Specify the width of the image. NA specifies a width of 9, 0 specifies a width
of 7, otherwise input a chosen width.}
}
\description{
Plot power profile, specifying FDR level and sample size. It will 
return the image as well as the predicted power at various sample sizes.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
