% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_mb.R
\name{plotProfile}
\alias{plotProfile}
\title{Plot the variable across time points (x)}
\usage{
plotProfile(mSetObj = NA, varName)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{varName}{Input the name of the variable}
}
\description{
Colored by experimental conditions, used in higher function
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
