% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{PlotRF.Classify}
\alias{PlotRF.Classify}
\title{Plot Random Forest}
\usage{
PlotRF.Classify(imgName, format, dpi, width)
}
\description{
Random Forest plot
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
