% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{PlotRF.Outlier}
\alias{PlotRF.Outlier}
\title{Plot Random Forest outliers}
\usage{
PlotRF.Outlier(imgName, format, dpi, width)
}
\description{
Random Forest plot of outliers
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
