% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{PlotRF.VIP}
\alias{PlotRF.VIP}
\title{Plot Random Forest variable importance}
\usage{
PlotRF.VIP(imgName, format, dpi, width)
}
\description{
Random Forest plot of variable importance ranked by MeanDecreaseAccuracy
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
