% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotROC.LRmodel}
\alias{PlotROC.LRmodel}
\title{Plot ROC for the logistic regression model}
\usage{
PlotROC.LRmodel(mSetObj = NA, imgName, format = "png", dpi = 72,
  show.conf = FALSE, sp.bin = 0.01)
}
\description{
Plot ROC for the logistic regression model
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
