% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotROCMulti}
\alias{PlotROCMulti}
\title{Plot ROC}
\usage{
PlotROCMulti(mSetObj=NA, imgName, format="png", dpi=72, mdl.inx, avg.method, show.conf, show.holdout, focus="fpr", cutoff = 1.0)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", of "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{mdl.inx}{Model index, 0 means to compare all models, input 1-6 to plot a ROC curve for one of the top six models}

\item{avg.method}{Input the method to compute the average ROC curve, either "threshold", "vertical" or "horizontal"}

\item{show.conf}{Logical, if 1, show confidence interval, if 0 do not show}

\item{show.holdout}{Logical, if 1, show the ROC curve for hold-out validation, if 0 do not show}

\item{focus}{"fpr"}

\item{cutoff}{Input the threshold to limit the calculation of the ROC curve, the number must be between 0 and 1.}
}
\description{
Pred and auroc are lists containing predictions
and labels from different cross-validations
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
