% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{PlotRSVM.Classification}
\alias{PlotRSVM.Classification}
\title{Recursive Support Vector Machine (R-SVM) plot}
\usage{
PlotRSVM.Classification(imgName, format, dpi, width)
}
\description{
Plot recursive SVM classification
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
