% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{PlotRSVM.Cmpd}
\alias{PlotRSVM.Cmpd}
\title{Recursive Support Vector Machine (R-SVM) plot of important variables}
\usage{
PlotRSVM.Cmpd(imgName, format, dpi, width)
}
\description{
Plot recursive SVM variables of importance
if too many, plot top 15
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
