% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{PlotSAM.Cmpd}
\alias{PlotSAM.Cmpd}
\title{Plot SAM}
\usage{
PlotSAM.Cmpd(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA)
}
\description{
Plot SAM with positive and negative metabolite sets
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
