% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{PlotSAM.FDR}
\alias{PlotSAM.FDR}
\title{Plot SAM Delta Plot}
\usage{
PlotSAM.FDR(mSetObj = NA, delta, imgName, format = "png", dpi = 72,
  width = NA)
}
\description{
Plot SAM Delta Plot (FDR)
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
