% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{PlotSampleNormSummary}
\alias{PlotSampleNormSummary}
\title{Two plot summary plot: Sample View of before and after normalization}
\usage{
PlotNormSummary(mSetObj, imgName, format="png", dpi=72, width=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", of "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}
}
\description{
For each plot, the top is a density plot and the bottom is a box plot.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}, Jasmine Chong 
McGill University, Canada
}
