% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{PlotSelectedFeature}
\alias{PlotSelectedFeature}
\title{Create a box-plot of a feature's expression pattern across the different datasets}
\usage{
PlotSelectedFeature(mSetObj = NA, gene.id)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object.}

\item{gene.id}{Input the name of the selected feature.}
}
\description{
This function plots a box-plot of the expression pattern of a user-selected feature
across the different datasets included in meta-analysis.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
