% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{PlotSigVar}
\alias{PlotSigVar}
\title{Supporting function for plotting important variables for each factor}
\usage{
PlotSigVar(x, y, xline, yline, title)
}
\arguments{
\item{x}{Input the X variable}

\item{y}{Input the Y variable}

\item{xline}{Input the xline}

\item{yline}{Input the yline}

\item{title}{Input the title}
}
\description{
Supporting function for plotting important variables for each factor
note, by control xpd to plot legend outside the plotting area
without using layout
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
