% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{plotSingleTIC}
\alias{plotSingleTIC}
\title{plotSingleTIC}
\usage{
plotSingleTIC(filename, imageNumber, format = "png", dpi = 72, width = NA)
}
\description{
plotSingleTIC is used to plot single TIC
}
\author{
Zhiqiang Pang
}
