% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{PlotSOM}
\alias{PlotSOM}
\title{SOM Plot}
\usage{
PlotSOM(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA)
}
\description{
Plot SOM map for  less than 20 clusters
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
