% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{PlotSPLS2DScore}
\alias{PlotSPLS2DScore}
\title{Score Plot SPLS-DA}
\usage{
PlotSPLS2DScore(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, inx1, inx2, reg = 0.95, show = 1, grey.scale = 0)
}
\description{
Sparse PLS-DA (from mixOmics) score plot
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
