% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{PlotSPLS3DScore}
\alias{PlotSPLS3DScore}
\title{3D SPLS-DA score plot}
\usage{
PlotSPLS3DScore(mSetObj = NA, imgName, format = "json", inx1 = 1,
  inx2 = 2, inx3 = 3)
}
\description{
Sparse PLS-DA (from mixOmics) 3D score plot
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
