% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{PlotSPLSDA.Classification}
\alias{PlotSPLSDA.Classification}
\title{Create SPLS-DA classification plot}
\usage{
PlotSPLSDA.Classification(mSetObj = NA, imgName, validOpt = "Mfold",
  format = "png", dpi = 72, width = NA)
}
\description{
Sparse PLS-DA (from mixOmics) plot of 
classification performance using different components
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
