% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{PlotSPLSLoading}
\alias{PlotSPLSLoading}
\title{Create SPLS-DA loading plot}
\usage{
PlotSPLSLoading(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, inx, viewOpt = "detail")
}
\description{
Sparse PLS-DA (from mixOmics) loading plot
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
