% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{PlotSPLSPairSummary}
\alias{PlotSPLSPairSummary}
\title{Plot SPLS-DA}
\usage{
PlotSPLSPairSummary(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, pc.num)
}
\description{
Sparse PLS-DA (from mixOmics) pairwise summary plot
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
