% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{PlotStaticCorrHeatMap}
\alias{PlotStaticCorrHeatMap}
\title{Create high resolution static HeatMap for download only}
\usage{
PlotStaticCorrHeatMap(
  mSetObj = NA,
  imgName,
  format = "png",
  dpi = default.dpi,
  width = NA,
  target,
  cor.method,
  colors,
  viewOpt,
  fix.col,
  no.clst,
  corrCutoff = 0
)
}
\description{
'@param #same as PlotCorrHeatMap
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
