% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{PlotStaticHeatMap}
\alias{PlotStaticHeatMap}
\title{Create high resolution static HeatMap for download only}
\usage{
PlotStaticHeatMap(
  mSetObj = NA,
  imgName,
  format = "png",
  dpi = 72,
  width = NA,
  dataOpt,
  scaleOpt,
  smplDist,
  clstDist,
  palette,
  fzCol,
  fzRow,
  viewOpt = "detail",
  rowV = T,
  colV = T,
  var.inx = NULL,
  border = T,
  grp.ave = F,
  show.legend = T,
  show.annot.legend = T,
  includeRowNames = T
)
}
\description{
'@param #same as PlotHeatMap
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
