% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{PlotStaticHeatMap2}
\alias{PlotStaticHeatMap2}
\title{Create high resolution static HeatMap for download only}
\usage{
PlotStaticHeatMap2(
  mSetObj = NA,
  imgName,
  dataOpt = "norm",
  scaleOpt = "row",
  format = "png",
  dpi = default.dpi,
  width = NA,
  smplDist = "pearson",
  clstDist = "average",
  colorGradient = "bwm",
  fzCol,
  fzRow,
  viewOpt = "overview",
  rankingMethod = "mean",
  topFeature = 2000,
  useTopFeature = F,
  drawBorder = T,
  show.legend = T,
  show.annot.legend = T,
  includeRowNames = T
)
}
\description{
'@param #same as PlotHeatMap2
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
