% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pca_anova2.R
\name{plotVennDiagram}
\alias{plotVennDiagram}
\title{Plot Venn diagram}
\usage{
plotVennDiagram(object, include = "both", names, mar = rep(0, 4),
  cex = 1.2, lwd = 1, circle.col, counts.col, show.include, ...)
}
\description{
Capabilities for multiple counts and colors by Francois Pepin
}
\author{
Gordon Smyth, James Wettenhall
4 July 2003.  Last modified 12 March 2010
}
