% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{PlotVolcano}
\alias{PlotVolcano}
\title{Create volcano plot}
\usage{
PlotVolcano(mSetObj=NA, imgName, plotLbl, format="png", dpi=72, width=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{imgName}{Input a name for the plot}

\item{plotLbl}{Logical, plot labels, 1 for yes and 0 for no.}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}
}
\description{
For labelling interesting points, it is defined by the following rules:
need to be signficant (sig.inx) and or 2. top 5 p, or 2. top 5 left, or 3. top 5 right.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
