% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{PlotXIC}
\alias{PlotXIC}
\title{PlotXIC}
\usage{
PlotXIC(featureNum, format = "png", dpi = 72, width = NA)
}
\description{
PlotXIC is used to plot both MS XIC/EIC features of different group and samples
}
\author{
Zhiqiang Pang
}
