% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PLSR.Anal}
\alias{PLSR.Anal}
\title{PLS analysis using oscorespls (Orthogonal scores algorithm)
so that VIP can be calculated
note: the VIP is calculated only after PLSDA-CV is performed
to determine the best # of comp. used for VIP}
\usage{
PLSR.Anal(mSetObj = NA, reg = FALSE)
}
\description{
PLS analysis using oscorespls
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
