% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Predict.class}
\alias{Predict.class}
\title{Get predicted class probability}
\usage{
Predict.class(x.train, y.train, x.test, clsMethod = "pls", lvNum, imp.out = F)
}
\arguments{
\item{x.train}{Input the x training samples}

\item{y.train}{Input the y training samples}

\item{x.test}{Input the x testing samples}

\item{clsMethod}{Se the classification method, default is set to pls}

\item{lvNum}{Input the number of levels}

\item{imp.out}{Logical, set to F by default}
}
\description{
Get predicted class probability
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
