% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{PrepareIntegData}
\alias{PrepareIntegData}
\title{Prepare integrated data}
\usage{
PrepareIntegData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Used for the pathinteg module.
}
