% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{PrepareLatex}
\alias{PrepareLatex}
\title{Remove HTML tag}
\usage{
PrepareLatex(stringVec)
}
\description{
Removes html tag
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
