% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{PrepareMetaPath}
\alias{PrepareMetaPath}
\title{PrepareMetaPath}
\usage{
PrepareMetaPath(
  mSetObj = NA,
  mode = "negative",
  ppm = 30,
  version = "v2",
  pcutoff = 0.05,
  rt.type = "seconds",
  dataName,
  dataName2
)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{mode}{ion mode, can be "positive" or "negative"}

\item{ppm}{mass error, default is 30}

\item{version}{mummichog version, can be "v1" or "v2"}

\item{pcutoff}{p value cut-off, default is 0.05}

\item{rt.type}{character, retention time type, can be "minutes" or "seconds"}

\item{dataName}{file name 1 with absolute path}

\item{dataName2}{file name 2 with absolute path or "null"}
}
\description{
PrepareMetaPath
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca} 
Zhiqiang Pang\email{zhiqiang.pang@mail.mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
