% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R
\name{PrepareNetworkData}
\alias{PrepareNetworkData}
\title{Prepare data for network exploration}
\usage{
PrepareNetworkData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Function for the network explorer module, prepares user's data for network exploration.
}
