% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{PreparePDFReport}
\alias{PreparePDFReport}
\title{Create report of analyses}
\usage{
PreparePDFReport(mSetObj = NA, usrName)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{usrName}{Input the name of the user}
}
\description{
Report generation using Sweave
Note: most analyses were already performed, only need to embed
the results to the right place without rerunning the whole analysis
through Sweave. Only some auxilliary info (i.e. time, version etc need to
run in R through Sweave
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
