% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PreparePeakTable4PSEA}
\alias{PreparePeakTable4PSEA}
\title{PreparePeakTable4PSEA}
\usage{
PreparePeakTable4PSEA(mSetObj = NA)
}
\arguments{
\item{mSetObj}{mSet Objective from previous step}
}
\description{
PreparePeakTable4PSEA
}
\author{
Zhiqiang Pang, Jeff Xia
}
