% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PreparePermResult}
\alias{PreparePermResult}
\title{Prepare report for permutation tests}
\usage{
PreparePermResult(perm.vec)
}
\arguments{
\item{perm.vec}{Input permutation vector}
}
\description{
Function to prepare a report for permutation tests, used in higher functions
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
