% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{PreparePrenormData}
\alias{PreparePrenormData}
\title{Prepare data for normalization}
\usage{
PreparePrenormData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Function should always be initialized (new or overwrite previous prenorm object).
}
