% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_data.R
\name{PrepareQueryJson}
\alias{PrepareQueryJson}
\title{Prepare user's query for mapping KEGG Global Metabolic Network}
\usage{
PrepareQueryJson(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
This function prepares the user's data for the 
KEGG Global Metabolic Network
}
\author{
Othman Soufan, Jeff Xia \email{jeff.xia@mcgill.ca}, {othman.soufan@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
