% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PrepareROCData}
\alias{PrepareROCData}
\title{Prepare data for ROC analysis}
\usage{
PrepareROCData(mSetObj = NA, sel.meta = "NA", factor1 = "NA", factor2 = "NA")
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{sel.meta}{Name of the metadata column to use for grouping; default is "NA" to use the default class labels}

\item{factor1}{Name of the first factor level for binary classification; default is "NA"}

\item{factor2}{Name of the second factor level for binary classification; default is "NA"}
}
\description{
Prepare data for ROC analysis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
