% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PrepareROCDetails}
\alias{PrepareROCDetails}
\title{ROC with CI for AUC}
\usage{
PrepareROCDetails(mSetObj = NA, feat.nm)
}
\description{
ROC with CI for AUC
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
