% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_utils.R
\name{PrepareUpsetData}
\alias{PrepareUpsetData}
\title{Prepare data for Upset diagram}
\usage{
PrepareUpsetData(fileNm)
}
\arguments{
\item{fileNm}{file name of the json file output}

\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}
}
\description{
Prepare data for Upset diagram
}
