% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PrepareVennData}
\alias{PrepareVennData}
\title{Prepare data for Venn diagram}
\usage{
PrepareVennData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Prepare data for Venn diagram
}
