% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem_ms.R
\name{processMSMSupload}
\alias{processMSMSupload}
\title{processMSMSupload}
\usage{
processMSMSupload(mSetObj = NA, spectrum)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{spectrum}{spectrum for uploading}
}
\description{
processMSMSupload
}
\author{
Zhiqiang Pang
}
