% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{RankFeatures}
\alias{RankFeatures}
\title{Rank features based on different importance measures}
\usage{
RankFeatures(x.in, y.in, method, lvNum)
}
\description{
Ranks features based on various importance measures,
return imp.vec which contains the importance measures
of unordered features
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
