% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_batch_check.R
\name{Read.BatchCSVdata}
\alias{Read.BatchCSVdata}
\title{Data I/O for batch effect checking}
\usage{
Read.BatchCSVdata(mSetObj = NA, filePath, format, label)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{filePath}{Input the path to the batch files}

\item{format}{Input the format of the batch files}

\item{label}{Input the label-type of the files}
}
\description{
Read multiple user uploaded CSV data one by one
format: row, col
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
