% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{Read.BatchDataTB}
\alias{Read.BatchDataTB}
\title{Data I/O for batch effect checking}
\usage{
Read.BatchDataTB(mSetObj = NA, filePath, format, missingEstimate)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{filePath}{Input the path to the batch files}

\item{format}{Input the format of the batch files}

\item{missingEstimate}{Approach to estimate the missing values}
}
\description{
Read peak data tale.
format: row, col
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
