% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{Read.PeakListData}
\alias{Read.PeakListData}
\title{License: GNU GPL (>= 2)
Constructor to read uploaded user files into the mummichog object}
\usage{
Read.PeakListData(mSetObj=NA, filename, sigmzs, anal.type, cutoff = 0)
}
\arguments{
\item{filename}{Input the path name for the CSV/TXT files to read.}

\item{mummichog}{Input the name of the created mummichog object (see mummichog_init).}
}
\description{
This function handles reading in CSV or TXT files and filling in the mSet object
for mummichog analysis. It makes sure that all necessary columns are present.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
