% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{Read.SignalDriftData}
\alias{Read.SignalDriftData}
\title{Data I/O for signal drift checking}
\usage{
Read.SignalDriftData(mSetObj = NA, filePath, format)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{filePath}{Input the path to the batch files}

\item{format}{Input the format of the batch files}
}
\description{
Read peak data tale.
format: row, col
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
