% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{Read.TextData}
\alias{Read.TextData}
\title{Constructor to read uploaded CSV or TXT files into the dataSet object}
\usage{
Read.TextData(mSetObj=NA, filePath, format, lbl.type)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects).}

\item{filePath}{Input the path name for the CSV/TXT files to read.}

\item{format}{Specify if samples are paired and in rows (rowp), unpaired and in rows (rowu),
in columns and paired (colp), or in columns and unpaired (colu).}

\item{lbl.type}{Specify the data label type, either discrete (disc) or continuous (cont).}
}
\description{
This function handles reading in CSV or TXT files and filling in the dataSet object created using "InitDataObjects".
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}, Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
