% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{Read.TextDataTs}
\alias{Read.TextDataTs}
\title{Read.TextDataTs}
\usage{
Read.TextDataTs(mSetObj = NA, filePath, format = "rowu")
}
\arguments{
\item{mSetObj}{metaboanalyst object, initialized by InitDataObjects("pktable", "ts", FALSE)}

\item{filePath}{file path of data}

\item{format}{format of data table, can be "rowu" or "colu"}
}
\description{
Read.TextDataTs is used to read metabolomics data for co-vairiate analysis
}
