% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{read_MetAnal_RDS}
\alias{read_MetAnal_RDS}
\title{Read RDS files from the internet}
\usage{
read_compound(fileURL, destfile)
}
\arguments{
\item{fileURL}{Name of the file path}

\item{destfile}{Name of the downloaded file to current working directory 

example: cmpd <- read_MetAnal_RDS("http://www.metaboanalyst.ca/resources/libs/syn_nms.rds", destfile="syn_nms.rds")}
}
\description{
Function downloads the required file and reads it only if not already in working directory.
Need to specify the file URL and the destfile.
}
