% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{ReadIndData}
\alias{ReadIndData}
\title{Read in individual data}
\usage{
ReadIndData(mSetObj = NA, dataName, format = "colu")
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{dataName}{Name of inputted dataset.}

\item{format}{Specify if samples are paired and in rows (rowp), unpaired and in rows (rowu),
in columns and paired (colp), or in columns and unpaired (colu).}
}
\description{
This function determines reads in user's individual data for meta-analysis.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
