% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maplist_utils.R
\name{ReadListFile}
\alias{ReadListFile}
\title{Read gene list}
\usage{
ReadListFile(fileName)
}
\arguments{
\item{fileName}{file name of the data, .txt format}
}
\description{
read gene list, Fold Change (optional) and store the data in dataSet list object;
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
