% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{ReadMetaPathTable}
\alias{ReadMetaPathTable}
\title{ReadMetaPathTable}
\usage{
ReadMetaPathTable(mSetObj = NA, dataNM, dataFormat, dataType)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{dataNM}{file name with absolute path}

\item{dataFormat}{data format, can be colu or rowu}

\item{dataType}{data type, usually "massPeaks"}
}
\description{
ReadMetaPathTable
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca} 
Zhiqiang Pang\email{zhiqiang.pang@mail.mcgill.ca}
McGill University, Canada
License: MIT License
}
