% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{ReadMetaPathTableMix}
\alias{ReadMetaPathTableMix}
\title{ReadMetaPathTableMix}
\usage{
ReadMetaPathTableMix(mSetObj = NA, dataNM, dataNM2, dataFormat, dataType)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{dataNM}{file name 1 with absolute path (should be ESI+)}

\item{dataNM2}{file name 2 with absolute path (should be ESI-)}

\item{dataFormat}{data format, can be colu or rowu}

\item{dataType}{data type, usually "massPeaks"}
}
\description{
ReadMetaPathTableMix
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca} 
Zhiqiang Pang\email{zhiqiang.pang@mail.mcgill.ca}
McGill University, Canada
License: MIT License
}
