% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_express_meta.R
\name{ReadOmicsData}
\alias{ReadOmicsData}
\title{Read individual dataset for meta-analysis.}
\usage{
ReadOmicsData(fileName)
}
\arguments{
\item{fileName}{File name of data table.}
}
\description{
parse gene expression dataset into R object.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
