% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{ReadPairFile}
\alias{ReadPairFile}
\title{Read paired peak or spectra files}
\usage{
ReadPairFile(filePath = "pairs.txt")
}
\arguments{
\item{filePath}{Set file path}
}
\description{
This function reads paired peak lists or spectra files. The pair information
is stored in a file where each line is a pair and names are separated by ":".
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
