% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{ReadRawMeta}
\alias{ReadRawMeta}
\title{#Raw Spectra Upload}
\usage{
ReadRawMeta(fileName)
}
\description{
function used to process during uploading stage
}
\author{
Guangyan Zhou, Zhiqiang Pang
}
