% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readtable_utils.R
\name{ReadTabExpressData}
\alias{ReadTabExpressData}
\title{Read tab delimited file}
\usage{
ReadTabExpressData(fileName)
}
\arguments{
\item{fileName}{file name of the data, .txt format}
}
\description{
read tab delimited file, used in single gene expression;
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
