% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{RecordRCommand}
\alias{RecordRCommand}
\title{Record R Commands}
\usage{
RecordRCommand(mSetObj = NA, cmd)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{cmd}{Commands}
}
\description{
Record R Commands
}
