% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{RegisterData}
\alias{RegisterData}
\title{Register data in R}
\usage{
RegisterData(mSetObj = NA, dataSet)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{dataSet}{Input dataset to be registered in R.}
}
\description{
When there are multiple datasets, record their name and save the inputted data as
a .qs file to save memory. Note, the memory will only contain one mSetObj$dataSet object. By default the last one
will be the most recent/current dataSet object. Users can switch which data to load into memory.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
