% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{RemoveCmpd}
\alias{RemoveCmpd}
\title{Remove selected compounds}
\usage{
RemoveCmpd(mSetObj = NA, inx)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{inx}{Input the index of compound to remove}
}
\description{
Remove compounds
}
