% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{RemoveData}
\alias{RemoveData}
\title{Remove data object, the current dataSet will be the last one by default}
\usage{
RemoveData(dataName)
}
\description{
Remove data object, the current dataSet will be the last one by default
}
