% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{RemoveDuplicates}
\alias{RemoveDuplicates}
\title{Given a data with duplicates, remove duplicates}
\usage{
RemoveDuplicates(data, lvlOpt = "mean", quiet = T)
}
\description{
Dups is the one with duplicates
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
