% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming.R
\name{RemoveEmptyScan}
\alias{RemoveEmptyScan}
\title{Empty scan removal}
\usage{
RemoveEmptyScan(datapath)
}
\arguments{
\item{datapath}{Character, the path of the raw MS data files' folder/path (.mzXML, .mzData, .CDF and .mzML).
The empty scan free MS will be save as .mzXML format automaticaly.}
}
\description{
Function for 'Empty scan' removal
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}
License: GNU GPL (>= 2)
}
