% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{RemoveFile}
\alias{RemoveFile}
\title{Remove file}
\usage{
RemoveFile(fileName)
}
\arguments{
\item{fileName}{Input name of file to remove}
}
\description{
Remove file
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
