% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{RemoveFolder}
\alias{RemoveFolder}
\title{Remove folder}
\usage{
RemoveFolder(folderName)
}
\description{
Remove folder
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
