% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{RemoveGene}
\alias{RemoveGene}
\title{Remove selected genes}
\usage{
RemoveGene(mSetObj = NA, inx)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{inx}{Input compound index}
}
\description{
Remove selected genes based on an index
}
