% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{ReplaceMin}
\alias{ReplaceMin}
\title{Replace missing or zero values}
\usage{
ReplaceMin(mSetObj=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
This function will replace zero/missing values by half of the smallest
positive value in the original dataset.  
This method will be called after all missing value imputations are conducted.
Also, it directly modifies the mSet$dataSet$proc if executed after normalization,
or the mSet$dataSet$norm if before normalization.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
