% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{RerenderMetPAGraph}
\alias{RerenderMetPAGraph}
\title{Redraw current graph for zooming or clipping then return a value}
\usage{
RerenderMetPAGraph(mSetObj = NA, imgName, width, height)
}
\description{
Redraw current graph for zooming or clipping then return a value
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
