% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{ResetCustomCmpds}
\alias{ResetCustomCmpds}
\title{Reset custom compounds}
\usage{
ResetCustomCmpds(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Reset custom compounds
}
