% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{resultIncreased_doe}
\alias{resultIncreased_doe}
\title{Identify whether results improved or not}
\usage{
resultIncreased_doe(history)
}
\arguments{
\item{history}{List, an interal media objects used to save the optimization results of peaks.}
}
\description{
Identify whether results improved or not
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
