% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{RF.Anal}
\alias{RF.Anal}
\title{Perform Random Forest Analysis}
\usage{
RF.Anal(mSetObj = NA, treeNum = 500, tryNum = 7, randomOn = 1)
}
\description{
Perform Random Forest
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
