% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_covariate.R
\name{RF.AnalMeta}
\alias{RF.AnalMeta}
\title{Perform Random Forest Analysis}
\usage{
RF.AnalMeta(
  mSetObj = NA,
  treeNum = 500,
  tryNum = 7,
  randomOn = 1,
  selectedMeta
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{treeNum}{Input the number of trees to create, default is set to 500}

\item{tryNum}{Set number of tries, default is 7}

\item{randomOn}{Set random, default is 1}

\item{selectedMeta}{selected Meta elements}
}
\description{
Perform Random Forest
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
