% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{ROCPredSamplesTable}
\alias{ROCPredSamplesTable}
\title{Create a table of newly classified samples}
\usage{
ROCPredSamplesTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)
Function to create the table of newly classified samples}
}
\description{
Function to create the table of newly classified samples
}
