% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{RSVM.Anal}
\alias{RSVM.Anal}
\title{Recursive Support Vector Machine (R-SVM)}
\usage{
RSVM.Anal(mSetObj = NA, cvType)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{cvType}{Cross-validation type}
}
\description{
recursive SVM for feature selection and classification
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
