% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{RSVM}
\alias{RSVM}
\title{R-SVM core code}
\usage{
RSVM(x, y, ladder, CVtype, CVnum = 0)
}
\arguments{
\item{x}{Row matrix of data}

\item{y}{Class label: 1 / -1 for 2 classes}

\item{ladder}{Input the ladder}

\item{CVtype}{Integer (N fold CV), "LOO" leave-one-out CV, "bootstrape" bootstrape CV}

\item{CVnum}{Number of CVs, LOO: defined as sample size, Nfold and bootstrape:  user defined, default as sample size
outputs a named list
Error: a vector of CV error on each level
SelFreq: a matrix for the frequency of each gene being selected in each level
with each column corresponds to a level of selection
and each row for a gene
The top important gene in each level are those high-freqent ones}
}
\description{
Core code to perform R-SVM
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
