% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{RT.Adjust_Slave}
\alias{RT.Adjust_Slave}
\title{RT.Adjust_Slave}
\usage{
RT.Adjust_Slave(
  peaks,
  peakIndex,
  rtime,
  minFraction = 0.9,
  extraPeaks = 1,
  smooth = c("loess", "linear"),
  span = 0.2,
  family = c("gaussian", "symmetric"),
  peakGroupsMatrix = matrix(ncol = 0, nrow = 0),
  subsetAdjust = c("average", "previous")
)
}
\description{
RT.Adjust_Slave
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
