% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem_ms.R
\name{SaintyCheckMSPfile}
\alias{SaintyCheckMSPfile}
\title{SaintyCheckMSPfile}
\usage{
SaintyCheckMSPfile(
  mSetObj = NA,
  filename = "",
  format_type = "mzmine",
  keepAllspec = FALSE
)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{filename}{filename with path}

\item{format_type}{format type, can be 'mzmine' or 'msdial'}

\item{keepAllspec}{if you want to search all spectra from your local, turn keepAllspec to TRUE. it is FALSE by default.}
}
\description{
SaintyCheckMSPfile
}
\author{
Zhiqiang Pang
}
