% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{SanityCheckData}
\alias{SanityCheckData}
\title{Sanity Check Data}
\usage{
SanityCheckData(mSetObj=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
SanityCheckData is used for data processing, and performs a basic sanity 
check of the uploaded content, ensuring that the data is suitable for further analysis. 
The function will return a message if the data has successfully passed the check
and is deemed suitable for further analysis. 
The function will perform the check directly onto the mSet$dataSet object, and must 
be performed immediatly after reading in data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
