% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{SanityCheckIndData}
\alias{SanityCheckIndData}
\title{Sanity check of individual datasets for meta-analysis}
\usage{
SanityCheckIndData(mSetObj = NA, dataName)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{dataName}{Input name of the dataset to perform the sanity check.}
}
\description{
Performs a sanity check on each-uploaded dataset for meta-analysis. Briefly, this function
will exclude empty rows, check class labels, ensure only 2 groups are being compared within the dataset, 
ensure sample names are unique, remove low quality samples/features, and replace missing values.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
