% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_metadata.R
\name{SanityCheckMeta}
\alias{SanityCheckMeta}
\title{SanityCheckMeta#'}
\usage{
SanityCheckMeta(mSetObj = NA, init = 1)
}
\arguments{
\item{mSetObj}{metaboanalyst object}

\item{init}{can be 0 or 1}
}
\description{
SanityCheckMeta#'
}
