% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{SanityCheckMetaPathTable}
\alias{SanityCheckMetaPathTable}
\title{SanityCheckMetaPathTable}
\usage{
SanityCheckMetaPathTable(mSetObj = NA, dataName, dataName2)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{dataName}{file name 1 with absolute path}

\item{dataName2}{file name 2 with absolute path or "null"}
}
\description{
SanityCheckMetaPathTable
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca} 
Zhiqiang Pang\email{zhiqiang.pang@mail.mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
