% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{SanityCheckMummichogData}
\alias{SanityCheckMummichogData}
\title{Sanity Check Data}
\usage{
SanityCheckMummichogData(mSetObj=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects).}
}
\description{
SanityCheckData is used for data processing, and performs a basic sanity
check of the uploaded data, ensuring that the data is suitable for further analysis.
The function ensure that all parameters are properly set based on updated parameters.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
