% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{Save2KEGGJSON}
\alias{Save2KEGGJSON}
\title{Utility function for PerformKOEnrichAnalysis_List}
\usage{
Save2KEGGJSON(hits.query, res.mat, file.nm)
}
\description{
Utility function for PerformKOEnrichAnalysis_List
}
