% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_wrapper_utils.R
\name{SaveListHeatmapJSON}
\alias{SaveListHeatmapJSON}
\title{Generate JSON file for heatmap visualization of gene lists}
\usage{
SaveListHeatmapJSON(dataName = "", fileName)
}
\arguments{
\item{{object}}{dataSetObj Input the name of the created datasetObj (see Init.Data).}

\item{{string}}{fileName file name of the output JSON file}
}
\description{
Generate JSON file for heatmap visualization of gene lists
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
