% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_graphics.R
\name{saveNetworkInSIF}
\alias{saveNetworkInSIF}
\title{Network functions}
\usage{
saveNetworkInSIF(network, name)
}
\arguments{
\item{network}{Input network}

\item{name}{Input file name}
}
\description{
Adapted from BioNet
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
