% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{savePeakListMetaData}
\alias{savePeakListMetaData}
\title{Function to save each mSetObj as a RDS file
to be used later in PerformMetaPSEA.
Should be called after SetPeakEnrichMethod/SetMummichogPval}
\usage{
savePeakListMetaData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{mSetObj}
}
\description{
Function to save each mSetObj as a RDS file
to be used later in PerformMetaPSEA.
Should be called after SetPeakEnrichMethod/SetMummichogPval
}
