% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{SaveTransformedData}
\alias{SaveTransformedData}
\title{Save the processed data with class names}
\usage{
SaveTransformedData(mSetObj = NA)
}
\description{
This function saves the processed data with class names as CSV files. 
Several files may be generated, the original data, processed data, peak normalized, and/or normalized data.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
