% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SearchByCompound}
\alias{SearchByCompound}
\title{Search for compound from all member compounds of metabolite set}
\usage{
SearchByCompound(mSetObj = NA, query)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{query}{Input the query to search}
}
\description{
Search for compound from all member compounds of metabolite set
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
