% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{SearchCompoundLib}
\alias{SearchCompoundLib}
\title{Merges user-input files with library files}
\usage{
SearchCompoundLib(mSetObj=NA)
}
\arguments{
\item{mummichog}{Input the name of the created mummichog object (see mummichog_init).}
}
\description{
This function matches user-uploaded files to library files specific to the selected organism. It
also creates a list of compounds IDs from the user-uploaded list of significant m/z features.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
