% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SearchMsetLibraries}
\alias{SearchMsetLibraries}
\title{Search metabolite set libraries}
\usage{
SearchMsetLibraries(mSetObj = NA, query, type)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{query}{Input the query to search}

\item{type}{Input the data type (name or compound)}
}
\description{
Search metabolite set libraries
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
