% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{SelectMultiData}
\alias{SelectMultiData}
\title{Select one or more datasets for meta-analysis}
\usage{
SelectMultiData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
This function selects one or more datasets to be used for meta-analysis. 1 is used to indicate that 
a dataset is selected and by default, all datasets will be selected for meta-analysis.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
