% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{SetAnalysisMode}
\alias{SetAnalysisMode}
\title{Set biomarker analysis mode}
\usage{
SetAnalysisMode(mSetObj, mode)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)s}

\item{mode}{Input the selected mode for biomarker analysis, "univ" for univariate ROC curve analysis,
"explore" for multivariate ROC curve analysis, and "test" for ROC curve based model creation and evaluation.
McGill University, Canada
License: GNU GPL (>= 2)}
}
\description{
ROC utilities
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
}
