% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{SetBarParams}
\alias{SetBarParams}
\title{Plot a scatterplot bubble chart overview of the matched pathways}
\usage{
SetBarParams(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
x axis is the pathway impact factor
y axis is the p value (from ORA) 
return the circle information
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
