% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SetCachexiaSetUsed}
\alias{SetCachexiaSetUsed}
\title{Set the cachexia set used}
\usage{
SetCachexiaSetUsed(mSetObj = NA, used)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{used}{Set data to be used}
}
\description{
Set cachexia set used
}
