% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{SetCandidate}
\alias{SetCandidate}
\title{Set matched name based on user selection from all potential hits}
\usage{
SetCandidate(mSetObj = NA, query_nm, can_nm)
}
\description{
Note: to change object in the enclosing enviroment, use "<<-"
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
