% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{SetCurrentGroups}
\alias{SetCurrentGroups}
\title{To choose from two groups}
\usage{
SetCurrentGroups(mSetObj = NA, grps)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{grps}{Input the groups}
}
\description{
Choose two groups (when more than two groups uploaded)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
