% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SetCurrentMsetLib}
\alias{SetCurrentMsetLib}
\title{Set current user selected metset library for search}
\usage{
SetCurrentMsetLib(mSetObj=NA, libname, excludeNum)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{libname}{Input user selected name of library, "self", "kegg_pathway",
"smpdb_pathway", "blood", "urine", "csf", "snp", "predicted", "location", and "drug".}

\item{excludeNum}{Users input the mimimum number compounds within selected metabolite sets (metabolitesets < excludeNum)}
}
\description{
if enrichment analysis, also prepare lib by
creating a list of metabolite sets
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
