% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{SetDesignType}
\alias{SetDesignType}
\title{For two factor time series only}
\usage{
SetDesignType(mSetObj = NA, design)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{design}{Input the design type}
}
\description{
For two factor time series only
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
