% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R
\name{setInclusionDataSets}
\alias{setInclusionDataSets}
\title{setInclusionDataSets#'}
\usage{
setInclusionDataSets(mSetObj = NA, datasVec)
}
\arguments{
\item{mSetObj}{mSetObj}

\item{datasVec}{a vector of all files}
}
\description{
setInclusionDataSets#'
}
\examples{
#setInclusionDataSets(c("A1_pos.csv","B1_pos.csv","C1_pos.csv"));
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca} 
Zhiqiang Pang\email{zhiqiang.pang@mail.mcgill.ca}
McGill University, Canada
License: MIT License
}
