% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SetKEGG.PathLib}
\alias{SetKEGG.PathLib}
\title{Set KEGG pathway library}
\usage{
SetKEGG.PathLib(mSetObj = NA, kegg.rda)
}
\description{
note, this process can be long, need to return a value
to force Java to wait
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
