% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{SetMass.PathLib}
\alias{SetMass.PathLib}
\title{Load the selected metabolic pathway database}
\usage{
LoadMummichogLib(mSetObj=NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
LoadMummichogLib is used to load in the user selected metabolic pathway database. If the user's input
is just a list of significant features (anal.type = "siglist"), this function will also provide a list of
reference features.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
