% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SetMetabolomeFilter}
\alias{SetMetabolomeFilter}
\title{Set metabolome filter}
\usage{
SetMetabolomeFilter(mSetObj = NA, TorF)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{TorF}{Input metabolome filter}
}
\description{
Set metabolome filter
}
