% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{SetMetaPeaksPvals}
\alias{SetMetaPeaksPvals}
\title{For meta-analysis, set the p-value
cutoff used to define the new
input_mzlist}
\usage{
SetMetaPeaksPvals(mSetObj = NA)
}
\description{
For meta-analysis, set the p-value
cutoff used to define the new
input_mzlist
}
