% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem_ms.R
\name{setMS2DBOpt}
\alias{setMS2DBOpt}
\title{setMS2DBOpt}
\usage{
setMS2DBOpt(mSetObj = NA, DBoption = "regular")
}
\arguments{
\item{mSetObj}{mSetObj object}

\item{DBoption}{database option to define neutral loss or not, can be either 'regualr" or 'nl'.}
}
\description{
setMS2DBOpt
}
\author{
Zhiqiang Pang
}
