% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{SetMummichogPval}
\alias{SetMummichogPval}
\title{Set the cutoff for mummichog analysis}
\usage{
SetMummichogPval(mSetObj = NA, cutoff)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}
}
\description{
Set the p-value cutoff for mummichog analysis.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
