% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{SetMummichogPvalFromPercent}
\alias{SetMummichogPvalFromPercent}
\title{Set the cutoff for mummichog analysis}
\usage{
SetMummichogPvalFromPercent(mSetObj = NA, fraction)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}

\item{fraction}{fraction value for mummichog running}
}
\description{
Set the p-value cutoff for mummichog analysis.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
