% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{SetPeakFormat}
\alias{SetPeakFormat}
\title{Set the peak format for the mummichog analysis}
\usage{
SetPeakFormat(mSetObj = NA, type = "mpt")
}
\arguments{
\item{mSetObj}{mSetObj}

\item{type}{Input the name of mummichog analysis type, usually 'mpt'.}
}
\description{
Set the peak format for mummichog analysis.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
