% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{SetPeakParam}
\alias{SetPeakParam}
\title{SetPeakParam}
\usage{
SetPeakParam(
  platform = "general",
  Peak_method = "centWave",
  RT_method = "loess",
  mzdiff,
  snthresh,
  bw,
  ppm,
  min_peakwidth,
  max_peakwidth,
  noise,
  prefilter,
  value_of_prefilter,
  fwhm,
  steps,
  sigma,
  peakBinSize,
  max,
  criticalValue,
  consecMissedLimit,
  unions,
  checkBack,
  withWave,
  profStep,
  minFraction,
  minSamples,
  maxFeatures,
  mzCenterFun,
  integrate,
  extra,
  span,
  smooth,
  family,
  fitgauss,
  polarity,
  perc_fwhm,
  mz_abs_iso,
  max_charge,
  max_iso,
  corr_eic_th,
  mz_abs_add,
  adducts,
  rmConts
)
}
\arguments{
\item{platform}{platform}

\item{Peak_method}{Peak_method}

\item{RT_method}{RT_method}

\item{mzdiff}{mzdiff}

\item{snthresh}{snthresh}

\item{bw}{bw}

\item{ppm}{ppm}

\item{min_peakwidth}{min_peakwidth}

\item{max_peakwidth}{max_peakwidth}

\item{noise}{noise}

\item{prefilter}{prefilter}

\item{value_of_prefilter}{value_of_prefilter}

\item{fwhm}{fwhm}

\item{steps}{steps}

\item{sigma}{sigma}

\item{peakBinSize}{peakBinSize}

\item{max}{max}

\item{criticalValue}{criticalValue}

\item{consecMissedLimit}{consecMissedLimit}

\item{unions}{unions}

\item{checkBack}{checkBack}

\item{withWave}{withWave}

\item{profStep}{profStep}

\item{minFraction}{minFraction}

\item{minSamples}{minSamples}

\item{maxFeatures}{maxFeatures}

\item{mzCenterFun}{mzCenterFun}

\item{integrate}{integrate}

\item{extra}{extra}

\item{span}{span}

\item{smooth}{smooth}

\item{family}{family}

\item{fitgauss}{fitgauss}

\item{polarity}{polarity}

\item{perc_fwhm}{perc_fwhm}

\item{mz_abs_iso}{mz_abs_iso}

\item{max_charge}{max_charge}

\item{max_iso}{max_iso}

\item{corr_eic_th}{corr_eic_th}

\item{mz_abs_add}{mz_abs_add}

\item{adducts}{adducts}

\item{rmConts}{rmConts}
}
\description{
SetPeakParam, used to set the peak param
}
\author{
Zhiqiang Pang
}
