% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{SetPlotParam}
\alias{SetPlotParam}
\title{Set generic Plotting Parameters}
\usage{
SetPlotParam(Plot = F, labels = TRUE, format = "png", dpi = 72, width = 9, ...)
}
\arguments{
\item{Plot}{Logical, if true, the function will plot internal figures for different functions.}

\item{labels}{Logical, if true, the labels in the plot will be added.}

\item{format}{Numeric, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}

\item{...}{Other specific parameters for specific function. Please set them according to the corresponding function.}
}
\description{
This function sets the generic Plotting Parameters for different functions
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
