% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{setRendAttrs}
\alias{setRendAttrs}
\title{Generate suitable features for nodes and edges}
\usage{
setRendAttrs(g, AllBorder = "transparent", AllFixedsize = FALSE,
  AllFontsize = 16, AllShape = "rectangle", fillcolor = "lightgreen", ...)
}
\description{
adapted from PathRender, used in higher functions.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
