% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{SetRTincluded}
\alias{SetRTincluded}
\title{SetRTincluded}
\usage{
SetRTincluded(mSetObj = NA, rt = "no")
}
\arguments{
\item{mSetObj}{mSetObj}

\item{rt}{retention time types, "minutes", "seconds" or "no"}
}
\value{
mSetObj
}
\description{
SetRTincluded
}
