% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{SetSAMSigMat}
\alias{SetSAMSigMat}
\title{Set Signifiance Analysis of Microarrays (SAM) analysis matrix}
\usage{
SetSAMSigMat(mSetObj = NA, delta)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Create SAM matrix
}
