% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SetSMPDB.PathLib}
\alias{SetSMPDB.PathLib}
\title{Set SMPDB pathway library}
\usage{
SetSMPDB.PathLib(mSetObj = NA, libNm)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{libNm}{Input library name}
}
\description{
note, this process can be long, need to return a value
to force Java to wait
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
