% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{Setup.AdductData}
\alias{Setup.AdductData}
\title{Save adduct names for mapping}
\usage{
Setup.AdductData(mSetObj = NA, qvec)

Setup.AdductData(mSetObj = NA, qvec)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{qvec}{Input the vector to query}
}
\description{
Save adduct names for mapping

Save adduct names for mapping
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
