% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{Setup.BiofluidType}
\alias{Setup.BiofluidType}
\title{Save biofluid type for SSP}
\usage{
Setup.BiofluidType(mSetObj = NA, type)
}
\description{
Save biofluid type for SSP
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
