% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{Setup.HMDBReferenceMetabolome}
\alias{Setup.HMDBReferenceMetabolome}
\title{Read user uploaded metabolome as a list of HMDB compound names}
\usage{
Setup.HMDBReferenceMetabolome(mSetObj = NA, filePath)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{filePath}{Input the path to the user's list of HMDB compound names}
}
\description{
Read user uploaded metabolome as a list of HMDB compound names
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
