% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{Setup.KEGGReferenceMetabolome}
\alias{Setup.KEGGReferenceMetabolome}
\title{Read user uploaded metabolome as a list of KEGG pathway ids}
\usage{
Setup.KEGGReferenceMetabolome(mSetObj = NA, filePath)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{filePath}{Input the path to the user's list of KEGG pathway ids}
}
\description{
Read user uploaded metabolome as a list of KEGG pathway ids
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
