% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{Setup.MapData}
\alias{Setup.MapData}
\title{Save compound name for mapping}
\usage{
Setup.MapData(mSetObj = NA, qvec)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{qvec}{Input the vector to query}
}
\description{
Save compound name for mapping
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
