% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{Setup.UserMsetLibData}
\alias{Setup.UserMsetLibData}
\title{Read user upload metabolite set library file}
\usage{
Setup.UserMsetLibData(mSetObj = NA, filePath)
}
\description{
Return two col csv file, first name, second cmpd list
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
