% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{SetupKEGGLinks}
\alias{SetupKEGGLinks}
\title{Only works for human (hsa.rda) data}
\usage{
SetupKEGGLinks(smpdb.ids)
}
\arguments{
\item{smpdb.ids}{Input the list of SMPD ids to add SMPDB links}
}
\description{
Only works for human (hsa.rda) data
2018 - works for ath, eco, mmu, sce
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
