% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{SetupMSdataMatrix}
\alias{SetupMSdataMatrix}
\title{Create a MS spectra data matrix of peak values for each group}
\usage{
SetupMSdataMatrix(mSetObj, intvalue)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{intvalue}{name of peak column to enter into the returned matrix, if intvalue = 'into', 
use integrated area of original (raw) peak intensities, 
if intvalue = 'intf', use integrated area of filtered peak intensities, 
if intvalue = 'intb', use baseline corrected integrated peak intensities,
if intvalue = 'maxo', use the maximum intensity of original (raw) peaks, 
or if intvalue = 'maxf' use the maximum intensity of filtered peaks}
}
\description{
This function sets up a MS spectra data matrix using the 'groupval' function
from XCMS. This generates a usable matrix of peak values for analysis where 
columns represent peak groups and rows represent samples. 
Collisions where more than one peak from a single sample are in the same group get resolved
utilizing "medret", which uses the peak closest to the median retention time.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
