% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{SetupSMPDBLinks}
\alias{SetupSMPDBLinks}
\title{Only works for human (hsa.rda) data}
\usage{
SetupSMPDBLinks(kegg.ids)
}
\arguments{
\item{kegg.ids}{Input the list of KEGG ids to add SMPDB links}
}
\description{
Only works for human (hsa.rda) data
2018 - works for ath, eco, mmu, sce
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
