% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{SetUserPath}
\alias{SetUserPath}
\title{Set User Path}
\usage{
SetUserPath(path)
}
\description{
play roles at the first uploading and login stage
}
\author{
Guangyan Zhou, Zhiqiang Pang
TODO: can be delete in the future
}
