% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{SlaveCluster_doe}
\alias{SlaveCluster_doe}
\title{Core Peak Picking Slave Cluster}
\usage{
SlaveCluster_doe(
  task,
  Set_parameters,
  object,
  object_mslevel,
  isotopeIdentification,
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{task}{Numeric, task order for XCMS paramters table to run the peak picking and alignment.}

\item{object_mslevel}{List, the parsed metabolomics scans produced by PeakPicking_prep.}

\item{isotopeIdentification}{Character, IsotopeIdentidication method, usually includes 'IPO' and 'CAMERA'.}

\item{BPPARAM}{MulticoreParam method, used to set the parallel method. Default is bpparam().}

\item{xcmsSet_parameters}{Matrix, the parameters combination produced automatically according to 
the primary parameters input.}

\item{MSnExp}{object, the trimmed or the original data.}
}
\description{
Core Peak Picking Slave Cluster
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
