% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{sparse.mint.block_iteration}
\alias{sparse.mint.block_iteration}
\title{Perform Sparse Generalized Canonical Correlation (sgccak)}
\usage{
sparse.mint.block_iteration(A, design, ncomp)
}
\arguments{
\item{A}{list of datasets each with the same number of rows (samples)}

\item{Design}{design matrix}

\item{ncomp}{vector specifying number of components to keep per datasets}
}
\description{
Runs sgccak() modified from RGCCA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
