% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{spectraInclusion}
\alias{spectraInclusion}
\title{spectraInclusion}
\usage{
spectraInclusion(files, number)
}
\description{
save the spectra file info for project loading
}
\author{
Zhiqiang Pang
}
