% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{splsda}
\alias{splsda}
\title{Perform sPLS-DA}
\usage{
splsda(X, Y, ncomp = 2, mode = c("regression", "canonical",
  "invariant", "classic"), keepX, keepX.constraint = NULL,
  scale = TRUE, tol = 1e-06, max.iter = 100, near.zero.var = FALSE,
  logratio = "none", multilevel = NULL)
}
\arguments{
\item{X}{numeric matrix of predictors}

\item{Y}{a factor or a class vector for the discrete outcome}

\item{ncomp}{the number of components to include in the model. Default to 2.}

\item{mode}{Default set to c("regression", "canonical", "invariant", "classic")}

\item{keepX}{Number of \eqn{X} variables kept in the model on the last components (once all keepX.constraint[[i]] are used).}

\item{keepX.constraint}{A list containing which variables of X are to be kept on each of the first PLS-components.}

\item{scale}{Boleean. If scale = TRUE, each block is standardized to zero means and unit variances (default: TRUE).}

\item{tol}{Convergence stopping value.}

\item{max.iter}{integer, the maximum number of iterations.}

\item{near.zero.var}{boolean, see the internal \code{\link{nearZeroVar}} function (should be set to TRUE in particular for data with many zero values). 
Setting this argument to FALSE (when appropriate) will speed up the computations}

\item{logratio}{"None" by default, or "CLR"}

\item{multilevel}{Designate multilevel design, "NULL" by default}
}
\description{
Sparse PLS functions (adapted from mixOmics package for web-based usage)
this function is a particular setting of internal_mint.block
the formatting of the input is checked in internal_wrapper.mint
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
