% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{SPLSR.Anal}
\alias{SPLSR.Anal}
\title{R script for MetaboAnalyst
Description: perform PCA/PLS-DA/Orthogonal PLS-DA/sparse PLS-DA}
\usage{
SPLSR.Anal(mSetObj = NA, comp.num, var.num, compVarOpt)
}
\description{
Sparse PLS-DA (from mixOmics)
}
\details{
Author: Jeff Xia, jeff.xia@mcgill.ca
McGill University, Canada

License: GNU GPL (>= 2)
Perform SPLS-DA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
