% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming.R
\name{ssm_trim}
\alias{ssm_trim}
\title{Standards Simulation Method}
\usage{
ssm_trim(raw_data, ms_list, rt.idx)
}
\arguments{
\item{raw_data}{MSnExp object, the raw data that has been read in memory.}

\item{ms_list}{List, the names list of all scans.}

\item{rt.idx}{Numeric, the retention time percentage, from 0 to 1. Default is 1/15.}
}
\description{
Whole mass spectra will be divided as 4 bins according to the mz range. Trimming 
the raw with slide window method in every bins and retained the windows with highest scan intensity
and remove other scan signal in mz dimension. Then the data will be trimed again in the RT dimension
with slide window method. The window with highest intensity scans will be kept. After the timming
alongside mz and RT dimension, the peaks not only the high intensity peaks, but also the relatively 
low intensity peaks will also be retained as the 'simulated standards' data for parameters optimization.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
