% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{Statistic_doe}
\alias{Statistic_doe}
\title{Analyze DoE Result}
\usage{
Statistic_doe(
  object,
  object_mslevel,
  isotopeIdentification,
  BPPARAM = bpparam(),
  mSet_OPT,
  subdir = NULL,
  plot = F,
  iterator,
  index.set,
  useNoise
)
}
\arguments{
\item{object}{MSnExp object, the trimmed or the original data.}

\item{object_mslevel}{List, the parsed metabolomics scans produced by PeakPicking_prep.}

\item{isotopeIdentification}{Character, IsotopeIdentidication method, usually includes 'IPO' and 'CAMERA'.}

\item{BPPARAM}{MulticoreParam method, used to set the parallel method. Default is bpparam().}

\item{mSet_OPT}{List, the result produced by 'ExperimentsCluster'.}

\item{subdir}{Logical, weather to creat a sub-directory (if true) or not (if false).}

\item{plot}{Logical, weather to plot the Contours plots of the DoE results.}

\item{iterator}{Numeric, the round number of the DoE.}

\item{index.set}{List, the indexes set (including PPS, CV, RCS, GS and Gaussian Index) produced by 
ExperiemntCluster.}

\item{useNoise}{Numeric, the noise level removed to evalute the gaussian peak.}
}
\description{
Analyze DoE Result
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
