% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{template.match}
\alias{template.match}
\title{Pattern hunter}
\usage{
template.match(x, template, dist.name)
}
\arguments{
\item{x}{Input data}

\item{template}{Input template}

\item{dist.name}{Input distance method}
}
\description{
Run template on all the high region effect genes
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
