% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{Ttests.Anal}
\alias{Ttests.Anal}
\title{Perform t-test analysis}
\usage{
Ttests.Anal(
  mSetObj = NA,
  nonpar = F,
  threshp = 0.05,
  paired = FALSE,
  equal.var = TRUE,
  pvalType = "fdr",
  all_results = FALSE
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{nonpar}{Logical, use a non-parametric test, T or F. False is default.}

\item{threshp}{Numeric, enter the adjusted p-value (FDR) cutoff}

\item{paired}{Logical, is data paired (T) or not (F).}

\item{equal.var}{Logical, evaluates if the group variance is equal (T) or not (F).}

\item{all_results}{Logical, if TRUE, returns T-Test analysis results
for all compounds.}
}
\description{
This function is used to perform t-test analysis.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
