% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{UnzipUploadedFile}
\alias{UnzipUploadedFile}
\title{Unzip .zip files}
\usage{
UnzipUploadedFile(mSetObj, inPath, outPath, rmFile = T)
}
\description{
Unzips uploaded .zip files, removes the uploaded file, checks for success
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
