% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{UpdateData}
\alias{UpdateData}
\title{Update data for filtering}
\usage{
UpdateData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Function to update the mSetObj after removing features or samples.
}
