% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{UpdateFeatureItems}
\alias{UpdateFeatureItems}
\title{Remove feature items}
\usage{
UpdateFeatureItems(mSetObj=NA, feature.nm.vec)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{feature.nm.vec}{Input the name of the feature to remove from the data in quotation marks. 
The name must be identical to the feature names found in the data set.}
}
\description{
This function removes user-selected features from the data set. 
This must be performed following data processing and filtering.
If the data was normalized prior to removal, you must re-normalize the data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}, Jasmine Chong 
McGill University, Canada
}
