% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{UpdateGraphSettings}
\alias{UpdateGraphSettings}
\title{Update graph settings}
\usage{
UpdateGraphSettings(mSetObj = NA, colVec, shapeVec)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Function to update the graph settings.
}
