% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{UpdateOPLS.Splot}
\alias{UpdateOPLS.Splot}
\title{Update OPLS loadings}
\usage{
UpdateOPLS.Splot(mSetObj = NA, plotType)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{plotType}{Set annotation type, "all" to label all variables and
"none" to label no variables.}
}
\description{
Update the OPLS loadings
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
