% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{UpdatePCA.Loading}
\alias{UpdatePCA.Loading}
\title{Update PCA loadings}
\usage{
UpdatePCA.Loading(mSetObj = NA, plotType)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{plotType}{Set annotation type, "all" to label all variables and
"none" to label no variables.}
}
\description{
Update the PCA loadings
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
