% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{UpdatePLS.Loading}
\alias{UpdatePLS.Loading}
\title{Update PLS loadings}
\usage{
UpdatePLS.Loading(mSetObj = NA, plotType)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{plotType}{Set annotation type, "all" to label all variables and
"none" to label no variables.}
}
\description{
Update the PLS loadings
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
