% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{updateRawSpectraParam}
\alias{updateRawSpectraParam}
\title{updateRawSpectraParam}
\usage{
updateRawSpectraParam(Params)
}
\arguments{
\item{Params}{object generated by SetPeakParams function.}
}
\description{
updateRawSpectraParam
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
}
