% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_meta.R
\name{UpdateSampleInfo}
\alias{UpdateSampleInfo}
\title{Updating meta-data info (meta-analysis)}
\usage{
UpdateSampleInfo(dataName, clsLbl)
}
\arguments{
\item{dataName}{File name of data set.}

\item{clsLbl}{vector of class labels to be assigned to samples}
}
\description{
Assign class labels to samples (meta-analysis)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
