% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{UpdateSampleItems}
\alias{UpdateSampleItems}
\title{Remove samples from user's data}
\usage{
UpdateSampleItems(mSetObj=NA, smpl.nm.vec)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{smpl.nm.vec}{Input the name of the sample to remove from the data in quotation marks. The name must be identical to the 
sample names found in the data set.}
}
\description{
This function removes samples from the data set. This must be performed following data processing and filtering.
If the data was normalized prior to removal, you must re-normalize the data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}, Jasmine Chong 
McGill University, Canada
}
