% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{updateSpectra3DPCA}
\alias{updateSpectra3DPCA}
\title{updateSpectra3DPCA}
\usage{
updateSpectra3DPCA(featureNM = 100)
}
\description{
update Spectra 3D PCA, mainly json file
}
\author{
Zhiqiang Pang
}
