% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{usr2png}
\alias{usr2png}
\title{Perform utilities for MetPa}
\usage{
usr2png(xy, im)
}
\arguments{
\item{xy}{Input coordinates}

\item{im}{Input coordinates}
}
\description{
Convert user coords (as used in current plot) to pixels in a png
adapted from the imagemap package
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
