% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{verifyParam}
\alias{verifyParam}
\title{verifyParam}
\usage{
verifyParam(param0_path, users.path)
}
\description{
Verify the example params has changed or not
}
\author{
Guangyan Zhou, Zhiqiang Pang
}
