% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_utils.R
\name{Volcano.Anal}
\alias{Volcano.Anal}
\title{Prepare data for volcano plot visualization}
\usage{
Volcano.Anal(
  dataName = "",
  fileNm = "name",
  paired = FALSE,
  fcthresh = 0,
  threshp = 0.05,
  analType = "NA",
  inx = 1
)
}
\description{
Prepare data for volcano plot visualization
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
