% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{Wrap.Names}
\alias{Wrap.Names}
\title{Utilities for creating pathway maps for MetPA}
\usage{
Wrap.Names(cName, wrap.len = 10, tol.len = 5)
}
\description{
Perform various utilities for MetPA
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

Function to deal with long string names
}
