% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{XSet2MSet}
\alias{XSet2MSet}
\title{Converts xset object from XCMS to mSet object for MetaboAnalyst}
\usage{
mSet <- XSet2MSet(xset, dataType, analType, paired=F, format, lbl.type)
}
\arguments{
\item{xset}{The name of the xcmsSet object created.}

\item{dataType}{The type of data, either list (Compound lists), conc (Compound concentration data), 
specbin (Binned spectra data), pktable (Peak intensity table), nmrpeak (NMR peak lists), mspeak (MS peak lists), 
or msspec (MS spectra data).}

\item{analType}{Indicate the analysis module to be performed: stat, pathora, pathqea, msetora, msetssp, msetqea, ts, 
cmpdmap, smpmap, or pathinteg.}

\item{paired}{Logical, is data paired (T) or not (F).}

\item{format}{Specify if samples are paired and in rows (rowp), unpaired and in rows (rowu),
in columns and paired (colp), or in columns and unpaired (colu).}

\item{lbl.type}{Specify the data label type, either discrete (disc) or continuous (cont).}
}
\description{
This function converts processed raw LC/MS data from XCMS 
to a usable data object (mSet) for MetaboAnalyst. The immediate next step following using this 
function is to perform a SanityCheck, and then further data processing and analysis can continue.
}
