% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{ANOVA.Anal}
\alias{ANOVA.Anal}
\title{Perform ANOVA analysis}
\usage{
ANOVA.Anal(mSetObj=NA, nonpar=FALSE, thresh=0.05, all_results=FALSE)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{nonpar}{Logical, use a non-parametric test (T) or not (F)}

\item{thresh}{Numeric, from 0 to 1, indicate the p-value threshold}

\item{all_results}{Logical, if TRUE, it will output the ANOVA results for all compounds}
}
\description{
ANOVA analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
