% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{CalculateOraScore}
\alias{CalculateOraScore}
\title{Calculate ORA score}
\usage{
CalculateOraScore(mSetObj=NA, nodeImp, method)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{nodeImp}{Indicate the pathway topology analysis, "rbc" for relative-betweeness centrality, 
and "dgr" for out-degree centrality.}

\item{method}{is "fisher" or "hyperg"}
}
\description{
Calculate the over representation analysis score
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
