% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{CalculatePairwiseDiff}
\alias{CalculatePairwiseDiff}
\title{Calculate Pairwise Differences}
\usage{
CalculatePairwiseDiff(mat)
}
\arguments{
\item{mat}{Input matrix of data to calculate pair-wise differences.}
}
\description{
Mat are log normalized, diff will be ratio. Used in higher functions.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
