% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{CentroidMSData}
\alias{CentroidMSData}
\title{Centroid MS data}
\usage{
CentroidMSData(in_path, ncore = 1, rm = F)
}
\arguments{
\item{in_path}{Character, the path of your MS data folder, only 
one format should be included in the folder. If you have different kinds of formats, 
please separate them into different folder and do this centroid multiple times.}

\item{ncore}{Numeric, how many cores you want to use for data centroid, default is 1.}

\item{rm}{logic, to remove the original data (TRUE) or not (FALSE, default).}
}
\description{
This function is used to centorid different MS data (specially designed for .CDF/.cdf, other formats are also supported). 
ProteoWizard (PMID: 27186799) based on docker is optional but strongly suggested to be installed. Current user of the docker also 
need to be added into thte sudo list to make it configured well to this function.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca},
}
