% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{centroidMSData}
\alias{centroidMSData}
\title{centroidMSData}
\usage{
centroidMSData(fileName, outFolder, ncore = 1)
}
\description{
centroidMSData is used to centroid MS Data
}
\author{
Zhiqiang Pang
}
