% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{CleanData}
\alias{CleanData}
\title{Perform data cleaning}
\usage{
CleanData(bdata, removeNA = T, removeNeg = T, removeConst = T)
}
\arguments{
\item{bdata}{Input data to clean}

\item{removeNA}{Logical, T to remove NAs, F to not.}

\item{removeNeg}{Logical, T to remove negative numbers, F to not.}

\item{removeConst}{Logical, T to remove samples/features with 0s, F to not.}
}
\description{
Cleans data and removes -Inf, Inf, NA, negative and 0s.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
