% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{ClearNegatives}
\alias{ClearNegatives}
\title{Data processing: Dealing with negative values}
\usage{
ClearNegatives(mSetObj = NA, method = "abs")
}
\description{
Operates on dataSet$procr after dealing with missing values
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
