% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{ComputeAverageCurve}
\alias{ComputeAverageCurve}
\title{Compute average ROC curve}
\usage{
ComputeAverageCurve(perf, avg.method)
}
\arguments{
\item{perf}{Input the average}

\item{avg.method}{Input the name of the method to compute the average curve}
}
\description{
Compute the average ROC curve
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
