% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{ComputeHighLow}
\alias{ComputeHighLow}
\title{Compute the 95 percent interval for threshold ROC}
\usage{
ComputeHighLow(perf)
}
\arguments{
\item{perf}{Input the performance}
}
\description{
Computes the 95 percent interval only for the y-axis.
Utility function, called upon by higher functions
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
