% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_covariate.R
\name{CovariateScatter.Anal}
\alias{CovariateScatter.Anal}
\title{CovariateScatter.Anal}
\usage{
CovariateScatter.Anal(
  mSetObj,
  imgName = "NA",
  format = "png",
  dpi = 72,
  theme = "default",
  analysis.var,
  ref = NULL,
  block = "NA",
  thresh = 0.05,
  all_results = FALSE
)
}
\arguments{
\item{mSetObj}{mSetObj object}

\item{imgName}{image name}

\item{format}{image format}

\item{dpi}{image dpi}

\item{theme}{image theme}

\item{analysis.var}{variable of analysis}

\item{ref}{reference group}

\item{block}{block name}

\item{thresh}{threshold}

\item{all_results}{all results}
}
\description{
CovariateScatter.Anal
}
