% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateBiomarkerRatioOverview}
\alias{CreateBiomarkerRatioOverview}
\title{Create biomarker analysis report: Normalization, ratio}
\usage{
CreateBiomarkerRatioOverview(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Biomarker analysis, ratio option
}
\author{
Jasmine Chong
McGill University, Canada
License: MIT License
}
