% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateBiomarkerRnwReport}
\alias{CreateBiomarkerRnwReport}
\title{Create report of analyses (Biomarker)}
\usage{
CreateBiomarkerRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Puts together the analysis report
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
