% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_raw_spectra.R,
%   R/sweave_report_stats.R
\name{CreateCorrDoc}
\alias{CreateCorrDoc}
\title{Create report of analyses}
\usage{
CreateCorrDoc(mSetObj = NA)

CreateCorrDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Create correlation document

Report generation using Sweave
Create correlation document
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
