% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateEnrichORAdoc}
\alias{CreateEnrichORAdoc}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateEnrichORAdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report, over
representation analysis (ORA)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
