% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_enrichment.R
\name{CreateEnrichRnwReport}
\alias{CreateEnrichRnwReport}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateEnrichRnwReport(mSetObj, usrName)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{usrName}{Input the name of the user}
}
\description{
Report generation using Sweave
Metabolite enrichment analysis report
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
