% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_time.R
\name{CreateiPCAdoc}
\alias{CreateiPCAdoc}
\title{Create report of analyses}
\usage{
CreateiPCAdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Interactive PCA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
