% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateMetaAnalysisDEdoc}
\alias{CreateMetaAnalysisDEdoc}
\title{Create MetaAnalysis analysis report: Data Normalization}
\usage{
CreateMetaAnalysisDEdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Meta-Analysis, data normalization documentation.
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
