% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_meta_analysis.R
\name{CreateMetaTable}
\alias{CreateMetaTable}
\title{Create MetaAnalysis table of results}
\usage{
CreateMetaTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Function to create a table containing meta-analysis results.
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
