% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_metamummi.R,
%   R/sweave_report_mummichog.R
\name{CreateMummichogAnalysisDoc}
\alias{CreateMummichogAnalysisDoc}
\title{Create mummichog analysis report}
\usage{
CreateMummichogAnalysisDoc(mSetObj = NA)

CreateMummichogAnalysisDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Mummichog analysis report

Report generation using Sweave
Mummichog analysis report
}
\author{
Jasmine Chong 
McGill University, Canada
License: GNU GPL (>= 2)

Jasmine Chong 
McGill University, Canada
License: GNU GPL (>= 2)
}
