% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_mummichog.R
\name{CreateMummichogOverview}
\alias{CreateMummichogOverview}
\title{Create Mummichog analysis report: Overview}
\usage{
CreateMummichogOverview()
}
\description{
Report generation using Sweave
Mummichog analysis report overview
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
