% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_network.R
\name{CreateNetworkExplorerRnwReport}
\alias{CreateNetworkExplorerRnwReport}
\title{Create report of analyses (Network Explorer)}
\usage{
CreateNetworkExplorerRnwReport(mSetObj, usrName)
}
\description{
Report generation using Sweave
Puts together the analysis report
}
\author{
Jasmine Chong
McGill University, Canada
License: GNU GPL (>= 2)
}
