% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateOPLSDAdoc}
\alias{CreateOPLSDAdoc}
\title{Create report of analyses}
\usage{
CreateOPLSDAdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Create OPLSDA document
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
