% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_pathway.R
\name{CreatePathIntr}
\alias{CreatePathIntr}
\title{Create report of analyses (Met Pathway)}
\usage{
CreatePathIntr()
}
\description{
Report generation using Sweave
Metabolomic pathway analysis
Introduction
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
