% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_stats.R
\name{CreatePCAdoc}
\alias{CreatePCAdoc}
\title{Create report of analyses}
\usage{
CreatePCAdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Create PCA document
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
