% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateRFdoc}
\alias{CreateRFdoc}
\title{Create report of analyses}
\usage{
CreateRFdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Create Random Forest document
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
