% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{CreateSemiTransColors}
\alias{CreateSemiTransColors}
\title{Create semitransparant colors}
\usage{
CreateSemiTransColors(cls)
}
\description{
Create semitransparant colors for a given class label
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
