% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_stats.R
\name{CreateStatRnwReport}
\alias{CreateStatRnwReport}
\title{Create report for statistical analysis module}
\usage{
CreateStatRnwReport(mSetObj, usrName)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{usrName}{Input the name of the user}
}
\description{
Report generation using Sweave
Write .Rnw file template
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
