% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateSummaryTable}
\alias{CreateSummaryTable}
\title{Create report of analyses}
\usage{
CreateSummaryTable(mSetObj = NA)
}
\description{
Report generation using Sweave
Create a summary table for each type of uploaded data
csv table has 5 col: sampleID, feature #, zero,  missing #
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
