% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_reporter.R
\name{CreateSVMdoc}
\alias{CreateSVMdoc}
\title{Create report of analyses}
\usage{
CreateSVMdoc(mSetObj = NA)
}
\description{
Report generation using Sweave
Create R-SVM document
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
