% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_biomarker.R
\name{CreateUnivarBiomarkersDoc}
\alias{CreateUnivarBiomarkersDoc}
\title{Create power analysis report: Biomarker Univariate Analysis}
\usage{
CreateUnivarBiomarkersDoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Biomarker analysis report, Univariate Analysis
}
\author{
Jasmine Chong 
McGill University, Canada
License: MIT License
}
