% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweave_report_stats.R
\name{CreateUNIVdoc}
\alias{CreateUNIVdoc}
\title{Create report of analyses}
\usage{
CreateUNIVdoc(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Report generation using Sweave
Create univariate analyses document
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
