% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{CrossReferencing}
\alias{CrossReferencing}
\title{Various functions for mapping b/w names & database identifiers
Given a list of compound names or ids, find matched name or ids from selected databases}
\usage{
CrossReferencing(mSetObj = NA, q.type, hmdb = T, pubchem = T,
  chebi = F, kegg = T, metlin = F)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects).}

\item{q.type}{Input the query type, "name" for compound names, "hmdb" for HMDB IDs, "kegg" for KEGG IDs, "pubchem"
for PubChem CIDs, "chebi" for ChEBI IDs, "metlin" for METLIN IDs, and "hmdb_kegg" for a both KEGG and HMDB IDs.}

\item{hmdb}{Logical, T to cross reference to HMDB, F to not.}

\item{pubchem}{Logical, T to cross reference to PubChem, F to not.}

\item{chebi}{Logical, T to cross reference to CheBI, F to not.}

\item{kegg}{Logical, T to cross reference to KEGG, F to not.}

\item{metlin}{Logical, T to cross reference to MetLin, F to not.}
}
\description{
Given a list of compound names or ids
find matched name or IDs from selected databases
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
