% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_data.R
\name{doEmblProtein2EntrezMapping}
\alias{doEmblProtein2EntrezMapping}
\title{Utility function for PerformNetEnrichment}
\usage{
doEmblProtein2EntrezMapping(emblprotein.vec)
}
\arguments{
\item{emblprotein.vec}{Input the vector containing protein embl ids}
}
\description{
Utility function for PerformNetEnrichment
}
