% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{.compute.mummichog.fgsea}
\alias{.compute.mummichog.fgsea}
\title{Internal function for calculating GSEA, no RT}
\usage{
.compute.mummichog.fgsea(mSetObj, permNum)
}
\description{
Internal function for calculating GSEA, no RT
}
