% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{.readDataTable}
\alias{.readDataTable}
\title{Read data table}
\usage{
.readDataTable(fileName)
}
\arguments{
\item{fileName}{Input filename}
}
\description{
note, try to use the fread, however, it has issues with 
some windows 10 files "Line ending is \r\r\n. .... appears to add the extra \r in text mode on Windows"
in such as, use the slower read.table method
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
