% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{.readDataTable}
\alias{.readDataTable}
\title{Read data table}
\usage{
.readDataTable(fileName)
}
\arguments{
\item{fileName}{Input filename}
}
\description{
Function to read in a data table. First, it will try to use fread, however, it has issues with 
some windows 10 files. In such case, use the slower read.table method.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
