% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_correlations.R
\name{FeatureCorrelation}
\alias{FeatureCorrelation}
\title{Pattern hunter}
\usage{
FeatureCorrelation(mSetObj = NA, dist.name, varName)
}
\description{
calculate correlation of all other feature to a given feature name
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
