% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_util_peaks.R, R/stats_peaks.R
\name{findEqualGreaterM}
\alias{findEqualGreaterM}
\title{Perform utilities for peak grouping}
\usage{
findEqualGreaterM(x, values)

findEqualGreaterM(x, values)
}
\arguments{
\item{x}{Input the data}

\item{values}{Input the values}
}
\description{
Perform various utilities for peak grouping

Perform various utilities for peak grouping
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
