% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{FisherLSD}
\alias{FisherLSD}
\title{Fisher for ANOVA}
\usage{
FisherLSD(aov.obj, thresh)
}
\description{
Perform  Fisher LSD for ANOVA, used in higher function
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
