% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Get.pAUC}
\alias{Get.pAUC}
\title{Calculate partial area under ROC curve}
\usage{
Get.pAUC(x, y, focus, cutoff)
}
\description{
Calculate partial area under ROC curve
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
