% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Get.pred}
\alias{Get.pred}
\title{Get predicted class probability}
\usage{
Get.pred(x.train, y.train, x.test, y.test, clsMethod = "pls")
}
\arguments{
\item{x.train}{Training X}

\item{y.train}{Training Y}

\item{x.test}{Test X}

\item{y.test}{Test Y}

\item{clsMethod}{Method to predict class, by default it is PLS}
}
\description{
Get predicted class probability, used in higher function
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
