% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{GetCandidateList}
\alias{GetCandidateList}
\title{Get all candidate compound names for a given index}
\usage{
GetCandidateList(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Returns 3 coloumns - inx, name, score
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
