% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetCIs}
\alias{GetCIs}
\title{Get confidence intervals}
\usage{
GetCIs(data, param = F)
}
\description{
For non-parametric tests, use quantiles,
use normal (1.96*std.err) if parametric
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
