% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{GetExtendRange}
\alias{GetExtendRange}
\title{Extend axis}
\usage{
GetExtendRange(vec, unit = 10)
}
\arguments{
\item{vec}{Input the vector}

\item{unit}{Numeric}
}
\description{
Extends the axis range to both ends
vec is the values for that axis
unit is the width to extend, 10 will increase by 1/10 of the range
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
