% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetFC}
\alias{GetFC}
\title{Used by higher functions to calculate fold change}
\usage{
GetFC(mSetObj = NA, paired = FALSE, cmpType, verbose = TRUE, top_k = 5)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{paired}{Logical, true of false}

\item{cmpType}{Numeric, 0 or 1}
}
\description{
Utility method to calculate FC, used in higher function
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
