% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetFeatureNumbers}
\alias{GetFeatureNumbers}
\title{Numbers for subset selection}
\usage{
GetFeatureNumbers(feat.len)
}
\arguments{
\item{feat.len}{Input the feature length}
}
\description{
Return a series of number for subsets selection
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
