% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{GetFinalNameMap}
\alias{GetFinalNameMap}
\title{Return the final (after user selection) map as dataframe}
\usage{
GetFinalNameMap(mSetObj = NA)
}
\description{
Returns three columns: original name, HMDB name and KEGG ID,
for enrichment and pathway analysis, respectively
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
