% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_stats.R
\name{GetHTMLMetSet}
\alias{GetHTMLMetSet}
\title{Given a metset inx, return hmtl highlighted metset cmpds and references}
\usage{
GetHTMLMetSet(mSetObj = NA, msetNm)
}
\description{
Given a metset inx, return hmtl highlighted metset cmpds and references
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
