% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{GetHTMLPathSet}
\alias{GetHTMLPathSet}
\title{Given a metset inx, return hmtl highlighted pathway cmpds}
\usage{
GetHTMLPathSet(mSetObj = NA, msetNm)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{msetNm}{Input the name of the metabolite set}
}
\description{
Given a metset inx, return hmtl highlighted pathway cmpds
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
