% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{GetKEGGNodeInfo}
\alias{GetKEGGNodeInfo}
\title{Retrieves KEGG node information}
\usage{
GetKEGGNodeInfo(pathName, g, width, height, usr = par("usr"))
}
\arguments{
\item{g}{Input data}

\item{width}{Input the width}

\item{height}{Input the height}

\item{usr}{Input the user}

\item{path.id}{Input the path ID}
}
\description{
Retrieves KEGG node information
}
