% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{GetKMClusterMembers}
\alias{GetKMClusterMembers}
\title{K-means analysis - cluster}
\usage{
GetKMClusterMembers(mSetObj = NA, i)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{i}{Input the cluster index}
}
\description{
Get the cluster members for given index
add HTML color to the names based on its group membership
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
