% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_stats.R
\name{GetMetSetName}
\alias{GetMetSetName}
\title{Given a metset inx, give its name}
\usage{
GetMetSetName(mSetObj = NA, msetInx)
}
\description{
Given a metset inx, give its name
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
