% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{GetNetworkGeneMappingResultTable}
\alias{GetNetworkGeneMappingResultTable}
\title{Exports Gene-Mapping result into a table}
\usage{
GetNetworkGeneMappingResultTable(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Exports Gene-Mapping result into a table
}
