% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{GetORA.smpdbIDs}
\alias{GetORA.smpdbIDs}
\title{Only for human pathways}
\usage{
GetORA.smpdbIDs(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Only for human pathways + ath, eco, mmu & sce
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
