% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_classification.R
\name{GetRFSigMat}
\alias{GetRFSigMat}
\title{Random Forest Significance matrix}
\usage{
GetRFSigMat(mSetObj = NA)
}
\description{
Significance measure, double[][]
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
