% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetROC.coords}
\alias{GetROC.coords}
\title{Return ROC corodinates with confidence intervals}
\usage{
GetROC.coords(mSetObj = NA, fld.nm, val, plot = TRUE, imgNm)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{fld.nm}{The kind of input coordinate}

\item{val}{The coordinates to look for}

\item{plot}{Logical, by default set to TRUE}

\item{imgNm}{Input the image name}
}
\description{
Return ROC corodinates with confidence intervals
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
