% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetROCTtestP}
\alias{GetROCTtestP}
\title{Get p-values for ROC}
\usage{
GetROCTtestP(data, cls)
}
\description{
ROC p-vaues, used in higher function
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
