% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{GetTtestRes}
\alias{GetTtestRes}
\title{Retrieve T-test p-values}
\usage{
GetTtestRes(mSetObj = NA, paired = FALSE, equal.var = TRUE, nonpar = F)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{paired}{Default set to FALSE}

\item{equal.var}{Default set to TRUE}

\item{nonpar}{Use non-parametric tests, default is set to FALSE}
}
\description{
Utility method to get p values
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
