% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{GetXYCluster}
\alias{GetXYCluster}
\title{Determine row/column number for plotting}
\usage{
GetXYCluster(total)
}
\description{
Determine the number of rows and columns for a given total
number of plots (used by Kmeans and SOM plots)
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
