% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R
\name{HMDBID2Name}
\alias{HMDBID2Name}
\title{Given a vector of HMDBIDs, return a vector of HMDB compound names}
\usage{
HMDBID2Name(ids)
}
\arguments{
\item{ids}{Input the vector of HMDB Ids}
}
\description{
This function, when given a vector of HMDBIDs, return a vector of HMDB compound names. HMDB standing
for the Human Metabolome Database.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
