% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{ImportRawMSData}
\alias{ImportRawMSData}
\title{Import raw MS data}
\usage{
ImportRawMSData(foldername, format = "png", dpi = 72, width = 9,
  par.cores = TRUE, plot = TRUE)
}
\arguments{
\item{foldername}{Character, input the file path to the folder containing
the raw MS spectra to be processed.}

\item{format}{Character, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}

\item{par.cores}{Logical, if true, the function will automatically 
set the number of parallel cores. If false, it will not.}

\item{plot}{Logical, if true the function will create BPIS and TICS plots.}
}
\description{
This function handles the reading in of 
raw MS data (.mzML, .CDF and .mzXML). Users must set 
their working directory to the folder containing their raw 
data, divided into two subfolders named their desired group labels. The  
function will output two chromatograms into the user's working directory, a 
base peak intensity chromatogram (BPIC) and a total ion 
chromatogram (TIC). Further, this function sets the number of cores
to be used for parallel processing. It first determines the number of cores 
within a user's computer and then sets it that number/2.
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
