% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{isEmptyMatrix}
\alias{isEmptyMatrix}
\title{Sig table matrix is empty}
\usage{
isEmptyMatrix(mat)
}
\arguments{
\item{mat}{Matrix to test if empty}
}
\description{
Test if a sig table matrix is empty
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
