% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{kwtest}
\alias{kwtest}
\title{Kruskal-Wallis}
\usage{
kwtest(x, cls)
}
\description{
Perform  Kruskal Wallis Test
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
