% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R, R/networks_data.R
\name{LoadKEGGKO_lib}
\alias{LoadKEGGKO_lib}
\title{Utility function for PerformKOEnrichAnalysis_KO01100}
\usage{
LoadKEGGKO_lib(category)

LoadKEGGKO_lib(category)
}
\arguments{
\item{category}{Module or pathway}
}
\description{
Utility function for PerformKOEnrichAnalysis_KO01100

Utility function for PerformKOEnrichAnalysis_KO01100
}
