% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_integ.R
\name{LoadKEGGLib}
\alias{LoadKEGGLib}
\title{Load KEGG library}
\usage{
LoadKEGGLib(libOpt)
}
\arguments{
\item{libOpt}{KEGG library option, "integ" for integrative, "genetic" for genetic, and "met" for metabolic}
}
\description{
Load KEGG library
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
