% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{LoadMsetLib}
\alias{LoadMsetLib}
\title{Load metabolite set library}
\usage{
LoadMsetLib(libname = "kegg_pathway")
}
\arguments{
\item{libname}{Input the name of the metabolite set library to load. Default set to "kegg_pathway" library.}
}
\description{
Metabolite set library
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
