% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{LogNorm}
\alias{LogNorm}
\title{Column-wise Normalization}
\usage{
Options for log, zero mean and unit variance, and
several zero mean and variance/SE
}
\description{
Column-wise norm methods, when x is a column
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
}
