% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{make_cpdlist}
\alias{make_cpdlist}
\title{Utility function to create compound lists for permutation analysis}
\usage{
make_cpdlist(mSetObj=NA, input_mzs)

make_cpdlist(mSetObj=NA, input_mzs)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj}

\item{input_mzs}{The vector of randomly drawn m/z features.}
}
\description{
From a vector of m/z features, this function outputs a vector of compounds.

From a vector of m/z features, this function outputs a vector of compounds.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
