% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R
\name{MapCmpd2KEGGNodes}
\alias{MapCmpd2KEGGNodes}
\title{Utility function for PrepareKeggQueryJson}
\usage{
MapCmpd2KEGGNodes(cmpds, net = "ko01100")
}
\arguments{
\item{cmpds}{Input the compounds}

\item{net}{Input the network name}
}
\description{
Utility function for PrepareKeggQueryJson
}
