% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_data.R
\name{MapKO2KEGGEdges}
\alias{MapKO2KEGGEdges}
\title{Utility function for PrepareQueryJson}
\usage{
MapKO2KEGGEdges(kos, net = "ko01100")
}
\arguments{
\item{kos}{Input the KOs}

\item{net}{Input the name of the network}
}
\description{
Utility function for PrepareQueryJson
}
