% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{MergeDuplicates}
\alias{MergeDuplicates}
\title{Merge duplicated columns or rows by their mean}
\usage{
MergeDuplicates(data, dim = 2)
}
\arguments{
\item{data}{Input the data}

\item{dim}{Numeric, input the dimensions, default is set to 2}
}
\description{
dim 1 => row,  dim 2 => column
}
