% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_c_utils.R
\docType{package}
\name{MetaboAnalystR}
\alias{MetaboAnalystR}
\title{MetaboAnalystR: A package for computating the notorious bar statistic.}
\description{
The MetaboAnalystR package provides a pipeline for metabolomics processing.
}
\section{MetaboAnalystR functions}{

The MetaboAnalystR functions ...
}

