% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_name_match.R
\name{MetaboliteMappingExact}
\alias{MetaboliteMappingExact}
\title{Mapping from different metabolite IDs}
\usage{
MetaboliteMappingExact(mSetObj = NA, q.type)
}
\description{
For compound names to other ids, can do exact or approximate matches
For other IDs, except HMDB ID, all others may return multiple/non-unique hits
Multiple hits or non-unique hits will allow users to manually select
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
