% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{OPLSR.Anal}
\alias{OPLSR.Anal}
\title{Perform OPLS-DA}
\usage{
OPLSR.Anal(mSetObj = NA, reg = FALSE)
}
\description{
Orthogonal PLS-DA (from ropls)
Add reg (regression i.e. if class order matters)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
