% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PCA.Flip}
\alias{PCA.Flip}
\title{Rotate PCA analysis}
\usage{
PCA.Flip(mSetObj = NA, axisOpt)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}
}
\description{
Rotate PCA analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
