% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{Perform.permutation}
\alias{Perform.permutation}
\title{need to check if necessary to include this code
use single core on the public server
Permutation}
\usage{
Perform.permutation(perm.num, fun)
}
\description{
Perform permutation, options to change number of cores used
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
