% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog.R
\name{PerformAdductMapping}
\alias{PerformAdductMapping}
\title{Read Adduct List}
\usage{
Read.AdductData(mSetObj=NA, adductList)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object}

\item{adductList}{Input the name of the adduct list}
}
\description{
This function reads in the user's adduct list and 
saves it as a matrix.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
