% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_batch_check.R
\name{PerformBatchCorrection}
\alias{PerformBatchCorrection}
\title{Set up two matrixes}
\usage{
PerformBatchCorrection(mSetObj = NA, imgName)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imgName}{Input the name of the plot to create}
}
\description{
One is a batch containing summed concentrations of each sample
the other contains the features aligned across all samples
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
