% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{PerformBatchCorrection}
\alias{PerformBatchCorrection}
\title{Batch Effect Correction}
\usage{
PerformBatchCorrection(
  mSetObj = NA,
  imgName = NULL,
  Method = NULL,
  center = NULL
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imgName}{Input the name of the plot to create}

\item{Method}{Batch effect correction method, default is "auto". Specific method, including "Combat",
"WaveICA","EigenMS","QC_RLSC","ANCOVA","RUV_random","RUV_2","RUV_s","RUV_r","RUV_g","NOMIS" and "CCMN".}

\item{center}{The center point of the batch effect correction, based on "QC" or "", which means correct 
to minimize the distance between batches.}
}
\description{
This function is designed to perform the batch effect correction
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
