% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PerformCurrencyMapping}
\alias{PerformCurrencyMapping}
\title{Map currency metabolites to KEGG & BioCyc}
\usage{
PerformCurrencyMapping(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object}
}
\description{
This function maps the user selected list
of compounds to its corresponding KEGG IDs and BioCyc IDs
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
