% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming.R
\name{PerformDataInspect}
\alias{PerformDataInspect}
\title{Data inspectation}
\usage{
PerformDataInspect(datapath, rt.range, mz.range, dimension = "3D", res = 100)
}
\arguments{
\item{datapath}{Character, the path of the raw MS data files (.mzXML, .CDF and .mzML) 
for the visual and intuitive data inspectation.}

\item{rt.range}{Numerics, a congregation of two values to define the lower and upper RT range (seconds) for 
users to inspect. This is an optional parameter, if absent, will display the MS of the whole RT range.}

\item{mz.range}{Numerics, a congregation of two values to define the lower and upper mz range for 
users to inspect. This is an optional parameter, if absent, will display the MS of the whole mz range.}

\item{dimension}{Character, the dimension for sample to display, including '2D' or '3D'. The default is '3D'.}

\item{res}{Numeric, the resolution for data inspectation. The larger the value, the higher the resolution.
The default value is 100. This value is usually clearly enough and also give consideration to the speed.}
}
\description{
This functions provide a path for users to visually inspect their raw data before the data 
trimming so as to remove the dirty or significantly uneluted peaks.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
