% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{PerformDataInspect}
\alias{PerformDataInspect}
\title{PerformDataInspect}
\usage{
PerformDataInspect(
  datapath = NULL,
  rt.range = c(0, 0),
  mz.range = c(0, 0),
  dimension = "3D",
  res = 100
)
}
\description{
PerformDataInspect is used to plot 2D/3D structure of the MS data
}
\author{
Zhiqiang Pang
}
