% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_utils.R
\name{PerformGSEA}
\alias{PerformGSEA}
\title{Perform Gene Set Enrichment Analysis test on single or multiple gene expression matrices}
\usage{
PerformGSEA(
  dataName,
  file.nm,
  fun.type,
  netNm,
  mType,
  selectedFactorInx = 1,
  mode = "multi",
  rankOpt = ""
)
}
\arguments{
\item{dataSetObj}{file name of the data, .txt format}
}
\description{
Perform Gene Set Enrichment Analysis test on single or multiple gene expression matrices
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
