% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_utils.R
\name{PerformLimmaDE}
\alias{PerformLimmaDE}
\title{Perform differential expression analysis using Limma for individually-uploaded data.}
\usage{
PerformLimmaDE(mSetObj = NA, dataName, p.lvl = 0.1, fc.lvl = 0)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{dataName}{Input the name of the individual dataset for normalization.}

\item{p.lvl}{Numeric, input the p-value (FDR) cutoff.}

\item{fc.lvl}{Numeric, input the fold-change (FC) cutoff.}
}
\description{
This function performs DE analysis of individually-uploaded data prior to meta-analysis.
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
