% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_optimization.R
\name{PerformParamsOptimization}
\alias{PerformParamsOptimization}
\title{Perform Parameters Optimization}
\usage{
PerformParamsOptimization(raw_data, param = p0, method = "DoE", ncore = 4)
}
\arguments{
\item{raw_data}{MSnExp object, can be the (trimmed) data in memory produced by 'PerformDataTrimming' or the 
orignal data read by ImportRawMSData with 'inMemory" mode.}

\item{param}{List, Parameters defined by 'SetPeakParam' function.}

\item{method}{Character, method of parameters optimization, including "DoE' only. Default is "DoE". Other method 
is under development.}

\item{ncore}{Numeric, CPU threads number used to perform the parallel based optimization. If thers is memory issue,
please reduce the 'ncore' used here. For default, 2/3 CPU threads of total will be used.}
}
\description{
This function is used to optimize the critical parameters of peak picking and alignment for 
the following data processing. It utilizes the trimed data and the internal instrument-specific parameters.
Parallel computing will be performed. The number of cores user want to use could be specified.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
