% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{PerformPeakAnnotation}
\alias{PerformPeakAnnotation}
\title{Perform peak annotation}
\usage{
PerformPeakAnnotation(xset, annParams)
}
\arguments{
\item{xset}{The object created using the PerformPeakPicking function,
containing the peak picked MS data.}

\item{annParams}{The object created using the SetAnnotationParam function, 
containing user's specified or default parameters for downstream 
raw MS data pre-processing.}
}
\description{
This function performs peak annotation on
the xset object created using the PerformPeakPicking function.
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
