% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{PerformPeakProfiling}
\alias{PerformPeakProfiling}
\title{Perform peak profiling
This function performs feature extraction of user's raw MS data using 
the rawData object created using the ImportRawMSData function.}
\usage{
PerformPeakProfiling(rawData, Params, plotSettings, ncore)
}
\arguments{
\item{rawData}{The object created using the ImportRawMSData function,
containing the raw MS data.}

\item{Params}{The object created using the SetPeakParam function, 
containing user's specified or default parameters for downstream 
raw MS data pre-processing.}

\item{plotSettings}{List, plotting parameters produced by SetPlotParam Function.
Defaut is set to true.}

\item{ncore}{Numeric, used to define the cores' number for Peak Profiling.}
}
\description{
Perform peak profiling
This function performs feature extraction of user's raw MS data using 
the rawData object created using the ImportRawMSData function.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
