% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PerformPSEA}
\alias{PerformPSEA}
\title{Function to perform peak set enrichment analysis}
\usage{
PerformPSEA(mSetObj=NA, lib, libVersion, minLib, permNum = 100)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object.}

\item{lib}{Input the name of the organism library, default is hsa_mfn.}

\item{libVersion}{Input the version of the KEGG pathway libraries ("current" or "old").}

\item{minLib}{Numeric, input the minimum number of metabolites needed to consider the pathway 
or metabolite set.}

\item{permNum}{Numeric, input the number of permutations to perform. Default is 100.}
}
\description{
This is the main function that performs either the mummichog
algorithm, GSEA, or both for peak set enrichment analysis.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
