% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{PlotCmpdSummary}
\alias{PlotCmpdSummary}
\title{Plot compound summary
change to use dataSet$procr instead of dataSet$orig in
case of too many NAs}
\usage{
PlotCmpdSummary(mSetObj = NA, cmpdNm, format = "png", dpi = 72,
  width = NA)
}
\description{
Plot compound summary
change to use dataSet$procr instead of dataSet$orig in
case of too many NAs
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
