% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotDetailROC}
\alias{PlotDetailROC}
\title{Plot detailed ROC}
\usage{
PlotDetailROC(mSetObj = NA, imgName, thresh, sp, se, dpi = 72,
  format = "png")
}
\description{
Plot detailed ROC
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
