% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_asca_heatmap2.R
\name{PlotHeatMap2}
\alias{PlotHeatMap2}
\title{Plot heatmap visualization for time-series data}
\usage{
PlotHeatMap2(mSetObj, imgName, format, dpi, width, smplDist, clstDist, colors, viewOpt, hiRes, sortInx, useSigFeature, drawBorder, var.inx)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{smplDist}{select distance measure: euclidean, pearson, or minkowski}

\item{clstDist}{select clustering algorithm: ward, average, complete, single}

\item{colors}{select heatmap colors: bwm, gray}

\item{viewOpt}{select overview or detailed view: overview or detail}

\item{useSigFeature}{use significant features only: F or T (default false)}

\item{drawborder}{show cell borders: F or T (default F)}
}
\description{
Plot heatmap visualization for time-series data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
