% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotImpVar}
\alias{PlotImpVar}
\title{Plot PLS important variables,}
\usage{
PlotImpVar(mSetObj = NA, imp.vec, xlbl, feat.num = 15, color.BW = FALSE)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imp.vec}{Input the vector of important variables}

\item{xlbl}{Input the x-label}

\item{feat.num}{Numeric, set the feature numbers, default is set to 15}

\item{color.BW}{Use black-white for plot (T) or colors (F)}
}
\description{
Plot PLS important variables, BHan: added bgcolor parameter for B/W color
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
