% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_kegg.R
\name{PlotinmexGraph}
\alias{PlotinmexGraph}
\title{Plot an igraph object and return the node information (position and labels)}
\usage{
PlotinmexGraph(mSetObj, path.id, g, width, height, bg.color = NULL,
  line.color = NULL)
}
\description{
Plot an igraph object and return the node information (position and labels)
Used in a higher function
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
