% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{PlotIntegPaths}
\alias{PlotIntegPaths}
\title{PlotIntegPaths}
\usage{
PlotIntegPaths(
  mSetObj = NA,
  imgName,
  format = "png",
  dpi = 72,
  width = 9,
  labels = "default",
  labels.x = 5,
  labels.y = 5
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj object}

\item{imgName}{Input a name for the plot}

\item{format}{Character, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}

\item{labels.x}{Numeric, indicate the number of top-ranked pathways using the fGSEA algorithm 
to annotate on the plot.}

\item{labels.y}{Numeric, indicate the number of top-ranked pathways using the original 
mummichog algorithm to annotate on the plot.}

\item{Labels}{Character, indicate if the plot should be labeled. By default
it is set to "default", and the 5 top-ranked pathways per each algorithm will be plotted.
Users can adjust the number of pathways to be annotated per pathway using the "labels.x" 
and "labels.y" parameters.
Users can set this to "none" for no annotations, or "all" to annotate all pathways.}
}
\description{
Plots both the original mummichog and the GSEA results by combining p-values
using the Fisher's method (sumlog).
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
