% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotKEGGPath}
\alias{PlotKEGGPath}
\title{Plot KEGG pathway}
\usage{
PlotKEGGPath(mSetObj = NA, pathName, format = "png", width = NA,
  dpi = 72)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{pathName}{Input the name of the selected KEGG pathway}

\item{format}{Select the image format, "png", or "pdf".}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}
}
\description{
Orthogonal PLS-DA (from ropls)
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
