% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotKEGGPath}
\alias{PlotKEGGPath}
\title{Plot metabolome pathway}
\usage{
PlotKEGGPath(
  mSetObj = NA,
  pathName,
  width = NA,
  height = NA,
  format = "png",
  dpi = NULL
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{pathName}{Input the name of the selected pathway}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{height}{Input the height of the created plot.}
}
\description{
Plot KEGG pathway graph
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
