% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_asca_heatmap2.R
\name{PlotMetaCorrHeatmap}
\alias{PlotMetaCorrHeatmap}
\title{Generate correlation heatmap for metadata}
\usage{
PlotMetaCorrHeatmap(
  mSetObj = NA,
  cor.opt = "pearson",
  imgName,
  format = "png",
  dpi = 96,
  width = NA
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{cor.opt}{Meethod for computing correlation coefficient}
}
\description{
Plot correlation coefficients between metadata
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
