% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotMetPath}
\alias{PlotMetPath}
\title{R script for generating various graphics for pathway analysis}
\usage{
PlotMetPath(mSetObj = NA, pathName, width, height)
}
\description{
Orthogonal PLS-DA (from ropls)
}
\details{
Plot metabolome pathway
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
