% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{PlotMS.RT}
\alias{PlotMS.RT}
\title{Plot rentention time corrected spectra}
\usage{
PlotMS.RT(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA)
}
\description{
Plot the retention time corrected spectra
}
