% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{plotMSfeature}
\alias{plotMSfeature}
\title{plotMSfeature}
\usage{
plotMSfeature(FeatureNM)
}
\description{
plotMSfeature is used to plot MS feature stats for different groups
}
\author{
Zhiqiang Pang
}
