% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{plotMSfeature}
\alias{plotMSfeature}
\title{plotMSfeature}
\usage{
plotMSfeature(FeatureNM, format = "png", dpi = default.dpi, width = NA)
}
\arguments{
\item{FeatureNM}{FeatureNM}

\item{format}{format}

\item{dpi}{dpi}

\item{width}{width}
}
\description{
plotMSfeature is used to plot MS feature stats for different groups
}
\author{
Zhiqiang Pang
}
