% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotOPLS.Imp}
\alias{PlotOPLS.Imp}
\title{PlotOPLS.Imp
OPLS VIP plotting function}
\usage{
PlotOPLS.Imp(
  mSetObj = NA,
  imgName,
  format = "png",
  dpi = 72,
  width = NA,
  type = "vip",
  feat.nm = "tscore",
  feat.num = 15,
  color.BW = FALSE
)
}
\arguments{
\item{mSetObj}{mSetObj objects generated from last step}

\item{imgName}{image name}

\item{format}{image format, can be "png", "jpg", "tiff", "pdf" and "svg"}

\item{dpi}{numeric, dpi number}

\item{width}{numeric, width number}

\item{type}{analysis type, can be "vip" only}

\item{feat.nm}{feature name, should be "tscore" for now}

\item{feat.num}{feature number}

\item{color.BW}{color information}
}
\description{
PlotOPLS.Imp
OPLS VIP plotting function
}
