% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{PlotPathSummary}
\alias{PlotPathSummary}
\title{Plot a scatterplot (circle) overview of the matched pathways}
\usage{
PlotPathSummary(
  mSetObj = NA,
  show.grid,
  imgName,
  format = "png",
  dpi = 72,
  width = NA,
  xlim = NA,
  ylim = NA
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{show.grid}{logical, show grid or not}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.}

\item{xlim}{limit of x axis}

\item{ylim}{limit of y axis}
}
\description{
x axis is the pathway impact factor
y axis is the p value (from ORA or globaltest) 
return the circle information
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
