% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PlotPLS3DScore}
\alias{PlotPLS3DScore}
\title{Plot PLS 3D score plot}
\usage{
PlotPLS3DScore(mSetObj = NA, imgName, format = "json", inx1, inx2, inx3)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", or "pdf".}

\item{inx1}{Numeric, indicate the number of the principal component for the x-axis of the loading plot.}

\item{inx2}{Numeric, indicate the number of the principal component for the y-axis of the loading plot.}

\item{inx3}{Numeric, indicate the number of the principal component for the z-axis of the loading plot.}
}
\description{
Plot PLS 3D score plot
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
