% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotRocUnivBoxPlot}
\alias{PlotRocUnivBoxPlot}
\title{Plot a boxplot view of a selected compound}
\usage{
PlotRocUnivBoxPlot(
  mSetObj,
  feat.nm,
  version,
  format = "png",
  dpi = 72,
  isOpt,
  isQuery
)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{feat.nm}{Input the name of the selected compound.}

\item{version}{version mark for image name}

\item{format}{Select the image format, "png", of "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{isOpt}{logical}

\item{isQuery}{logical}
}
\description{
Plots a boxplot of the selected compound's concentration
between the groups.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
