% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_processing.R
\name{plotSingleTIC}
\alias{plotSingleTIC}
\title{plotSingleTIC}
\usage{
plotSingleTIC(filename, imageNumber, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{filename}{filename}

\item{imageNumber}{imageNumber}

\item{format}{format}

\item{dpi}{dpi}

\item{width}{width}
}
\description{
plotSingleTIC is used to plot single TIC
}
\author{
Zhiqiang Pang
}
