% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{PlotSubHeatMap}
\alias{PlotSubHeatMap}
\title{Create Sub Heat Map Plot}
\usage{
PlotSubHeatMap(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, dataOpt, scaleOpt, smplDist, clstDist, palette, method.nm,
  top.num, viewOpt, rowV = T, colV = T, border = T, grp.ave = F)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", or "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}
}
\description{
Plot a sub heatmap based on results from t-tests/ANOVA, VIP or randomforest
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
