% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{PlotSubHeatMap}
\alias{PlotSubHeatMap}
\title{Create Sub Heat Map Plot}
\usage{
PlotSubHeatMap(mSetObj = NA, imgName, format = "png", dpi = 72,
  width = NA, dataOpt, scaleOpt, smplDist, clstDist, palette, method.nm,
  top.num, viewOpt, rowV = T, colV = T, border = T, grp.ave = F)
}
\description{
Plot a sub heatmap based on results from t-tests/ANOVA, VIP or randomforest
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
