% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R, R/time_pca_anova2.R
\name{plotVennDiagram}
\alias{plotVennDiagram}
\alias{plotVennDiagram}
\title{Utility function: Plot Venn diagram
Gordon Smyth, James Wettenhall.
Capabilities for multiple counts and colors by Francois Pepin.
4 July 2003.  Last modified 12 March 2010.}
\usage{
plotVennDiagram(object, include = "both", names, mar = rep(0, 4),
  cex = 1.2, lwd = 1, circle.col, counts.col, show.include, ...)

plotVennDiagram(object, include = "both", names, mar = rep(0, 4),
  cex = 1.2, lwd = 1, circle.col, counts.col, show.include, ...)
}
\description{
Utility function: Plot Venn diagram
Gordon Smyth, James Wettenhall.
Capabilities for multiple counts and colors by Francois Pepin.
4 July 2003.  Last modified 12 March 2010.

Capabilities for multiple counts and colors by Francois Pepin
}
\author{
Gordon Smyth, James Wettenhall
4 July 2003.  Last modified 12 March 2010
}
