% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_chemometrics.R
\name{PLSDA.CV}
\alias{PLSDA.CV}
\title{PLS-DA classification and feature selection}
\usage{
PLSDA.CV(
  mSetObj = NA,
  methodName = "T",
  compNum = GetDefaultPLSCVComp(mSetObj),
  choice = "Q2",
  segments = 10
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{methodName}{Logical, by default set to TRUE}

\item{compNum}{GetDefaultPLSCVComp()}

\item{choice}{Input the choice, by default it is Q2}
}
\description{
PLS-DA classification and feature selection
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
