% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Predict.class}
\alias{Predict.class}
\title{Get predicted class probability}
\usage{
Predict.class(x.train, y.train, x.test, clsMethod = "pls", lvNum,
  imp.out = F)
}
\description{
Get predicted class probability
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
