% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PrepareUpsetData}
\alias{PrepareUpsetData}
\title{Prepare data for Upset diagram}
\usage{
PrepareUpsetData(mSetObj = NA, fileNm)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{fileNm}{json file name to save}
}
\description{
Prepare data for Upset diagram
}
