% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_batch_check.R
\name{Read.BatchCSVdata}
\alias{Read.BatchCSVdata}
\title{Data I/O for batch effect checking}
\usage{
Read.BatchCSVdata(mSetObj = NA, filePath, format, label)
}
\description{
Read multiple user uploaded CSV data one by one
format: row, col
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
