% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_utils.R
\name{Read.BatchDataBC}
\alias{Read.BatchDataBC}
\title{Data I/O for batch effect checking}
\usage{
Read.BatchDataBC(mSetObj = NA, filePath, format, label, missingEstimate)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{filePath}{Input the path to the batch files}

\item{format}{Input the format of the batch files}

\item{label}{Input the label-type of the files}

\item{missingEstimate}{Approach to estimate the missing values}
}
\description{
Read multiple user uploaded CSV data one by one
format: row, col
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
