% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{Read.PeakListData}
\alias{Read.PeakListData}
\title{Constructor to read uploaded user files into the mummichog object}
\usage{
Read.PeakListData(mSetObj=NA, filename = NA)

Read.PeakListData(mSetObj=NA, filename = NA)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}

\item{filename}{Input the path name for the CSV/TXT files to read.}

\item{meta.anal}{Logical, TRUE if data will be used for meta-analysis.}

\item{method}{Input the type of statistical scores included in the
mummichog input. "pvalue" for p-values, "es" for effect-sizes, and
"both" for both p-values and effect-sizes.}
}
\description{
This function handles reading in CSV or TXT files and filling in the mSet object
for mummichog analysis. It makes sure that all necessary columns are present.

This function handles reading in CSV or TXT files and filling in the mSet object
for mummichog analysis. It makes sure that all necessary columns are present.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
