% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{ReadMetaData}
\alias{ReadMetaData}
\title{ReadMetaData}
\usage{
ReadMetaData(mSetObj = NA, metafilename)
}
\arguments{
\item{mSetObj}{metaboanalyst object, initialized by InitDataObjects("pktable", "ts", FALSE)}

\item{metafilename}{file path of data}
}
\description{
ReadMetaData
}
