% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{rectUnique}
\alias{rectUnique}
\title{Perform utilities for peak grouping}
\usage{
rectUnique(m, order = seq(length = nrow(m)), xdiff = 0, ydiff = 0)
}
\description{
Perform various utilities for peak grouping
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
