% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_graphics.R
\name{RerenderMetPAGraph}
\alias{RerenderMetPAGraph}
\title{Redraw current graph for zooming or clipping then return a value}
\usage{
RerenderMetPAGraph(mSetObj = NA, imgName, width, height, zoom.factor = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{imgName}{Input the name of the plot}

\item{width}{Input the width, there are 2 default widths, the first, width = NULL, is 10.5.
The second default is width = 0, where the width is 7.2. Otherwise users can input their own width.}

\item{height}{Input the height of the created plot.}
}
\description{
Redraw current graph for zooming or clipping then return a value
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
