% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{SAM.Anal}
\alias{SAM.Anal}
\title{Perform Signifiance Analysis of Microarrays (SAM) analysis}
\usage{
SAM.Anal(mSetObj = NA, method = "d.stat", paired = FALSE,
  varequal = TRUE)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{method}{Method for SAM analysis, default is set to "d.stat", alternative is "wilc.stat"}

\item{paired}{Logical, indicates if samples are paired or not. Default is set to FALSE}

\item{varequal}{Logical, indicates if variance is equal. Default is set to TRUE}
}
\description{
Perform SAM
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
