% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SearchByName}
\alias{SearchByName}
\title{Given a metabolite set name, search its index}
\usage{
SearchByName(mSetObj = NA, query)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{query}{Input the query to search}
}
\description{
Given a metabolite set name, search its index
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
