% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sigfeatures.R
\name{SetEBAMSigMat}
\alias{SetEBAMSigMat}
\title{For EBAM analysis}
\usage{
SetEBAMSigMat(mSetObj = NA, delta)
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{delta}{Input the delta for EBAM analysis}
}
\description{
return double matrix with 3 columns - z.value, posterior, local.fdr
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
