% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{SetKEGG.PathLib}
\alias{SetKEGG.PathLib}
\title{Set KEGG pathway library}
\usage{
SetKEGG.PathLib(mSetObj = NA, libNm, lib.version)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{libNm}{lib name option
KEGG pathway library or "v2018" for the KEGG pathway library version prior to November 2019.}

\item{lib.version}{Input the KEGG pathway version. "current" for the latest}
}
\description{
note, this process can be long, need to return a value
to force Java to wait
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
