% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pathway.R, R/peaks_to_function.R
\name{SetMummichogPvalFromPercent}
\alias{SetMummichogPvalFromPercent}
\title{Set the cutoff for mummichog analysis}
\usage{
SetMummichogPvalFromPercent(mSetObj = NA, fraction)

SetMummichogPvalFromPercent(mSetObj = NA, fraction)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}
}
\description{
Set the p-value cutoff for mummichog analysis.

Set the p-value cutoff for mummichog analysis.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)

Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
