% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anot_utils.R, R/peaks_to_function.R
\name{SetOrganism}
\alias{SetOrganism}
\title{Set organism for further analysis}
\usage{
SetOrganism(mSetObj = NA, org)

SetOrganism(mSetObj = NA, org)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{org}{Set organism ID}
}
\description{
Set organism for further analysis

Set organism for further analysis
}
