% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{SetPeakEnrichMethod}
\alias{SetPeakEnrichMethod}
\title{Set the peak enrichment method for the MS Peaks to Paths module}
\usage{
SetPeakEnrichMethod(mSetObj = NA, algOpt)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj.}
}
\description{
This function sets the peak enrichment method.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
