% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc_utils.R
\name{SetPeakList.GroupValues}
\alias{SetPeakList.GroupValues}
\title{Set peak list group values}
\usage{
SetPeakList.GroupValues(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of mSetObj, the data used is the nmr.xcmsSet object}
}
\description{
Set peak list group values
}
