% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{SetPeakParam}
\alias{SetPeakParam}
\title{SetPeakParam}
\usage{
SetPeakParam(
  platform = "general",
  Peak_method = "centWave",
  RT_method = "loess",
  mzdiff,
  snthresh,
  bw,
  ppm,
  min_peakwidth,
  max_peakwidth,
  noise,
  prefilter,
  value_of_prefilter,
  fwhm,
  steps,
  sigma,
  peakBinSize,
  max,
  criticalValue,
  consecMissedLimit,
  unions,
  checkBack,
  withWave,
  profStep,
  minFraction,
  minSamples,
  maxFeatures,
  mzCenterFun,
  integrate,
  extra,
  span,
  smooth,
  family,
  fitgauss,
  polarity,
  perc_fwhm,
  mz_abs_iso,
  max_charge,
  max_iso,
  corr_eic_th,
  mz_abs_add,
  adducts,
  rmConts
)
}
\description{
SetPeakParam, used to set the peak param
}
\author{
Zhiqiang Pang
}
