% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_utils.R
\name{Setup.ConcData}
\alias{Setup.ConcData}
\title{Save concentration data}
\usage{
Setup.ConcData(mSetObj = NA, conc)
}
\description{
Save concentration data
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
