% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_mset.R
\name{Setup.KEGGReferenceMetabolome}
\alias{Setup.KEGGReferenceMetabolome}
\title{Read user uploaded metabolome as a list of KEGG pathway ids}
\usage{
Setup.KEGGReferenceMetabolome(mSetObj = NA, filePath)
}
\description{
Read user uploaded metabolome as a list of KEGG pathway ids
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
