% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_path_stats.R
\name{SetupSMPDBLinks}
\alias{SetupSMPDBLinks}
\title{Only works for human (hsa.rda) data}
\usage{
SetupSMPDBLinks(kegg.ids)
}
\description{
Only works for human (hsa.rda) data
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
