% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_clustering.R
\name{SOM.Anal}
\alias{SOM.Anal}
\title{SOM analysis}
\usage{
SOM.Anal(mSetObj = NA, x.dim, y.dim, initMethod, neigb = "gaussian")
}
\arguments{
\item{mSetObj}{Input name of the created mSet Object}

\item{x.dim}{Input X dimension for SOM analysis}

\item{y.dim}{Input Y dimension for SOM analysis}

\item{initMethod}{Input the method}

\item{neigb}{Default is set to 'gaussian'}
}
\description{
SOM analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
