% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_spls.R
\name{sparse.mint.block_iteration}
\alias{sparse.mint.block_iteration}
\title{Perform Sparse Generalized Canonical Correlation (sgccak)}
\usage{
sparse.mint.block_iteration(
  A,
  design,
  study = NULL,
  keepA.constraint = NULL,
  keepA = NULL,
  scheme = "horst",
  init = "svd",
  max.iter = 100,
  tol = 1e-06,
  verbose = TRUE,
  bias = FALSE,
  penalty = NULL
)
}
\arguments{
\item{A}{Data}

\item{design}{Set design}

\item{study}{Default set to NULL}

\item{keepA.constraint}{Default set to NULL}

\item{keepA}{Default set to NULL}

\item{scheme}{Scheme, default set to "horst"}

\item{init}{Init mode, default set to "svd"}

\item{max.iter}{Max number of iterations, numeric, default set to 100}

\item{tol}{Tolerance, numeric, default set to 1e-06}

\item{verbose}{Default set to TRUE}

\item{bias}{Default set to FALSE}

\item{penalty}{Default set to NULL}
}
\description{
Runs sgccak() modified from RGCCA
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
