% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_norm_utils.R
\name{SumNorm}
\alias{SumNorm}
\title{Row-wise Normalization}
\usage{
Options for normalize by sum median, reference sample,
reference reference (compound), or quantile normalization
}
\description{
Row-wise norm methods, when x is a row.
Normalize by a sum of each sample, assume constant sum (1000).
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
}
