% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{UnzipUploadedFile}
\alias{UnzipUploadedFile}
\title{Unzip .zip files}
\usage{
UnzipUploadedFile(inPath, outPath, rmFile = T)
}
\arguments{
\item{inPath}{Input the path of the zipped files}

\item{outPath}{Input the path to directory where the unzipped files will be deposited}

\item{rmFile}{Logical, input whether or not to remove files. Default set to T}
}
\description{
Unzips uploaded .zip files, removes the uploaded file, checks for success
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
