% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_to_function.R
\name{UpdateInstrumentParameters}
\alias{UpdateInstrumentParameters}
\title{Update the mSetObj with user-selected parameters for MS Peaks to Pathways.}
\usage{
UpdateInstrumentParameters(mSetObj=NA, instrumentOpt, 
msModeOpt, force_primary_ion, rt_frac, rt_tol)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects).}

\item{instrumentOpt}{Numeric. Define the mass-spec instrument used to perform untargeted metabolomics.}

\item{msModeOpt}{Character. Define the mass-spec mode of the instrument used to perform untargeted metabolomics.}

\item{force_primary_ion}{Character, if "yes", only mz features that match compounds with a primary ion are kept.}

\item{rt_frac}{rt_frac.}

\item{rt_tol}{rt_tol.}
}
\description{
This functions handles updating the mSet object for mummichog analysis. 
It is necessary to utilize this function
to specify to the organism's pathways to use (libOpt), the mass-spec mode (msModeOpt) 
and mass-spec instrument (instrumentOpt).
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT License
}
