% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_univariates.R
\name{Volcano.Anal}
\alias{Volcano.Anal}
\title{Perform Volcano Analysis}
\usage{
Volcano.Anal(mSetObj=NA, paired=FALSE, fcthresh, cmpType, percent.thresh, nonpar=F, threshp, equal.var=TRUE, pval.type="raw")
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}

\item{paired}{Logical, T if data is paired, F if data is not.}

\item{fcthresh}{Numeric, input the fold change threshold}

\item{cmpType}{Comparison type, 1 indicates group 1 vs group 2, and 2 indicates group 2 vs group 1}

\item{percent.thresh}{Only for paired data, numeric, indicate the significant count threshold}

\item{nonpar}{Logical, indicate if a non-parametric test should be used (T or F)}

\item{threshp}{Numeric, indicate the p-value threshold}

\item{equal.var}{Logical, indicates if the group variance is equal (T) or unequal (F)}

\item{pval.type}{To indicate raw p-values, use "raw". To indicate FDR-adjusted p-values, use "fdr".}
}
\description{
Perform volcano analysis
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
